% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{add_points_shaper}
\alias{add_points_shaper}
\title{Add confidence ellipse or convex hull for group of points}
\usage{
add_points_shaper(plot_data, lwd, lty, conf.level, opacity, shape_function,
  ...)
}
\arguments{
\item{plot_data}{plot data returned by function `mdaplot()`}

\item{lwd}{thickness of line used to show the hull}

\item{lty}{type of line used to show the hull}

\item{conf.level}{confidence level to make the ellipse for (between 0 and 1)}

\item{opacity}{of opacity is larger than 0 a semi-transparent polygon is shown over points}

\item{shape_function}{function which calculates and return coordinates of the shape}
}
\description{
Add confidence ellipse or convex hull for group of points
}
