% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{apply.constraint}
\alias{apply.constraint}
\title{Applies constraint to a dataset}
\usage{
\method{apply}{constraint}(obj, x)
}
\arguments{
\item{obj}{object with constraint}

\item{x}{matrix with data values}
}
\description{
Applies constraint to a dataset
}
