% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{as.data.frame.ddsimcares}
\alias{as.data.frame.ddsimcares}
\title{Creates a data frame from DD-SIMCA classification results.}
\usage{
\method{as.data.frame}{ddsimcares}(x, ncomp = x$ncomp.selected, limType = "classic", ...)
}
\arguments{
\item{x}{classification results (object of class \code{ddsimcamres}).}

\item{ncomp}{model complexity (number of components) to compute the classiciation results for.}

\item{limType}{limit type to use ('classic' or 'robust')}

\item{...}{other arguments}
}
\value{
a data frame.
}
\description{
The data frame contains object label, class label (if provided), role, decision, as well
ass valued for distances (h/h0, q/q0, f) for every object from dataset used to
create the results object.
}
