% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{as.matrix.ddsimcares}
\alias{as.matrix.ddsimcares}
\title{Creates a matrix from DD-SIMCA classification results.}
\usage{
\method{as.matrix}{ddsimcares}(x, limType = "classic", ...)
}
\arguments{
\item{x}{classification results (object of class \code{ddsimcamres}).}

\item{limType}{limit type to use ('classic' or 'robust')}

\item{...}{other arguments}
}
\value{
a matrix.
}
\description{
The matrix contains results characteristics computed for different number
of components, including number of objects accepted/rejected by the model,
explained and cumulative explained variance, as well as (if reference classes
are provided) main figures of merits (TP, FP, TN, FN, sensitivity, specificity,
efficiency and selectivity).

If dataset uuse to create the results contained only target class members or only
objects from alternative classes, the FoMs will be reduced accordingly.
}
