\name{as.matrix.ldecomp}
\alias{as.matrix.ldecomp}
\title{Converts ldecomp object to a matrix}
\usage{
as.matrix.ldecomp(x, ...)
}
\arguments{
  \item{x}{object of class \code{ldecomp}}

  \item{...}{other arguments}
}
\description{
Generic \code{as.matrix} function for linear decomposition.
Returns a matrix with information about the decomposition.
}

