% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcamres.R
\name{as.matrix.simcamres}
\alias{as.matrix.simcamres}
\title{as.matrix method for SIMCAM results}
\usage{
\method{as.matrix}{simcamres}(x, nc = seq_len(x$nclasses), ...)
}
\arguments{
\item{x}{classification results (object of class \code{plsdares}, \code{simcamres}, etc.).}

\item{nc}{vector with classes to use.}

\item{...}{other arguments}
}
\description{
Generic \code{as.matrix} function for SIMCAM results. Returns matrix with performance
values for specific class.
}
