% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{asjson.pca}
\alias{asjson.pca}
\title{Converts object with PCA model to JSON string compatible with web-application.}
\usage{
\method{asjson}{pca}(obj, ...)
}
\arguments{
\item{obj}{Object with PCA model (from \code{\link{pca}}).}

\item{...}{other arguments}
}
\value{
stringified JSON
}
\description{
Converts object with PCA model to JSON string compatible with web-application.
}
