% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{asjson}
\alias{asjson}
\title{S3 implementation of asjson() method}
\usage{
asjson(obj)
}
\arguments{
\item{obj}{object of any class, e.g. \code{pca}}
}
\description{
S3 implementation of asjson() method
}
