% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{asvector.pca}
\alias{asvector.pca}
\title{Converts object with PCA model to numeric vector compatible with web-application.}
\usage{
\method{asvector}{pca}(obj)
}
\arguments{
\item{obj}{Object with PCA model (from \code{\link{pca}}).}
}
\value{
vector with values.
}
\description{
Converts object with PCA model to numeric vector compatible with web-application.
}
