% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{asvector.pls}
\alias{asvector.pls}
\title{Converts object with PLS model to numeric vector compatible with web-application.}
\usage{
\method{asvector}{pls}(obj)
}
\arguments{
\item{obj}{Object with PLS model (from \code{\link{pls}}).}
}
\value{
vector with values.
}
\description{
Converts object with PLS model to numeric vector compatible with web-application.
}
