% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{asvector}
\alias{asvector}
\title{S3 implementation of as.vector() method}
\usage{
asvector(obj)
}
\arguments{
\item{obj}{object of any class, e.g. \code{pca}}
}
\description{
S3 implementation of as.vector() method
}
