% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize text or vector with text values}
\usage{
capitalize(s)

capitalize(s)
}
\arguments{
\item{s}{string or vector of strings}

\item{str}{text of vector with text values}
}
\description{
Capitalize text or vector with text values

Capitalize first letter of sting (or vector of strings)
}
