% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{categorize}
\alias{categorize}
\title{Categorize PCA results}
\usage{
categorize(obj, ...)
}
\arguments{
\item{obj}{object with PCA model}

\item{...}{other parameters}
}
\description{
Categorize PCA results
}
