% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{chisq.crit}
\alias{chisq.crit}
\title{Calculates critical limits for distance values using Chi-square distribution}
\usage{
chisq.crit(U, ncomp, alpha = 0.05, gamma = 0.01)
}
\arguments{
\item{U}{matrix or vector with distance values}

\item{ncomp}{number of components}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}
}
\description{
The method is based on Chi-squared distribution with DF = 2 * (m(u)/s(u)^2
}
