% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{chisq.prob}
\alias{chisq.prob}
\title{Calculate probabilities for distance values using Chi-square distribution}
\usage{
chisq.prob(u, ncomp, lim)
}
\arguments{
\item{u}{vector with distances}

\item{ncomp}{number of components}

\item{lim}{vector with parameters}
}
\description{
Calculate probabilities for distance values using Chi-square distribution
}
