% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{clamp.dof}
\alias{clamp.dof}
\title{Round and clamp degrees of freedom to valid range [1, 250]}
\usage{
clamp.dof(x)
}
\arguments{
\item{x}{vector with raw degrees of freedom values}
}
\value{
vector with rounded and clamped values
}
\description{
Round and clamp degrees of freedom to valid range [1, 250]
}
