% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{classify.plsda}
\alias{classify.plsda}
\title{PLS-DA classification}
\usage{
classify.plsda(model, y)
}
\arguments{
\item{model}{a PLS-DA model (object of class \code{plsda})}

\item{y}{a matrix with predicted y values}
}
\value{
Classification results (an object of class \code{classres})
}
\description{
Converts PLS predictions of y values to predictions of classes
}
\details{
This is a service function for PLS-DA class, do not use it manually.
}

