% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simca.R
\name{classify.simca}
\alias{classify.simca}
\title{SIMCA classification}
\usage{
classify.simca(obj, pca.res, c.ref = NULL)
}
\arguments{
\item{obj}{a SIMCA model (object of class \code{simca})}

\item{pca.res}{results of projection data to PCA space}

\item{c.ref}{vector with class reference values}
}
\value{
vector with predicted class values (\code{c.pred})
}
\description{
Make classification based on calculated T2 and Q values and corresponding limits
}
\details{
This is a service function for SIMCA class, do not use it manually.
}
