% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmodel.R
\name{classmodel.processRefValues}
\alias{classmodel.processRefValues}
\title{Check reference class values and convert it to a factor if necessary}
\usage{
classmodel.processRefValues(c.ref, classnames = NULL)
}
\arguments{
\item{c.ref}{class reference values provided by user}

\item{classnames}{text with class name in case of logical reference values}
}
\description{
Check reference class values and convert it to a factor if necessary
}
