% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cleanLabels}
\alias{cleanLabels}
\title{Clean text labels from extra elements so they are compatible with JSON}
\usage{
cleanLabels(str)
}
\arguments{
\item{str}{label string (for example produced by \code{expression}).}
}
\description{
Clean text labels from extra elements so they are compatible with JSON
}
