% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint}
\alias{constraint}
\title{Class for MCR-ALS constraint}
\usage{
constraint(name, params = NULL, method = NULL)
}
\arguments{
\item{name}{short text with name for the constraint}

\item{params}{a list with parameters for the constraint method (if NULL - default parameters will be used)}

\item{method}{method to call when applying the constraint, provide it only for user defined constraints}
}
\description{
Class for MCR-ALS constraint
}
\details{
Use this class to create constraints and add them to a list for MCR-ALS curve resuliton (see
\code{\link{mcrals}}). Either provide name and parameters to one of the existing constraint
implementations or make your own. See the list of implemented constraints by running
\code{constraints()}

For your own constraint you need to create a method, which takes matrix with values (either
spectra or contributions being resolved) as the first argument, does something and then return
a matrix with the same dimension as the result. The method can have any number of optional
parameters.

See help for \code{\link{mcrals}} or Bookdown tutorial for details.
}
