% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraintAngle}
\alias{constraintAngle}
\title{Method for angle constraint}
\usage{
constraintAngle(x, d, weight = 0.05)
}
\arguments{
\item{x}{data matrix (spectra or contributions)}

\item{d}{matrix with the original spectral values}

\item{weight}{how many percent of mean to add (between 0 and 1)}
}
\description{
Adds a small portion of mean to contributions or spectra to increase contrast
}
