% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraintClosure}
\alias{constraintClosure}
\title{Method for closure constraint}
\usage{
constraintClosure(x, d, sum = 1)
}
\arguments{
\item{x}{data matrix (spectra or contributions)}

\item{d}{matrix with the original spectral values}

\item{sum}{which value the specra or contributions should sum up to}
}
\description{
Force rows of data sum up to given value
}
