% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraintNonNegativity}
\alias{constraintNonNegativity}
\title{Method for non-negativity constraint}
\usage{
constraintNonNegativity(x)
}
\arguments{
\item{x}{data matrix (spectra or contributions)}
}
\description{
Set all negative values in the matrix to 0
}
