% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraintNorm}
\alias{constraintNorm}
\title{Method for normalization constraint}
\usage{
constraintNorm(x, type = "area")
}
\arguments{
\item{x}{data matrix (spectra or contributions)}

\item{type}{type of normalization ("area" or "length")}
}
\description{
Normalize rows of matrix to unit length or area
}
