% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_categories}
\alias{create_categories}
\title{Create a factor with categories (regular, extreme, outlier)}
\usage{
create_categories(nobj, extremes_ind, outliers_ind)
}
\arguments{
\item{nobj}{number of objects}

\item{extremes_ind}{logical or numeric indices for extreme objects}

\item{outliers_ind}{logical or numeric indices for outlier objects}
}
\description{
Create a factor with categories (regular, extreme, outlier)
}
