% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{create_plot_data}
\alias{create_plot_data}
\title{Create dataset to show as plot}
\usage{
create_plot_data(data, type, xlim, ylim, bwd, show.excluded, show.colorbar,
  show.labels, show.lines, show.axes, labels)
}
\arguments{
\item{data}{dataset (vector, matrix or data frame)}

\item{type}{type of the plot}

\item{xlim}{limits for x-axis provided by user}

\item{ylim}{limits for y-axis provided by user}

\item{bwd}{width of bars provided by user}

\item{show.excluded}{logical, show or not excluded values}

\item{show.colorbar}{logical, show or not colorbar (needed to compute limits correctly)}

\item{show.labels}{logical, show or not labels for data points}

\item{show.lines}{logical, show or not lines}

\item{show.axes}{logical, show or not axes}

\item{labels}{vector with labels}
}
\description{
Create dataset to show as plot
}
