% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval.getParams}
\alias{crossval.getParams}
\title{Define parameters based on 'cv' value}
\usage{
crossval.getParams(cv, nobj)
}
\arguments{
\item{cv}{settings for cross-validation provided by user}

\item{nobj}{number of objects in calibration set}
}
\description{
Define parameters based on 'cv' value
}
