% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodel.R
\name{crossval.regmodel}
\alias{crossval.regmodel}
\title{Cross-validation of a regression model}
\usage{
crossval.regmodel(obj, x, y, cv, cal.fun)
}
\arguments{
\item{obj}{a regression model (object of class \code{regmodel})}

\item{x}{a matrix with x values (predictors from calibration set)}

\item{y}{a matrix with y values (responses from calibration set)}

\item{cv}{number of segments (if cv = 1, full cross-validation will be used)}

\item{cal.fun}{reference to function for model calibration}
}
\value{
object of class \code{plsres} with results of cross-validation
}
\description{
Does cross-validation of a regression model
}
