% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simca.R
\name{crossval.simca}
\alias{crossval.simca}
\title{Cross-validation of a SIMCA model}
\usage{
crossval.simca(obj, x, cv)
}
\arguments{
\item{obj}{a SIMCA model (object of class \code{simca})}

\item{x}{a matrix with x values (predictors from calibration set)}

\item{cv}{number of segments (if cv = 1, full cross-validation will be used)}
}
\value{
object of class \code{simcares} with results of cross-validation
}
\description{
Does the cross-validation of a SIMCA model
}
