% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval.str}
\alias{crossval.str}
\title{String with description of cross-validation method}
\usage{
crossval.str(cv)
}
\arguments{
\item{cv}{a list with cross-validation settings}
}
\value{
a string with the description text
}
\description{
String with description of cross-validation method
}
