% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{dd.crit}
\alias{dd.crit}
\title{Calculates critical limits for distance values using Data Driven moments approach}
\usage{
dd.crit(paramQ, paramT2, alpha = 0.05, gamma = 0.01)
}
\arguments{
\item{paramQ}{matrix with parameters for distribution of Q distances}

\item{paramT2}{matrix with parameters for distribution of T2 distances}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}
}
\description{
Calculates critical limits for distance values using Data Driven moments approach
}
