% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ddmoments.crit}
\alias{ddmoments.crit}
\title{Calculates critical limits for distance values using Data Driven moments approach}
\usage{
ddmoments.crit(U, ncomp, alpha = 0.05, gamma = 0.01)
}
\arguments{
\item{U}{matrix or vector with distance values}

\item{ncomp}{number of components}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}
}
\description{
Calculates critical limits for distance values using Data Driven moments approach
}
