% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ddmoments.param}
\alias{ddmoments.param}
\title{Calculates critical limits for distance values using Data Driven moments approach}
\usage{
ddmoments.param(U)
}
\arguments{
\item{U}{matrix or vector with distance values}
}
\description{
Calculates critical limits for distance values using Data Driven moments approach
}
