% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ddrobust.crit}
\alias{ddrobust.crit}
\title{Calculates critical limits for distance values using Data Driven robust approach}
\usage{
ddrobust.crit(U, ncomp, alpha, gamma)
}
\arguments{
\item{U}{matrix or vector with distance values}

\item{ncomp}{number of components}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}
}
\description{
Calculates critical limits for distance values using Data Driven robust approach
}
