% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ddrobust.param}
\alias{ddrobust.param}
\title{Calculates critical limits for distance values using Data Driven robust approach}
\usage{
ddrobust.param(U)
}
\arguments{
\item{U}{matrix or vector with distance values}
}
\description{
Calculates critical limits for distance values using Data Driven robust approach
}
