% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{ddsimca.fromjson}
\alias{ddsimca.fromjson}
\title{Converts JSON string created in mda.tools/ddsimca app to \code{ddsimca} object}
\usage{
ddsimca.fromjson(str)
}
\arguments{
\item{str}{stringified JSON (from model file)}
}
\value{
object of \code{\link{ddsimca}} class
}
\description{
Converts JSON string created in mda.tools/ddsimca app to \code{ddsimca} object
}
