% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{ddsimca}
\alias{ddsimca}
\title{Data Driven SIMCA}
\usage{
ddsimca(
  x,
  classname,
  ncomp = min(nrow(x) - 1, ncol(x) - 1, 20),
  center = TRUE,
  scale = FALSE,
  pcv = list("ven", 10),
  alpha = 0.05,
  gamma = 0.01,
  exclrows = NULL,
  exclcols = NULL,
  prep = NULL,
  ...
)
}
\arguments{
\item{x}{a numerical matrix with data values.}

\item{classname}{short text (up to 20 symbols) with class name.}

\item{ncomp}{maximum number of components to calculate.}

\item{center}{logical, do mean centering of data or not.}

\item{scale}{logical, do standardization of data or not.}

\item{pcv}{Procrurstes cross-validation settings (see details).}

\item{alpha}{significance level for making the predictions (can be also adjusted when model is applied to data).}

\item{gamma}{significance level for detection of outliers (can be also adjusted when model is applied to data).}

\item{exclrows}{rows to be excluded from calculations (numbers, names or vector with logical values)}

\item{exclcols}{columns to be excluded from calculations (numbers, names or vector with logical values)}

\item{prep}{optional list with preprocessing methods created using `\code{\link{prep}}` function.}

\item{...}{any other parameters suitable for \code{\link{pca}} method.}
}
\value{
Returns an object of \code{ddsimca} class with following fields:
\item{classname }{a short text with class name.}
\item{calres }{an object of class \code{\link{simcares}} with classification results for a
calibration data.}
\item{pvres }{an object of class \code{\link{simcares}} with classification results for a Procrustes
validation set.}

Fields, inherited from \code{\link{pca}} class:
\item{ncomp }{number of components included to the model.}
\item{ncomp.selected }{selected (optimal) number of components.}
\item{loadings }{matrix with loading values (nvar x ncomp).}
\item{eigenvals }{vector with eigenvalues for all existent components.}
\item{expvar }{vector with explained variance for each component (in percent).}
\item{cumexpvar }{vector with cumulative explained variance for each component (in percent).}
\item{info }{information about the model, provided by user when build the model.}
}
\description{
\code{ddsimca} is used to develop DD-SIMCA (Data Driven SIMCA) model for
one-class classification.
}
\details{
DD-SIMCA is based on PCA model with additional functionality, so \code{ddsimca} class inherits most
of the functionality of \code{\link{pca}} class.

In order to make a decision, DDSIMCA uses score and orthogonal distances to PCA model. It combines
the two distances to joint full distance and uses chi-distribution for finding a critical value
which is employed as decision rule. More details about DD-SIMCA can be found in [1] (open access).

Procrustes cross-validation (PCV) is used to generate a validation set in order to find optimal model
complexity (number of components). The PCV settings are similar to the ones used for conventional
cross-validation. The best way is to set `pcv` value to a list, for example:  \code{pcv = list('ven', nseg)}
for systematic splits or \code{pcv = list('rand', nseg)} for random splits. In case if full cross-validation
must be employed, use \code{pcv = list('loo')}.
}
\examples{
## make a SIMCA model for Iris setosa class with full cross-validation
library(mdatools)

data = iris[, 1:4]
class = iris[, 5]

# take first 20 objects of setosa as calibration set
se = data[1:20, ]

# make SIMCA model and apply to test set
model = ddsimca(se, "setosa", pcv = list("ven", 10))
model = selectCompNum(model, 1)

# show infromation, summary and plot overview
print(model)
summary(model)
plot(model)


}
\references{
1. Kucheryavskiy S, Rodionova O, Pomerantsev A. A comprehensive tutorial on Data-Driven SIMCA:
Theory and implementation in web. Journal of Chemometrics. 2024; 38(7):e3556. doi:10.1002/cem.3556

2. S. Kucheryavskiy, O. Rodionova, A. Pomerantsev, Procrustes cross-validation of multivariate
regression models. Analytica Chimica Acta. 2023; 1255:341096. doi:10.1016/j.aca.2023.341096.
}
\seealso{
Methods for \code{ddsimca} objects:
\tabular{ll}{
 \code{print.ddsimca} \tab shows information about the object.\cr
 \code{summary.ddsimca} \tab shows summary statistics for the model.\cr
 \code{plot.ddsimca} \tab makes an overview of DD-SIMCA model with four plots.\cr
 \code{\link{predict.ddsimca}} \tab applies DD-SIMCA model to a new data.\cr
}

Methods, inherited from \code{classmodel} class:
\tabular{ll}{
 \code{\link{plotPredictions.classmodel}} \tab shows plot with predicted values.\cr
 \code{\link{plotSensitivity.classmodel}} \tab shows sensitivity plot.\cr
 \code{\link{plotSpecificity.classmodel}} \tab shows specificity plot.\cr
 \code{\link{plotMisclassified.classmodel}} \tab shows misclassified ratio plot.\cr
}

Methods, inherited from \code{\link{pca}} class:
\tabular{ll}{
 \code{\link{selectCompNum.pca}} \tab set number of optimal components in the model\cr
 \code{\link{plotScores.pca}} \tab shows scores plot.\cr
 \code{\link{plotLoadings.pca}} \tab shows loadings plot.\cr
 \code{\link{plotVariance.pca}} \tab shows explained variance plot.\cr
 \code{\link{plotCumVariance.pca}} \tab shows cumulative explained variance plot.\cr
}
}
\author{
Sergey Kucheryavskiy (svkucheryavski@gmail.com)
}
