% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{ddsimca.readJSON}
\alias{ddsimca.readJSON}
\title{Reads DD-SIMCA model from JSON file made in web-application (mda.tools/ddsimca).}
\usage{
ddsimca.readJSON(fileName)
}
\arguments{
\item{fileName}{file name (or full path) to JSON file.}
}
\value{
list with DD-SIMCA model similar to what \code{ddsimca()} creates.
}
\description{
Reads DD-SIMCA model from JSON file made in web-application (mda.tools/ddsimca).
}
