% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{ddsimcares}
\alias{ddsimcares}
\title{Results of DD-SIMCA one-class classification}
\usage{
ddsimcares(pcares, outcomes, indices, numbers, alpha, classname, c.ref = NULL)
}
\arguments{
\item{pcares}{results of PCA decomposition of data (class \code{pcares}).}

\item{outcomes}{outcomes of DD-SIMCA classification procedure.}

\item{indices}{list with the object indices (members, strangers, unknown).}

\item{numbers}{list with the object numbers in each subset (members, strangers, unknown).}

\item{alpha}{significance level for making the predictions (can be also adjusted when model is applied to data).}

\item{classname}{short text (up to 20 symbols) with class name.}

\item{c.ref}{optional, vector with reference classes.}
}
\value{
Returns an object (list) of class \code{ddsimcares} with the same fields as \code{\link{pcares}}
plus additional field \code{simcares} which is a list with all DD-SIMCA outcomes and related
properties:
}
\description{
@description
\code{ddsimcares} is used to store results for DD-SIMCA one-class classification. Do not create this object manually,
it will be created automatically by applying DD-SIMCA model.
}
\details{
Class \code{ddsimcares} inherits all properties and methods of class \code{\link{pcares}}, and
has additional properties and functions for representing of classification results and other
DD-SIMCA outcomes.

Do not create a \code{ddsimcares} object manually, it is created automatically when
a DD-SIMCA model is developed (see \code{\link{ddsimca}}) or when the model is applied to a new
data (see \code{\link{predict.ddsimca}}). The object can be used to show summary and plots for
the results.
}
\examples{
## make a SIMCA model for Iris setosa class and show results for calibration set
library(mdatools)

data = iris[, 1:4]
class = iris[, 5]

# take first 30 objects of setosa as calibration set
se = data[1:30, ]

# make SIMCA model and apply to test set
model = ddsimca(se, 'Se')
model = selectCompNum(model, 1)

# show infromation and summary
print(model$calres)
summary(model$calres)


}
\seealso{
Methods specific for \code{ddsimcares} objects:
\tabular{ll}{
 \code{print.ddsimcares} \tab shows information about the object.\cr
 \code{summary.ddsimcares} \tab shows statistics for results of classification.\cr
 \code{as.data.frame.ddsimcares} \tab shows converts DD-SIMCA results into data frame.\cr
 \code{as.matrix.ddsimcares} \tab shows converts summary of DD-SIMCA results into matrix.\cr
 \code{writeCSV.ddsimcares} \tab saves DD-SIMCA results into a CSV file.\cr
 \code{plotAcceptance.ddsimcares} \tab shows acceptance plot (q/q0 vs h/h0) with decision and outlier boundaries.\cr
}

Methods, inherited from \code{\link{ldecomp}} class:
\tabular{ll}{
 \code{\link{plotResiduals.ldecomp}} \tab makes Q2 vs. T2 residuals plot.\cr
 \code{\link{plotScores.ldecomp}} \tab makes scores plot.\cr
 \code{\link{plotVariance.ldecomp}} \tab makes explained variance plot.\cr
 \code{\link{plotCumVariance.ldecomp}} \tab makes cumulative explained variance plot.\cr
}
Check also \code{\link{simca}} and \code{\link{pcares}}.
}
