% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{ellipse}
\alias{ellipse}
\title{Create ellipse on the current plot}
\usage{
ellipse(xc = 0, yc = 0, a, b, col = "black", lty = 1, ...)
}
\arguments{
\item{xc}{coordinate of center (x)}

\item{yc}{coordinate of center (y)}

\item{a}{major axis}

\item{b}{minor axis}

\item{col}{color of the ellipse line}

\item{lty}{type of the ellipse line}

\item{...}{any argument suitable for \code{lines} function}
}
\description{
Create ellipse on the current plot
}
