% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{employ.constraint}
\alias{employ.constraint}
\title{Applies constraint to a dataset}
\usage{
\method{employ}{constraint}(obj, x, d)
}
\arguments{
\item{obj}{object with constraint}

\item{x}{matrix with pure spectra or contributions}

\item{d}{matrix with original spectral values}
}
\description{
Applies constraint to a dataset
}
