% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{employ.prep}
\alias{employ.prep}
\title{Applies a list with preprocessing methods to a dataset}
\usage{
\method{employ}{prep}(obj, data)
}
\arguments{
\item{obj}{list with preprocssing methods (created using \code{prep} function).}

\item{data}{matrix with dataset}
}
\description{
Applies a list with preprocessing methods to a dataset
}
