% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{employ}
\alias{employ}
\title{Generic function to apply a method (e.g. constraint)}
\usage{
employ(obj, x)
}
\arguments{
\item{obj}{constraint object}

\item{x}{data matrix}
}
\description{
Generic function to apply a method (e.g. constraint)
}
