% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{extract_array}
\alias{extract_array}
\title{Extract numeric array from JSON string}
\usage{
extract_array(js, key)
}
\arguments{
\item{js}{stringified JSON.}

\item{key}{name of the element.}
}
\value{
array of numbers.
}
\description{
Extract numeric array from JSON string
}
