% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{extract_string_array}
\alias{extract_string_array}
\title{Extract string array from JSON string}
\usage{
extract_string_array(js, key)
}
\arguments{
\item{js}{stringified JSON.}

\item{key}{name of the element.}
}
\value{
array of strings.
}
\description{
Extract string array from JSON string
}
