% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{extract_value}
\alias{extract_value}
\title{Extract single value from JSON string}
\usage{
extract_value(js, key)
}
\arguments{
\item{js}{stringified JSON.}

\item{key}{name of the element.}
}
\value{
value
}
\description{
Extract single value from JSON string
}
