% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extractArray}
\alias{extractArray}
\title{Extract numeric array from JSON string}
\usage{
extractArray(js, key)
}
\arguments{
\item{js}{stringified JSON.}

\item{key}{name of the element.}
}
\value{
array of numbers.
}
\description{
Extract numeric array from JSON string
}
