% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extractBlock}
\alias{extractBlock}
\title{Extracts a JSON subset in main JSON structure}
\usage{
extractBlock(js, name)
}
\arguments{
\item{js}{stringified JSON with model}

\item{name}{name of the JSON block to extract}
}
\value{
string with part of the JSON string related to preprocessing
}
\description{
Extracts a JSON subset in main JSON structure
}
