% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extractPrep}
\alias{extractPrep}
\title{Extracts JSON related to preprocessing model}
\usage{
extractPrep(js)
}
\arguments{
\item{js}{stringified JSON with model}
}
\value{
string with part of the JSON string related to preprocessing
}
\description{
Extracts JSON related to preprocessing model
}
