% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extractStringArray}
\alias{extractStringArray}
\title{Extract string array from JSON string}
\usage{
extractStringArray(js, key)
}
\arguments{
\item{js}{stringified JSON.}

\item{key}{name of the element.}
}
\value{
array of strings.
}
\description{
Extract string array from JSON string
}
