% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extractValue}
\alias{extractValue}
\title{Extract single value from JSON string}
\usage{
extractValue(js, key)
}
\arguments{
\item{js}{stringified JSON.}

\item{key}{name of the element.}
}
\value{
value
}
\description{
Extract single value from JSON string
}
