% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{genhash}
\alias{genhash}
\title{Generates unique pseudo-hash number based on current time and date}
\usage{
genhash()
}
\value{
string with the hash
}
\description{
Generates unique pseudo-hash number based on current time and date
}
