% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{get_convex_hull}
\alias{get_convex_hull}
\title{Compute coordinates of a closed convex hull for data points}
\usage{
get_convex_hull(points)
}
\arguments{
\item{points}{matrix of data frame with coordinates of the points}
}
\description{
Compute coordinates of a closed convex hull for data points
}
