% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getCalibrationData.pca}
\alias{getCalibrationData.pca}
\title{Returns matrix with original calibration data}
\usage{
\method{getCalibrationData}{pca}(obj)
}
\arguments{
\item{obj}{object with PCA model}
}
\description{
Returns matrix with original calibration data
}
