% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcam.R
\name{getCalibrationData.simcam}
\alias{getCalibrationData.simcam}
\title{Get calibration data}
\usage{
\method{getCalibrationData}{simcam}(obj)
}
\arguments{
\item{obj}{SIMCAM model (object of class \code{simcam})}
}
\description{
Get data, used for calibration of the SIMCAM individual models and combine to one dataset.
}
\details{
See examples in help for \code{\link{simcam}} function.
}
