% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{getConfidenceEllipse}
\alias{getConfidenceEllipse}
\title{Compute confidence ellipse for a set of points}
\usage{
getConfidenceEllipse(points, conf.level = 0.95, n = 100)
}
\arguments{
\item{points}{matrix of data frame with coordinates of the points}

\item{conf.level}{confidence level for the ellipse}

\item{n}{number of points in the ellipse coordinates}
}
\value{
matrix with coordinates of the ellipse points (x and y)
}
\description{
Compute confidence ellipse for a set of points
}
