% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classres.R
\name{getConfusionMatrix.classres}
\alias{getConfusionMatrix.classres}
\title{Confusion matrix for classification results}
\usage{
\method{getConfusionMatrix}{classres}(obj, ncomp = NULL, ...)
}
\arguments{
\item{obj}{classification results (object of class \code{simcares}, \code{simcamres}, etc)}

\item{ncomp}{number of components to make the matrix for (NULL - use selected for a model).}

\item{...}{other arguments}
}
\description{
The columns of the matrix correspond to classification results, rows - to the real classes. In
case of soft classification with multiple classes (e.g. SIMCAM) sum of values for every row 
will not correspond to the total number of class members as the same object can be classified
as a member of several classes or non of them.
}
\details{
Returns confusion matrix for classification results represented by the object.
}
