% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{getConfusionMatrix}
\alias{getConfusionMatrix}
\title{Confusion matrix for classification results}
\usage{
getConfusionMatrix(obj, ...)
}
\arguments{
\item{obj}{classification results (object of class \code{simcares}, \code{simcamres}, etc)}

\item{...}{other parameters.}
}
\description{
Confusion matrix for classification results
}
\details{
Returns confusion matrix for classification results represented by the object.
}
