% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{getConvexHull}
\alias{getConvexHull}
\title{Compute coordinates of a closed convex hull for data points}
\usage{
getConvexHull(points)
}
\arguments{
\item{points}{matrix of data frame with coordinates of the points}
}
\description{
Compute coordinates of a closed convex hull for data points
}
