% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{getDataLabels}
\alias{getDataLabels}
\title{Create a vector with labels for plot series}
\usage{
getDataLabels(ps, labels = NULL)
}
\arguments{
\item{ps}{`plotseries` object}

\item{labels}{vector with user defined labels or type of labels to show ("values", "names", "indices")}
}
\description{
For scatter plots labels correspond to rows of the data (names, values, indices, etc.). For
non-scatter plots labels correspond to the columns (names, indices or max value for each column)
}
