% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{getImplementedConstraints}
\alias{getImplementedConstraints}
\title{Shows a list with implemented constraints}
\usage{
getImplementedConstraints()
}
\description{
Shows a list with implemented constraints
}
