% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getImplementedPrepMethods}
\alias{getImplementedPrepMethods}
\title{Shows a list with implemented preprocessing methods}
\usage{
getImplementedPrepMethods()
}
\description{
Shows a list with implemented preprocessing methods
}
