% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{getLabelsAsIndices}
\alias{getLabelsAsIndices}
\title{Create labels as column or row indices}
\usage{
getLabelsAsIndices(ps)
}
\arguments{
\item{ps}{`plotseries` object}
}
\description{
Create labels as column or row indices
}
