% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{getLabelsAsValues}
\alias{getLabelsAsValues}
\title{Create labels from data values}
\usage{
getLabelsAsValues(ps)
}
\arguments{
\item{ps}{`plotseries` object}
}
\description{
Create labels from data values
}
