% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getLimitsCoordinates.pca}
\alias{getLimitsCoordinates.pca}
\title{Compute coordinates of lines or curves with critical limits}
\usage{
getLimitsCoordinates.pca(obj, ncomp, norm, log)
}
\arguments{
\item{obj}{object of class \code{pca}}

\item{ncomp}{number of components for computing the coordinates}

\item{norm}{logical, shall distance values be normalized or not}

\item{log}{logical, shall log transformation be applied or not}
}
\value{
list with two matrices (x and y coordinates of corresponding limits)
}
\description{
Compute coordinates of lines or curves with critical limits
}
