\name{getMainTitle}
\alias{getMainTitle}
\title{Get main title}
\usage{
getMainTitle(main, ncomp, default)
}
\arguments{
  \item{main}{main title of a plot, provided by user}

  \item{ncomp}{number of components to select, provided by
  user}

  \item{default}{default title for the plot}
}
\description{
returns main title for a plot depending on a user choice
}
\details{
Depedning on a user choice it returns main title for a plot
}

