% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getModelRes}
\alias{getModelRes}
\title{Returns list with all model results available (cal, cv and tes)}
\usage{
getModelRes(model)
}
\arguments{
\item{model}{object with model}
}
\description{
Returns list with all model results available (cal, cv and tes)
}
