% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{getPlotColors}
\alias{getPlotColors}
\title{Define colors for plot series}
\usage{
getPlotColors(ps, col, opacity, cgroup, colmap)
}
\arguments{
\item{ps}{`plotseries` object}

\item{col}{color specified by user (if any)}

\item{opacity}{opacity for the color}

\item{cgroup}{vector for color grouping (if any)}

\item{colmap}{name or values for colormap}
}
\description{
Define colors for plot series
}
