% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getProbabilities.pca}
\alias{getProbabilities.pca}
\title{Probabilities for residual distances}
\usage{
\method{getProbabilities}{pca}(obj, ncomp, q, h, ...)
}
\arguments{
\item{obj}{object with PCA model}

\item{ncomp}{number of components to compute the probability for}

\item{q}{vector with squared orthogonal distances for given number of components}

\item{h}{vector with score distances for given number of components}

\item{...}{other parameters}
}
\description{
Probabilities for residual distances
}
\details{
Computes p-value for every object being from the same populaion as calibration set
based on its orthogonal and score distances.
}
