% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{getProbabilities}
\alias{getProbabilities}
\title{Class belongings probability for SIMCA results}
\usage{
getProbabilities(obj, ...)
}
\arguments{
\item{obj}{a SIMCA model}

\item{...}{other parameters}
}
\description{
Calculates probability of class belongings for each object from SIMCA results
}
