% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcv.R
\name{getR}
\alias{getR}
\title{Creates rotation matrix to map a set vectors \code{base1} to a set of vectors \code{base2}.}
\usage{
getR(base1, base2)
}
\arguments{
\item{base1}{Matrix (JxA) with A orthonormal vectors as columns to be rotated (A <= J)}

\item{base2}{Matrix (JxA) with A orthonormal vectors as columns, \code{base1} should be aligned with}
}
\value{
Rotation matrix (JxJ)
}
\description{
In both sets vectors should be orthonormal.
}
