% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R
\name{getRegcoeffs}
\alias{getRegcoeffs}
\title{Get regression coefficients}
\usage{
getRegcoeffs(obj, ...)
}
\arguments{
\item{obj}{a PLS model}

\item{...}{other parameters}
}
\description{
Generic function for getting regression coefficients from PLS model
}

