% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getRes}
\alias{getRes}
\title{Return list with valid results}
\usage{
getRes(res, classname = "ldecomp")
}
\arguments{
\item{res}{list with results}

\item{classname}{name of class (for result object) to look for}
}
\description{
Return list with valid results
}
