\name{getSelectedComponents.classres}
\alias{getSelectedComponents.classres}
\title{Get selected components}
\usage{
getSelectedComponents.classres(obj, ncomp = NULL)
}
\arguments{
  \item{obj}{object with classification results (e.g.
  \code{plsdares} or \code{simcamres}).}

  \item{ncomp}{number of components specified by user.}
}
\description{
Returns number of components depending on user selection
and object properites
}
\details{
This is a technical function used for selection proper
value for number of components in plotting functions.
}

