% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{getSelectivityRatio.pls}
\alias{getSelectivityRatio.pls}
\title{Selectivity ratio for PLS model}
\usage{
\method{getSelectivityRatio}{pls}(obj, ncomp = obj$ncomp.selected, ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{ncomp}{number of components to get the values for (if NULL user selected as optimal will be used)}

\item{...}{other parameters}
}
\value{
vector with selectivity ratio values
}
\description{
Returns vector with Selectivity ratio values. This function is a proxy for \code{\link{selratio}}
and will be removed in future releases.
}
\references{
[1] Tarja Rajalahti et al. Chemometrics and Laboratory Systems, 95 (2009), pp. 35-48.
}
