% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{getSelectivityRatio}
\alias{getSelectivityRatio}
\title{Selectivity ratio}
\usage{
getSelectivityRatio(obj, ...)
}
\arguments{
\item{obj}{a regression model}

\item{...}{other parameters}
}
\description{
Generic function for returning selectivity ratio values for regression model (PCR, PLS, etc)
}

