% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{getVariance.mcr}
\alias{getVariance.mcr}
\title{Compute explained variance for MCR case}
\usage{
getVariance.mcr(obj, x)
}
\arguments{
\item{obj}{object of class \code{mcr}}

\item{x}{original spectral data}
}
\description{
Compute explained variance for MCR case
}
