% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{getVIPScores.pls}
\alias{getVIPScores.pls}
\title{VIP scores for PLS model}
\usage{
\method{getVIPScores}{pls}(obj, ncomp = obj$ncomp.selected, ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{ncomp}{number of components to count}

\item{...}{other parameters}
}
\value{
matrix \code{nvar x 1} with VIP score values
}
\description{
Returns vector with VIP scores values. This function is a proxy for \code{\link{vipscores}}
and will be removed in future releases.
}
