% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{getVIPScores}
\alias{getVIPScores}
\title{VIP scores}
\usage{
getVIPScores(obj, ...)
}
\arguments{
\item{obj}{a regression model}

\item{...}{other parameters}
}
\description{
Generic function for returning VIP scores values for regression model (PCR, PLS, etc)
}

