% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hotelling.prob}
\alias{hotelling.prob}
\title{Calculate probabilities for distance values and given parameters using Hotelling T2 distribution}
\usage{
hotelling.prob(U, ncomp, lim)
}
\arguments{
\item{U}{matrix or vector with distances}

\item{ncomp}{number of components}

\item{lim}{vector with parameters}
}
\description{
Calculate probabilities for distance values and given parameters using Hotelling T2 distribution
}
