% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{imshow}
\alias{imshow}
\title{show image data as an image}
\usage{
imshow(data, channels = 1, show.excluded = FALSE, main = NULL)
}
\arguments{
\item{data}{data with image}

\item{channels}{indices for one or three columns to show as image channels}
}
\description{
show image data as an image
}

