% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipls.R
\name{ipls.backward}
\alias{ipls.backward}
\title{Runs the backward iPLS algorithm}
\usage{
ipls.backward(x, y, obj, int.stat, glob.stat)
}
\arguments{
\item{x}{a matrix with predictor values}

\item{y}{a vector with response values}

\item{obj}{object with initial settings for iPLS algorithm}

\item{int.stat}{data frame with initial interval statistics}

\item{glob.stat}{data frame with initial global statistics}
}
\description{
Runs the backward iPLS algorithm
}
