% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{jm.crit}
\alias{jm.crit}
\title{Calculate critical limits for distance values using Jackson-Mudholkar approach}
\usage{
jm.crit(residuals, eigenvals, alpha = 0.05, gamma = 0.01)
}
\arguments{
\item{residuals}{matrix with PCA residuals}

\item{eigenvals}{vector with eigenvalues for PCA components}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}
}
\value{
vector with four values: critical limits for given alpha and gamma, mean distance and DoF.
}
\description{
Calculate critical limits for distance values using Jackson-Mudholkar approach
}
