% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{jm.prob}
\alias{jm.prob}
\title{Calculate probabilities for distance values and given parameters using Hotelling T2 distribution}
\usage{
jm.prob(u, eigenvals, ncomp)
}
\arguments{
\item{u}{vector with distances}

\item{eigenvals}{vector with eigenvalues for PCA components}

\item{ncomp}{number of components}
}
\description{
Calculate probabilities for distance values and given parameters using Hotelling T2 distribution
}
