% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R, R/pca.R
\name{ldecomp.getDistances}
\alias{ldecomp.getDistances}
\title{Compute score and residual distances}
\usage{
ldecomp.getDistances(scores, loadings, residuals, tnorm = NULL)

ldecomp.getDistances(scores, loadings, residuals, tnorm = NULL)
}
\arguments{
\item{scores}{matrix with scores (T).}

\item{loadings}{matrix with loadings (P).}

\item{residuals}{matrix with residuals (E).}

\item{tnorm}{vector with singular values for scores normalisation (can be provided)}

\item{scores}{matrix with scores (T).}

\item{loadings}{matrix with loadings (P).}

\item{residuals}{matrix with residuals (E).}

\item{tnorm}{vector with singular values for scores normalisation (can be provided)}
}
\value{
Returns a list with Q, T2 and tnorm values for each component.

Returns a list with Q, T2 and tnorm values for each component.
}
\description{
Compute orthogonal Euclidean distance from object to PC space (Q, q) and Mahalanobis
squared distance between projection of the object to the space and its origin (T2, h).

Compute orthogonal Euclidean distance from object to PC space (Q, q) and Mahalanobis
squared distance between projection of the object to the space and its origin (T2, h).
}
\details{
The distances are calculated for every 1:n components, where n goes from 1 to ncomp
(number of columns in scores and loadings).

The distances are calculated for every 1:n components, where n goes from 1 to ncomp
(number of columns in scores and loadings).
}
