% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.getLimits}
\alias{ldecomp.getLimits}
\title{Compute critical limits for orthogonal distance}
\usage{
ldecomp.getLimits(Q, T2, alpha, gamma, lim.type)
}
\arguments{
\item{Q}{matrix with distances}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}

\item{lim.type}{which method to use ("chisq", "ddrobust", "ddmoments")}
}
\description{
Compute critical values for extremes and outliers based on the matrix with distances.

If data driven approach is used ("ddrobust" or "ddmoments") then instead of the limit,
number of degrees of freedom (Nq) and scaling factor (q0) are computed instead.
}
