% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.getQLimits}
\alias{ldecomp.getQLimits}
\title{Compute critical limits for orthogonal distances (Q)}
\usage{
ldecomp.getQLimits(lim.type, alpha, gamma, params, residuals, eigenvals)
}
\arguments{
\item{lim.type}{which method to use for calculation of critical limits for residuals}

\item{alpha}{significance level for extreme limits.}

\item{gamma}{significance level for outlier limits.}

\item{params}{distribution parameters returned by ldecomp.getLimParams}

\item{residuals}{matrix with residuals (E)}

\item{eigenvals}{egenvalues for the components used to decompose the data}
}
\description{
Compute critical limits for orthogonal distances (Q)
}
