% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.getResLimits}
\alias{ldecomp.getResLimits}
\title{Statistical limits for Q and T2 residuals}
\usage{
ldecomp.getResLimits(model, lim.type = "classic", alpha = 0.05,
  gamma = 0.01)
}
\arguments{
\item{model}{object with PCA model}

\item{lim.type}{which method to use for calculation of the limits (see \code{pca} for details)}

\item{alpha}{significance level for detection of extreme objects}

\item{gamma}{significance level for detection of outliers (for data driven approach)}
}
\value{
Returns a list with two vectors:  \code{T2lim} and \code{Qlim}.
}
\description{
Computes statisticsl limits for Q and T2 residuals
}
\details{
See help for \code{\link{pca}} for details.
}
