% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.getT2Limits}
\alias{ldecomp.getT2Limits}
\title{Compute critical limits for score distances (T2)}
\usage{
ldecomp.getT2Limits(lim.type, alpha, gamma, params)
}
\arguments{
\item{lim.type}{which method to use for calculation ("chisq", "ddmoments", "ddrobust")}

\item{alpha}{significance level for extreme limits.}

\item{gamma}{significance level for outlier limits.}

\item{params}{distribution parameters returned by ldecomp.getLimParams}
}
\description{
Compute critical limits for score distances (T2)
}
