% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.plotLimits}
\alias{ldecomp.plotLimits}
\title{Shows lines with critical limits on residuals plot}
\usage{
ldecomp.plotLimits(lim, lim.type, lim.col, lim.lwd, lim.lty)
}
\arguments{
\item{lim}{matrix with residual limits (2x2)}

\item{lim.type}{type of limits}

\item{lim.col}{vector with two values - line color for extreme and outlier borders}

\item{lim.lwd}{vector with two values - line width for extreme and outlier borders}

\item{lim.lty}{vector with two values - line type for extreme and outlier borders}
}
\description{
Shows lines with critical limits on residuals plot
}
