% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.plotResiduals}
\alias{ldecomp.plotResiduals}
\title{Residuals distance plot for a set of ldecomp objects (legacy, use \code{\link{ldecomp.plotDistances}} instead).}
\usage{
ldecomp.plotResiduals(res, ...)
}
\arguments{
\item{res}{list with result objects to show the plot for}

\item{...}{other parameters}
}
\description{
Shows a plot with score (T2, h) vs orthogonal (Q, q) distances and corresponding critical
limits for given number of components.
}
