% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{mcr}
\alias{mcr}
\title{General class for Multivariate Curve Resolution model}
\usage{
mcr(x, ncomp, method, exclrows = NULL, exclcols = NULL, info = "", ...)
}
\arguments{
\item{x}{spectra of mixtures (as matrix or data frame)}

\item{ncomp}{number of pure components to resolve}

\item{method}{function for computing spectra of pure components}

\item{exclrows}{rows to be excluded from calculations (numbers, names or vector with logical values)}

\item{exclcols}{columns to be excluded from calculations (numbers, names or vector with logical values)}

\item{info}{text with information about the MCR model}

\item{...}{other parameters realted to specific method}
}
\description{
\code{mcr} is used to store and visualise general MCR data and results.
}
