% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrals.R
\name{mcrals.cal}
\alias{mcrals.cal}
\title{MCR-ALS calibration}
\usage{
mcrals.cal(
  D,
  ncomp,
  cont.constraints,
  spec.constraints,
  spec.ini,
  cont.forced,
  spec.forced,
  cont.solver,
  spec.solver,
  max.niter,
  tol,
  verbose
)
}
\arguments{
\item{D}{matrix with the spectra}

\item{ncomp}{number of pure components}

\item{cont.constraints}{a list with constraints to be applied to contributions (see details).}

\item{spec.constraints}{a list with constraints to be applied to spectra (see details).}

\item{spec.ini}{a matrix with initial estimation of the pure components spectra.}

\item{cont.forced}{a matrix that allows forcing some of the concentration values (see details).}

\item{spec.forced}{a matrix that allows forcing some of the spectra values (see details).}

\item{cont.solver}{which function to use as a solver for resolving of pure components contributions (see details).}

\item{spec.solver}{which function to use as a solver for resolving of pure components spectra (see details).}

\item{max.niter}{maximum number of iterations.}

\item{tol}{tolerance, when explained variance change is smaller than this value, iterations stop.}

\item{verbose}{logical, if TRUE information about every iteration will be shown.}
}
\value{
The function returns a list with the following fields:
\item{ncomp }{number of pure components.}
\item{resspec}{matrix with resolved spectra.}
\item{rescont}{matrix with resolved contributions.}
\item{cont.constraints}{list with contribution constraints provided by user.}
\item{spec.constraints}{list with spectra constraints provided by user.}
\item{max.niter}{maximum number of iterations}
}
\description{
The method resolves spectral data into pure component spectra and contributions
using the Alternating Least Squares algorithm with constraints.
}
