% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrals.R
\name{mcrals.ols}
\alias{mcrals.ols}
\title{Ordinary least squares}
\usage{
mcrals.ols(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\description{
Ordinary least squares
}
\details{
Computes OLS solution for A = CB
}
