% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.cbind}
\alias{mda.cbind}
\title{A wrapper for cbind() method with proper set of attributes}
\usage{
mda.cbind(...)
}
\arguments{
\item{...}{datasets (data frames or matrices) to bind}
}
\value{
the merged datasets
}
\description{
A wrapper for cbind() method with proper set of attributes
}
