% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.exclcols}
\alias{mda.exclcols}
\title{Exclude/hide columns in a dataset}
\usage{
mda.exclcols(x, ind)
}
\arguments{
\item{x}{dataset (data frame or matrix).}

\item{ind}{indices of columns to exclude (numbers, names or logical values)}
}
\value{
dataset with excluded columns
}
\description{
Exclude/hide columns in a dataset
}
\details{
The method assign attribute 'exclcols', which contains number of columns, which should be excluded/hidden
from calculations and plots (without removing them physically). The argument \code{ind} should contain 
column numbers (excluding already hidden), names or logical values.
}

