% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.exclrows}
\alias{mda.exclrows}
\title{Exclude/hide rows in a dataset}
\usage{
mda.exclrows(x, ind)
}
\arguments{
\item{x}{dataset (data frame or matrix).}

\item{ind}{indices of rows to exclude (numbers, names or logical values)}
}
\value{
dataset with excluded rows
}
\description{
Exclude/hide rows in a dataset
}
\details{
The method assign attribute 'exclrows', which contains number of rows, which should be excluded/hidden
from calculations and plots (without removing them physically). The argument \code{ind} should contain 
rows numbers (excluding already hidden), names or logical values.
}
