% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.getattr}
\alias{mda.getattr}
\title{Get data attributes}
\usage{
mda.getattr(x)
}
\arguments{
\item{x}{a dataset}
}
\description{
Returns a list with important data attributes (name, xvalues, excluded rows and columns, etc.)
}

