% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.getexclind}
\alias{mda.getexclind}
\title{Get indices of excluded rows or columns}
\usage{
mda.getexclind(excl, names, n)
}
\arguments{
\item{excl}{vector with excluded values (logical, text or numbers)}

\item{names}{vector with names for rows or columns}

\item{n}{number of rows or columns}
}
\description{
Get indices of excluded rows or columns
}
