% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.inclcols}
\alias{mda.inclcols}
\title{Include/unhide the excluded columns}
\usage{
mda.inclcols(x, ind)
}
\arguments{
\item{x}{dataset (data frame or matrix).}

\item{ind}{number of excluded columns to include}
}
\value{
dataset with included columns.
}
\description{
include colmns specified by user (earlier excluded using mda.exclcols)
}

