% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.inclrows}
\alias{mda.inclrows}
\title{include/unhide the excluded rows}
\usage{
mda.inclrows(x, ind)
}
\arguments{
\item{x}{dataset (data frame or matrix).}

\item{ind}{number of excluded rows to include}
}
\value{
dataset with included rows
}
\description{
include rows specified by user (earlier excluded using mda.exclrows)
}

