% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.purge}
\alias{mda.purge}
\title{Removes excluded (hidden) rows and colmns from data}
\usage{
mda.purge(data)
}
\arguments{
\item{data}{data frame or matrix with data}
}
\description{
Removes excluded (hidden) rows and colmns from data
}
