% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.purgeCols}
\alias{mda.purgeCols}
\title{Removes excluded (hidden) colmns from data}
\usage{
mda.purgeCols(data)
}
\arguments{
\item{data}{data frame or matrix with data}
}
\description{
Removes excluded (hidden) colmns from data
}
