% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.purgeRows}
\alias{mda.purgeRows}
\title{Removes excluded (hidden) rows from data}
\usage{
mda.purgeRows(data)
}
\arguments{
\item{data}{data frame or matrix with data}
}
\description{
Removes excluded (hidden) rows from data
}
