% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.rbind}
\alias{mda.rbind}
\title{A wrapper for rbind() method with proper set of attributes}
\usage{
mda.rbind(...)
}
\arguments{
\item{...}{datasets (data frames or matrices) to bind}
}
\value{
the merged datasets
}
\description{
A wrapper for rbind() method with proper set of attributes
}
