% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.setattr}
\alias{mda.setattr}
\title{Set data attributes}
\usage{
mda.setattr(x, attrs, type = "all")
}
\arguments{
\item{x}{a dataset}

\item{attrs}{list with attributes}

\item{type}{a text variable telling which attributes to set ('all', 'row', 'col')}
}
\description{
Set most important data attributes (name, xvalues, excluded rows and columns, etc.) to a dataset
}
