% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.setimbg}
\alias{mda.setimbg}
\title{Remove background pixels from image data}
\usage{
mda.setimbg(data, bgpixels)
}
\arguments{
\item{data}{a matrix with image data}

\item{bgpixels}{vector with indices or logical values corresponding to background pixels}
}
\description{
Remove background pixels from image data
}
