% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.subset}
\alias{mda.subset}
\title{A wrapper for subset() method with proper set of attributed}
\usage{
mda.subset(x, subset = NULL, select = NULL)
}
\arguments{
\item{x}{dataset (data frame or matrix)}

\item{subset}{which rows to keep (indices, names or logical values)}

\item{select}{which columns to select (indices, names or logical values)}
}
\value{
a data with the subset
}
\description{
A wrapper for subset() method with proper set of attributed
}
\details{
The method works similar to the standard \code{subset()} method, with minor differences. First of all 
it keeps (and correct, if necessary) all important attributes. If only columns are selected, it keeps 
all excluded rows as excluded. If only rows are selected, it keeps all excluded columns. If both rows
and columns are selected it removed all excluded elements first and then makes the subset.

The parameters \code{subset} and \code{select} may each be a vector with numbers or nanes without excluded 
elements, or a logical expression.
}

