% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mda.t}
\alias{mda.t}
\title{A wrapper for t() method with proper set of attributes}
\usage{
mda.t(x)
}
\arguments{
\item{x}{dataset (data frames or matrices) to transpose}
}
\value{
the transposed dataset
}
\description{
A wrapper for t() method with proper set of attributes
}

