% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.areColors}
\alias{mdaplot.areColors}
\title{Check color values}
\usage{
mdaplot.areColors(palette)
}
\arguments{
\item{palette}{vector with possibly color values (names, RGB, etc.)}
}
\description{
Checks if elements of argument are valid color values
}
