% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.getAxesLimits}
\alias{mdaplot.getAxesLimits}
\title{Calculate axes limits}
\usage{
mdaplot.getAxesLimits(plot_data, show.colorbar = F, show.lines = F,
  show.labels = F, show.excluded = F)
}
\arguments{
\item{plot_data}{a list with plot data and related parameters returned by `mdaplot.createPlotData()`.}

\item{show.colorbar}{logical, show or not the colorbar on the plot.}

\item{show.lines}{logical or numeric with line coordinates to be shown on the plot.}

\item{show.labels}{logical, show or not labels for the data objects}

\item{show.excluded}{logical, show or not excluded values}
}
\value{
Returns a list with four limits for the x and y axes.
}
\description{
Calculates axes limits depending on data values that have to be plotted,
extra plot elements that have to be shown and margins.
}
\details{
Data can be a list with several matrices or just one matrix. The matrices can have single.x
configuration, where first column is x values and the others are y values or normal
configuration, where every odd column is x values and every even is corresponding y values.
}
