% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.getXAxisLim}
\alias{mdaplot.getXAxisLim}
\title{Calculate limits for x-axis.}
\usage{
mdaplot.getXAxisLim(
  ps,
  xlim,
  show.labels = FALSE,
  show.lines = FALSE,
  show.excluded = FALSE,
  bwd = 0.8
)
}
\arguments{
\item{ps}{`plotseries` object.}

\item{xlim}{limits provided by user}

\item{show.labels}{logical, will data labels be shown on the plot}

\item{show.lines}{logical or numeric with line coordinates to be shown on the plot.}

\item{show.excluded}{logical, will excluded values be shown on the plot}

\item{bwd}{if limits are computed for bar plot, this is a bar width (otherwise NULL)}
}
\value{
Returns a vector with two limits.
}
\description{
Calculates limits for x-axis depending on data values that have to be plotted,
extra plot elements that have to be shown and margins.
}
