% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.getXTickLabels}
\alias{mdaplot.getXTickLabels}
\title{Prepare xticklabels for plot}
\usage{
mdaplot.getXTickLabels(xticklabels, xticks, excluded_cols)
}
\arguments{
\item{xticklabels}{xticklables provided by user (if any)}

\item{xticks}{xticks (provided or computed)}

\item{excluded_cols}{columns excluded from plot data (if any)}
}
\description{
Prepare xticklabels for plot
}
