% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.getXTicks}
\alias{mdaplot.getXTicks}
\title{Prepare xticks for plot}
\usage{
mdaplot.getXTicks(xticks, x_values, xlim, type)
}
\arguments{
\item{xticks}{xticks provided by user (if any)}

\item{x_values}{x values for the plot data object}

\item{xlim}{limits for x axis}

\item{type}{type of the plot}
}
\description{
Prepare xticks for plot
}
