% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.getYAxisLim}
\alias{mdaplot.getYAxisLim}
\title{Calculate limits for y-axis.}
\usage{
mdaplot.getYAxisLim(
  ps,
  ylim,
  show.lines = FALSE,
  show.excluded = FALSE,
  show.labels = FALSE,
  show.colorbar = FALSE
)
}
\arguments{
\item{ps}{`plotseries` object.}

\item{ylim}{limits provided by user}

\item{show.lines}{logical or numeric with line coordinates to be shown on the plot.}

\item{show.excluded}{logical, will excluded values be shown on the plot}

\item{show.labels}{logical, will data labels be shown on the plot}

\item{show.colorbar}{logical, will colorbar be shown on the plot}
}
\value{
Returns a vector with two limits.
}
\description{
Calculates limits for y-axis depending on data values that have to be plotted,
extra plot elements that have to be shown and margins.
}
