% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.getYTickLabels}
\alias{mdaplot.getYTickLabels}
\title{Prepare yticklabels for plot}
\usage{
mdaplot.getYTickLabels(yticklabels, yticks)
}
\arguments{
\item{yticklabels}{yticklabels provided by user (if any)}

\item{yticks}{yticks (provided or computed)}
}
\description{
Prepare yticklabels for plot
}
