% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.getYTicks}
\alias{mdaplot.getYTicks}
\title{Prepare yticks for plot}
\usage{
mdaplot.getYTicks(yticks, y_values, ylim, type)
}
\arguments{
\item{yticks}{yticks provided by user (if any)}

\item{y_values}{y values for the plot data object}

\item{ylim}{limits for y axis}

\item{type}{type of the plot}
}
\description{
Prepare yticks for plot
}
