% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.plotAxes}
\alias{mdaplot.plotAxes}
\title{Create axes plane}
\usage{
mdaplot.plotAxes(xticklabels = NULL, yticklabels = NULL, lim = NULL,
  main = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{xticklabels}{labels for x ticks}

\item{yticklabels}{labels for y ticks}

\item{lim}{vector with limits for x and y axis}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}
}
\description{
Creates an empty axes plane for given parameters
}

