% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.plotBars}
\alias{mdaplot.plotBars}
\title{Show bars on axes}
\usage{
mdaplot.plotBars(x, y, col = NULL, bwd = 0.8, border = NA,
  force.x.values = NA)
}
\arguments{
\item{x}{vector with x values (centers of bars)}

\item{y}{vector with y values (height of bars)}

\item{col}{colors of the bars}

\item{bwd}{width of the bars (as a ratio for max width)}

\item{border}{color of bar edges}

\item{force.x.values}{vector with corrected x-values for a bar plot (do not specify this manually).}
}
\description{
Shows bars (bar plot) on predefined axes
}
