% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.plotDensity}
\alias{mdaplot.plotDensity}
\title{Make density plot using hexagon binning}
\usage{
mdaplot.plotDensity(x, y, nbins = 60, colmap = "default")
}
\arguments{
\item{x}{vector with x coordinates of data points}

\item{y}{vector with y coordinates of data points}

\item{nbins}{number of bins to split x and y axes into}

\item{colmap}{colormap to use for showing points density}
}
\description{
Make density plot using hexagon binning
}
