% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.plotLines}
\alias{mdaplot.plotLines}
\title{Show set of lines on a plot}
\usage{
mdaplot.plotLines(x, y, x.excluded, y.excluded, type, pch, col,
  col.excluded, ...)
}
\arguments{
\item{x}{vector with x values}

\item{y}{matrix with y values}

\item{x.excluded}{vector with x values for excluded rows}

\item{y.excluded}{matrix with y values for excluded rows}

\item{type}{type of the plot ("l" or "b").}

\item{pch}{a character for markers (same as \code{plot} parameter).}

\item{col}{a color for markers or lines (same as \code{plot} parameter).}

\item{col.excluded}{color for the excluded lines.}

\item{...}{other arguments for function `points()`.}
}
\description{
Show line plot on predefined axes
}
