% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.plotPointsShape}
\alias{mdaplot.plotPointsShape}
\title{Add confidence ellipse or convex hull for group of points}
\usage{
mdaplot.plotPointsShape(plot_data, lwd, lty, conf.level, opacity,
  shape_function, ...)
}
\arguments{
\item{plot_data}{plot data returned by function `mdaplot()`}

\item{lwd}{thickness of line used to show the hull}

\item{lty}{type of line used to show the hull}

\item{conf.level}{confidence level to make the ellipse for (between 0 and 1)}

\item{opacity}{of opacity is larger than 0 a semi-transparent polygon is shown over points}

\item{shape_function}{function which calculates and return coordinates of the shape}

\item{...}{extra parameters for shape_function}
}
\description{
Add confidence ellipse or convex hull for group of points
}
