% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.plotScatter}
\alias{mdaplot.plotScatter}
\title{Show set of points on a plot}
\usage{
mdaplot.plotScatter(x, y, x.excluded, y.excluded, pch.colinv, pch, col, bg,
  col.excluded, ...)
}
\arguments{
\item{x}{vector with x values}

\item{y}{matrix with y values}

\item{x.excluded}{vector with x values for excluded rows}

\item{y.excluded}{matrix with y values for excluded rows}

\item{pch.colinv}{allows to swap values for `col` and `bg` for scatter plots with `pch` valyes from 21 to 25.}

\item{pch}{a character for markers (same as \code{plot} parameter).}

\item{col}{a color for markers or lines (same as \code{plot} parameter).}

\item{bg}{background color for scatter plots wich `pch=21:25`.}

\item{col.excluded}{color for the excluded points.}

\item{...}{other arguments for function `points()`.}
}
\description{
Show scatter plot on predefined axes
}
