% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.prepareColors}
\alias{mdaplot.prepareColors}
\title{Prepare colors based on palette and opacity value}
\usage{
mdaplot.prepareColors(palette, ncolors, opacity)
}
\arguments{
\item{palette}{vector with main colors for current pallette}

\item{ncolors}{number of colors to generate}

\item{opacity}{opacity for the colors (one value or individual for each color)}
}
\value{
vector with colors
}
\description{
Prepare colors based on palette and opacity value
}
