% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.prepareDataForPlot}
\alias{mdaplot.prepareDataForPlot}
\title{Check if data argument looks correct for creatins plots, correct if needed}
\usage{
mdaplot.prepareDataForPlot(data, type)
}
\arguments{
\item{data}{dataset provided to mdaplot function}

\item{type}{type of plot}
}
\description{
Check if data argument looks correct for creatins plots, correct if needed
}
