% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.prepareDataLabels}
\alias{mdaplot.prepareDataLabels}
\title{Create a vector with labels for plot data series}
\usage{
mdaplot.prepareDataLabels(plot_data, type, labels)
}
\arguments{
\item{plot_data}{list with data and parameters prepared for plot by `mdaplot.prepareDataForPlot()` function}

\item{type}{type of plot}

\item{labels}{vector or type of labels to show}
}
\description{
For scatter plots labels correspond to rows of the data (names, values, indices, etc.). For
non-scatter plots labels correspond to the columns (names, indices or max value for each column)
}
