% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.processExcludedRows}
\alias{mdaplot.processExcludedRows}
\title{Prepare x and y values for excluded rows}
\usage{
mdaplot.processExcludedRows(plot_data, type)
}
\arguments{
\item{plot_data}{list with data and parameters prepared for plot by `mdaplot.prepareDataForPlot()` function}

\item{type}{type of plot}
}
\description{
Prepare x and y values for excluded rows
}
