% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.showColorbar}
\alias{mdaplot.showColorbar}
\title{Plot colorbar}
\usage{
mdaplot.showColorbar(cgroup, colmap = "default", lab.col = "darkgray",
  lab.cex = 0.65)
}
\arguments{
\item{cgroup}{a vector with values used to make color grouping of the elements}

\item{colmap}{a colormap to be used for color generation}

\item{lab.col}{color for legend labels}

\item{lab.cex}{size for legend labels}
}
\description{
Shows a colorbar if plot has color grouping of elements (points or lines).
}
