% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.splitPlotData}
\alias{mdaplot.splitPlotData}
\title{Split dataset to x and y values depending on plot type}
\usage{
mdaplot.splitPlotData(plot_data, type)
}
\arguments{
\item{plot_data}{list with data and parameters prepared for plot by `mdaplot.prepareDataForPlot()` function}

\item{type}{type of plot}
}
\description{
Split dataset to x and y values depending on plot type
}
