% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplotg.R
\name{mdaplotg.getLegend}
\alias{mdaplotg.getLegend}
\title{Create and return vector with legend values}
\usage{
mdaplotg.getLegend(ps, data.names, legend = NULL)
}
\arguments{
\item{data.names}{names of the series}

\item{legend}{legend values provided by user}

\item{ngroups}{number of grops (plot series)}
}
\description{
Create and return vector with legend values
}
