% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplotg.R
\name{mdaplotg.getXLim}
\alias{mdaplotg.getXLim}
\title{Compute x-axis limits for mdaplotg}
\usage{
mdaplotg.getXLim(
  ps,
  xlim,
  show.excluded,
  show.legend,
  show.labels,
  legend.position,
  bwd = NULL
)
}
\arguments{
\item{ps}{list with plotseries}

\item{xlim}{limits provided by user}

\item{show.excluded}{logical, will excluded values also be shown}

\item{show.legend}{will legend be shown on the plot}

\item{show.labels}{will labels be shown on the plot}

\item{legend.position}{position of legend on the plot (if shown)}

\item{bwd}{size of bar for bar plot}
}
\value{
vector with two values
}
\description{
Compute x-axis limits for mdaplotg
}
