% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplotg.R
\name{mdaplotg.getYLim}
\alias{mdaplotg.getYLim}
\title{Compute y-axis limits for mdaplotg}
\usage{
mdaplotg.getYLim(
  ps,
  ylim,
  show.excluded,
  show.legend,
  legend.position,
  show.labels
)
}
\arguments{
\item{ps}{list with plotseries}

\item{ylim}{limits provided by user}

\item{show.excluded}{logical, will excluded values also be shown}

\item{show.legend}{will legend be shown on the plot}

\item{legend.position}{position of legend on the plot (if shown)}

\item{show.labels}{logical, will data ponit labels also be shown}
}
\value{
vector with two values
}
\description{
Compute y-axis limits for mdaplotg
}
