% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplotg.R
\name{mdaplotg.prepareData}
\alias{mdaplotg.prepareData}
\title{Prepare data for mdaplotg}
\usage{
mdaplotg.prepareData(data, type, groupby)
}
\arguments{
\item{data}{datasets (in form of list, matrix or data frame)}

\item{type}{vector with type for dataset}

\item{groupby}{factor or  data frame with factors - used to split data matrix into groups}
}
\value{
list of datasets

The method should prepare data as a list of datasets (matrices or data frames). One list
element will be used to create one plot series.

If `data` is matrix or data frame and not `groupby` parameter is provided, then every row
will be taken as separate set. This option is available only for line or bar plots.
}
\description{
Prepare data for mdaplotg
}
