% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplotg.R
\name{mdaplotg.processParam}
\alias{mdaplotg.processParam}
\title{Check mdaplotg parameters and replicate them if necessary}
\usage{
mdaplotg.processParam(param, name, is.type, ngroups)
}
\arguments{
\item{param}{A parameter to check}

\item{name}{name of the parameter (needed for error message)}

\item{is.type}{function to use for checking parameter type}

\item{ngroups}{number of groups (plot series)}
}
\description{
Check mdaplotg parameters and replicate them if necessary
}
