% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplotg.R
\name{mdaplotg.showLegend}
\alias{mdaplotg.showLegend}
\title{Show legend for mdaplotg}
\usage{
mdaplotg.showLegend(
  legend,
  col,
  pt.bg = NA,
  pch = NULL,
  lty = NULL,
  lwd = NULL,
  cex = 1,
  bty = "o",
  position = "topright",
  plot = TRUE,
  ...
)
}
\arguments{
\item{legend}{vector with text elements for the legend items}

\item{col}{vector with color values for the legend items}

\item{pt.bg}{vector with background colors for the legend items (e.g. for pch = 21:25)}

\item{pch}{vector with marker symbols for the legend items}

\item{lty}{vector with line types for the legend items}

\item{lwd}{vector with line width values for the legend items}

\item{cex}{vector with cex factor for the points}

\item{bty}{border type for the legend}

\item{position}{legend position ("topright", "topleft', "bottomright", "bottomleft", "top", "bottom")}

\item{plot}{logical, show legend or just calculate and return its size}

\item{...}{other parameters}
}
\description{
Shows a legend for plot elements or their groups.
}
