% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste1}
\alias{paste1}
\title{Paste values together with no separator and collapse into a single string}
\usage{
paste1(...)
}
\arguments{
\item{...}{values to paste}
}
\description{
Paste values together with no separator and collapse into a single string
}
