% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.fromjson}
\alias{pca.fromjson}
\title{Converts JSON string created in mda.tools/pca app to \code{pca} object}
\usage{
pca.fromjson(str)
}
\arguments{
\item{str}{stringified JSON (from model file)}
}
\value{
object of \code{\link{pca}} class
}
\description{
Converts JSON string created in mda.tools/pca app to \code{pca} object
}
