% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.getB}
\alias{pca.getB}
\title{Low-dimensional approximation of data matrix X}
\usage{
pca.getB(X, k = NULL, rand = NULL, dist = "unif")
}
\arguments{
\item{X}{data matrix}

\item{k}{rank of X (number of components)}

\item{rand}{a vector with two values - number of iterations (q) and oversmapling parameter (p)}

\item{dist}{distribution for generating random numbers, 'unif' or 'norm'}
}
\description{
Low-dimensional approximation of data matrix X
}
