% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.getLimParams}
\alias{pca.getLimParams}
\title{Compute parameters for critical limits based on calibration results}
\usage{
pca.getLimParams(res)
}
\arguments{
\item{res}{results returned by \code{pca.run()}}
}
\description{
Compute parameters for critical limits based on calibration results
}
