% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.getQLimits}
\alias{pca.getQLimits}
\title{Compute critical limits for orthogonal distances (Q)}
\usage{
pca.getQLimits(model, lim.type, alpha, gamma)
}
\arguments{
\item{model}{PCA model}

\item{lim.type}{which method to use for calculation of critical limits for residuals (see details)}

\item{alpha}{significance level for extreme limits.}

\item{gamma}{significance level for outlier limits.}
}
\description{
Compute critical limits for orthogonal distances (Q)
}
