% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.getT2Limits}
\alias{pca.getT2Limits}
\title{Compute critical limits for score distances (T2)}
\usage{
pca.getT2Limits(model, lim.type, alpha, gamma)
}
\arguments{
\item{model}{PCA model}

\item{lim.type}{which method to use for calculation ("chisq", "ddmoments", "ddrobust")}

\item{alpha}{significance level for extreme limits.}

\item{gamma}{significance level for outlier limits.}
}
\description{
Compute critical limits for score distances (T2)
}
