% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.readJSON}
\alias{pca.readJSON}
\title{Reads PCA model from JSON file made in web-application (mda.tools/pca).}
\usage{
pca.readJSON(fileName)
}
\arguments{
\item{fileName}{file name (or full path) to JSON file.}
}
\value{
list with PCA model similar to what \code{pca()} creates.
}
\description{
Reads PCA model from JSON file made in web-application (mda.tools/pca).
}
