% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.svd}
\alias{pca.svd}
\title{Singular Values Decomposition based PCA algorithm}
\usage{
pca.svd(x, ncomp = min(ncol(x), nrow(x) - 1))
}
\arguments{
\item{x}{a matrix with data values (preprocessed)}

\item{ncomp}{number of components to calculate}
}
\value{
a list with scores, loadings and eigenvalues for the components
}
\description{
Computes principal component space using Singular Values Decomposition
}
