% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.syncResAliases}
\alias{pca.syncResAliases}
\title{Sync calres/testres aliases with the canonical res[["cal"]]/res[["test"]] fields.}
\usage{
pca.syncResAliases(obj)
}
\arguments{
\item{obj}{model object}
}
\value{
model object with aliases updated
}
\description{
Sync calres/testres aliases with the canonical res[["cal"]]/res[["test"]] fields.
}
