\name{pellets}
\alias{pellets}
\docType{data}

\title{
Image data
}

\description{
Dataset for showing how mdatools works with images. It is an RGB image represented as 3-way array.
}

\usage{data(people)}
\format{
  a 3-way array (height x width x channels).
}

\details{
This is an image with pellets of four different colours mixed in a glas volume.
}

\keyword{datasets}
