% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prep.R
\name{pinv}
\alias{pinv}
\title{Pseudo-inverse matrix}
\usage{
pinv(data)
}
\arguments{
\item{data}{a matrix with data values to compute inverse for}
}
\description{
Computes pseudo-inverse matrix using SVD
}

