% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plot.ddsimcares}
\alias{plot.ddsimcares}
\title{Plot method for DD-SIMCA results.}
\usage{
\method{plot}{ddsimcares}(x, ...)
}
\arguments{
\item{x}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{...}{other arguments}
}
\description{
Plot method for DD-SIMCA results.
}
