% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ipls.R
\name{plot.ipls}
\alias{plot.ipls}
\title{Overview plot for iPLS results}
\usage{
\method{plot}{ipls}(x, ...)
}
\arguments{
\item{x}{a  (object of class \code{pca})}

\item{...}{other arguments}
}
\description{
Shows a plot for iPLS results.
}
\details{
See details for \code{\link{plotSelection.ipls}}.
}

