% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{plot.mcr}
\alias{plot.mcr}
\title{Plot summary for MCR model}
\usage{
\method{plot}{mcr}(x, ...)
}
\arguments{
\item{x}{\code{mcr} model object}

\item{...}{other parameters}
}
\description{
Plot summary for MCR model
}
