% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtest.R
\name{plot.randtest}
\alias{plot.randtest}
\title{Plot for randomization test results}
\usage{
\method{plot}{randtest}(x, main = "Alpha", xlab = "Components", ylab = "",
  ...)
}
\arguments{
\item{x}{results of randomization test (object of class `randtest`)}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{...}{other optional arguments}
}
\description{
Makes a bar plot with alpha values for each component.
}
\details{
See examples in help for \code{\link{randtest}} function.
}

