% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simcamres.R
\name{plot.simcamres}
\alias{plot.simcamres}
\title{Model overview plot for SIMCAM results}
\usage{
\method{plot}{simcamres}(x, ...)
}
\arguments{
\item{x}{SIMCAM results (object of class \code{simcamres})}

\item{...}{other arguments}
}
\description{
Just shows a prediction plot for SIMCAM results.
}
\details{
See examples in help for \code{\link{simcamres}} function.
}

