% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{plot.supure}
\alias{plot.supure}
\title{Overview plot for supure object}
\usage{
\method{plot}{supure}(x, ...)
}
\arguments{
\item{x}{a  (object of class \code{pca})}

\item{...}{other arguments}
}
\description{
Shows a set of plots () for .
}
\details{
See examples in help for \code{\link{supure}} function.
}

