% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{plotAcceptance.ddsimca}
\alias{plotAcceptance.ddsimca}
\title{Acceptance plot for DD-SIMCA model.}
\usage{
\method{plotAcceptance}{ddsimca}(obj, res = "cal", ...)
}
\arguments{
\item{obj}{DD-SIMCA model  (object of class \code{ddsimca}).}

\item{res}{name of the results (either 'cal' or 'pv')}

\item{...}{any parameters suitable for \code{\link{plotAcceptance.ddsimcares}}.}
}
\description{
Shows Acceptance plot for calibration of Procrustes validation results (if available).
}
