% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotAcceptance}
\alias{plotAcceptance}
\title{Acceptance plot for DDSIMCA model and results (generic function)}
\usage{
plotAcceptance(obj, ...)
}
\arguments{
\item{obj}{model or result object, e.g. \code{ddsimcares}}

\item{...}{other parameters relevant for the plot}
}
\description{
Acceptance plot for DDSIMCA model and results (generic function)
}
