% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotAliens}
\alias{plotAliens}
\title{Aliens plot for DD-SIMCA results (generic function)}
\usage{
plotAliens(obj, ...)
}
\arguments{
\item{obj}{result object, e.g. \code{ddsimcares}}

\item{...}{other parameters relevant for the plot}
}
\description{
Aliens plot for DD-SIMCA results (generic function)
}
