% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{plotBars}
\alias{plotBars}
\title{Show plot series as bars}
\usage{
plotBars(ps, col = ps$col, bwd = 0.8, border = NA, force.x.values = NA)
}
\arguments{
\item{ps}{`plotseries` object}

\item{col}{colors of the bars}

\item{bwd}{width of the bars (as a ratio for max width)}

\item{border}{color of bar edges}

\item{force.x.values}{vector with corrected x-values for a bar plot (needed for group plots, do not change manually).}
}
\description{
First row of the data matrix is taken for creating the bar series. In case of barplot
color grouping is made based on columns (not rows as for all other plots).
}
