% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotBiplot}
\alias{plotBiplot}
\title{Biplot}
\usage{
plotBiplot(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Biplot
}
\details{
Generic function for biplot
}

