% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{plotContributions.mcr}
\alias{plotContributions.mcr}
\title{Show plot with resolved contributions}
\usage{
\method{plotContributions}{mcr}(
  obj,
  comp = seq_len(obj$ncomp),
  type = "l",
  col = mdaplot.getColors(obj$ncomp),
  ...
)
}
\arguments{
\item{obj}{object of clacc \code{mcr}}

\item{comp}{vector with number of components to make the plot for}

\item{type}{type of the plot}

\item{col}{vector with colors for individual components}

\item{...}{other parameters suitable for \code{mdaplotg}}
}
\description{
Show plot with resolved contributions
}
