% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotContributions}
\alias{plotContributions}
\title{Plot resolved contributions}
\usage{
plotContributions(obj, ...)
}
\arguments{
\item{obj}{object with mcr case}

\item{...}{other parameters}
}
\description{
Plot resolved contributions
}
