% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCooman}
\alias{plotCooman}
\title{Cooman's plot}
\usage{
plotCooman(obj, ...)
}
\arguments{
\item{obj}{classification model or result object}

\item{...}{other arguments}
}
\description{
Cooman's plot
}
\details{
Generic function for Cooman's plot
}

