\name{plotCooman.simcamres}
\alias{plotCooman.simcamres}

\title{
Cooman's plot for SIMCA multiclass classification results
}

\description{
\code{plotCooman.simcamres} is used to show Cooman's plot for SIMCA multiclass classification results for two classes. 
}

\usage{
plotCooman(obj, nc = c(1, 2), ...)
}

\arguments{
  \item{obj}{an object of class \code{simcamres} (SIMCA mutliclass classification results).}
  \item{nc}{number of individual models (classes) to make the plot for.}
  \item{\dots}{most of \code{\link{mdaplotg}} parameters can be used.}
}

\details{
See examples for \code{\link{simcamres}} class.
}

\seealso{
\code{\link{simcam}} to make SIMCAM model, \code{\link{simcamres}} to store and explore SIMCA multiclass classification results.
}

\examples{
}
