% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtest.R
\name{plotCorr.randtest}
\alias{plotCorr.randtest}
\title{Correlation plot for randomization test results}
\usage{
\method{plotCorr}{randtest}(
  obj,
  ncomp = obj$ncomp.selected,
  ylim = NULL,
  xlab = expression(r^2),
  ylab = "Test statistic",
  ...
)
}
\arguments{
\item{obj}{results of randomization test (object of class `randtest`)}

\item{ncomp}{number of component to make the plot for}

\item{ylim}{limits for y axis}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{...}{other optional arguments}
}
\description{
Makes a plot with statistic values vs. coefficient of determination between permuted
and reference y-values.
}
\details{
See examples in help for \code{\link{randtest}} function.
}
