% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotCorr}
\alias{plotCorr}
\title{Correlation plot}
\usage{
plotCorr(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Correlation plot
}
\details{
Generic function for correlation plot
}
