% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{plotCumVariance.mcr}
\alias{plotCumVariance.mcr}
\title{Show plot with cumulative explained variance}
\usage{
\method{plotCumVariance}{mcr}(
  obj,
  type = "b",
  labels = "values",
  main = "Cumulative variance",
  xticks = seq_len(obj$ncomp),
  ...
)
}
\arguments{
\item{obj}{object of clacc \code{mcr}}

\item{type}{type of the plot}

\item{labels}{what to use as data labels}

\item{main}{title of the plot}

\item{xticks}{vector with ticks for x-axis}

\item{...}{other parameters suitable for \code{mdaplot}}
}
\description{
Show plot with cumulative explained variance
}
