% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotCumVariance.pca}
\alias{plotCumVariance.pca}
\title{Cumulative explained variance plot for PCA model}
\usage{
\method{plotCumVariance}{pca}(
  obj,
  type = "b",
  main = "Cumulative variance",
  xlab = "Components",
  ylab = "Explained variance, \%",
  show.legend = TRUE,
  labels = "values",
  res = obj$res,
  ...
)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{type}{type of the plot ("b", "l", "h")}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{show.legend}{logical, show or not a legend on the plot}

\item{labels}{what to use as labels (if \code{show.labels = TRUE})}

\item{res}{list with result objects to show the variance for}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows a plot with cumulative explained variance for components.
}
\details{
See examples in help for \code{\link{pca}} function.
}
