\name{plotCumVariance}
\alias{plotCumVariance}
\title{Variance plot}
\usage{
plotCumVariance(obj, ...)
}
\arguments{
  \item{obj}{a model or result object}

  \item{...}{other arguments}
}
\description{
Variance plot
}
\details{
Generic function for plotting explained variance for data
decomposition
}

