% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{plotDensity}
\alias{plotDensity}
\title{Show plot series as density plot (using hex binning)}
\usage{
plotDensity(ps, nbins = 60, colmap = ps$colmap)
}
\arguments{
\item{ps}{`plotseries` object}

\item{nbins}{number of bins in one dimension}

\item{colmap}{colormap name or values used to create color gradient}
}
\description{
Show plot series as density plot (using hex binning)
}
