% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{plotDistances.ddsimca}
\alias{plotDistances.ddsimca}
\title{Show with distance values (score, orthogonal or full) vs object indices for calibration
and PV-set results.}
\usage{
\method{plotDistances}{ddsimca}(
  obj,
  res = "both",
  ncomp = obj$ncomp.selected,
  limType = "classic",
  distance = "h",
  log = FALSE,
  lim.lty = c(2, 3),
  lim.col = c("darkgray", "darkgray"),
  lim.lwd = c(1, 1),
  ylim = NULL,
  xlim = NULL,
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA model (object of class \code{ddsimca})}

\item{res}{name of the results (can be 'cal', 'pv', 'both', or 'diff', in the latter case shows difference
of the distance values between the training and the PV-set).}

\item{ncomp}{number of components to show the plot for.}

\item{limType}{limit type to show the plot for ('classic' or 'robust')}

\item{distance}{which distance to show (\code{"h"} for score, \code{"q"} for orthogonal or \code{"f"} for full distance).}

\item{log}{logical, apply log transformation or not.}

\item{lim.lty}{vector with two values - types of the lines showing critical limits (extreme, outlier).}

\item{lim.col}{vector with two values - colors of the lines showing critical limits (extreme, outlier).}

\item{lim.lwd}{vector with two values - thickness of the lines showing critical limits (extreme, outlier).}

\item{ylim}{limits for y-axis, if NULL will be estimated automatically.}

\item{xlim}{limits for x-axis, if NULL will be estimated automatically.}

\item{...}{other parameters compatible with \code{\link{plotDistances.ddsimcares}} method.}
}
\description{
Show with distance values (score, orthogonal or full) vs object indices for calibration
and PV-set results.
}
