% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotDistances.ddsimcares}
\alias{plotDistances.ddsimcares}
\title{Show with distance values (score, orthogonal or full) vs object indices for DD-SIMCA results.}
\usage{
\method{plotDistances}{ddsimcares}(
  obj,
  ncomp = obj$ncomp.selected,
  limType = "classic",
  distance = "h",
  log = FALSE,
  show.legend = TRUE,
  pch = 1,
  lim.lty = c(2, 3),
  lim.col = c("darkgray", "darkgray"),
  lim.lwd = c(1, 1),
  ylim = NULL,
  xlim = NULL,
  show.plot = TRUE,
  show.excluded = FALSE,
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{ncomp}{number of components to show the plot for.}

\item{limType}{limit type to show the plot for ('classic' or 'robust')}

\item{distance}{which distance to show (\code{"h"} for score, \code{"q"} for orthogonal or \code{"f"} for full distance).}

\item{log}{logical, apply log transformation or not.}

\item{show.legend}{logica, show or not colorbar legend on the plot.}

\item{pch}{marker type}

\item{lim.lty}{vector with two values - types of the lines showing critical limits (extreme, outlier).}

\item{lim.col}{vector with two values - colors of the lines showing critical limits (extreme, outlier).}

\item{lim.lwd}{vector with two values - thickness of the lines showing critical limits (extreme, outlier).}

\item{ylim}{limits for y-axis, if NULL will be estimated automatically.}

\item{xlim}{limits for x-axis, if NULL will be estimated automatically.}

\item{show.plot}{logical, if FALSE returns only values and does not show the plot.}

\item{show.excluded}{logical, show or not the excluded rows (objects).}

\item{...}{other parameters compatible with \code{\link{mdaplot}} method.}
}
\description{
Show with distance values (score, orthogonal or full) vs object indices for DD-SIMCA results.
}
