% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotDistDoF}
\alias{plotDistDoF}
\alias{plotDistDoF.pca}
\title{Degrees of freedom plot for both distances}
\usage{
plotDistDoF(obj, ...)

\method{plotDistDoF}{pca}(
  obj,
  type = "b",
  labels = "values",
  xticks = seq_len(obj$ncomp),
  ...
)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}

\item{type}{type of the plot ("b", "l", "h")}

\item{labels}{what to show as data points labels}

\item{xticks}{vector with tick values for x-axis}
}
\description{
Shows a plot with degrees of freedom computed for score and orthogonal distances at given number
of components using data driven approach ("ddmoments" or "ddrobust").
}
\details{
Works only if parameter \code{lim.type} equal to "ddmoments" or "ddrobust".
}
