% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotEigenvalues}
\alias{plotEigenvalues}
\title{Eigenvalues plot}
\usage{
plotEigenvalues(obj, ...)
}
\arguments{
\item{obj}{a model object}

\item{...}{other arguments}
}
\description{
Eigenvalues plot
}
\details{
Generic function for plotting eigenvalues vs. number of components
}
