% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{plotErrorbars}
\alias{plotErrorbars}
\title{Show plot series as error bars}
\usage{
plotErrorbars(ps, col = ps$col, pch = 16, lwd = 1, cex = 1, ...)
}
\arguments{
\item{ps}{`plotseries` object}

\item{col}{color for the error bars}

\item{pch}{marker symbol for the plot}

\item{lwd}{line width for the error bars}

\item{cex}{scale factor for the marker}

\item{...}{other arguments for function `points()`.}
}
\description{
It is assumed that first row of dataset contains the y-coordinates of points,
second rows contains size of lower error bar and third - size for upper error bar. If
only two rows are provided it is assumed that error bars are symmetric.
}
