% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{plotExtreme.ddsimca}
\alias{plotExtreme.ddsimca}
\title{A shortcut to \code{\link{plotExtremes.ddsimca}}.}
\usage{
\method{plotExtreme}{ddsimca}(obj, ...)
}
\arguments{
\item{obj}{DD-SIMCA model (object of class \code{ddsimca}).}

\item{...}{any parameters suitable for \code{\link{plotExtremes.ddsimca}}.}
}
\description{
A shortcut to \code{\link{plotExtremes.ddsimca}}.
}
