% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotExtreme.ddsimcares}
\alias{plotExtreme.ddsimcares}
\title{Extremes plot (shortcut to \code{\link{plotExtremes.ddsimcares}}).}
\usage{
\method{plotExtreme}{ddsimcares}(obj, ...)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{...}{any parameter acceptable by \code{\link{plotExtremes.ddsimcares}}.}
}
\description{
Extremes plot (shortcut to \code{\link{plotExtremes.ddsimcares}}).
}
