% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotExtreme.pca}
\alias{plotExtreme.pca}
\title{A shortcut to `\code{\link{plotExtremes.pca}}.}
\usage{
\method{plotExtreme}{pca}(obj, ...)
}
\arguments{
\item{obj}{PCA model (object of class \code{pca}).}

\item{...}{any parameters suitable for `\code{\link{plotExtremes.pca}}.}
}
\description{
A shortcut to `\code{\link{plotExtremes.pca}}.
}
