% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simca.R
\name{plotExtreme.simca}
\alias{plotExtreme.simca}
\title{Shows extreme plot for SIMCA model}
\usage{
\method{plotExtreme}{simca}(obj, ncomp = NULL, main = NULL,
  xlab = "Expected", ylab = "Observed", ...)
}
\arguments{
\item{obj}{SIMCA model}

\item{ncomp}{Number of components to show the plot for}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{...}{other arguments}
}
\description{
The plot shows the number of extreme objects rejected by the model vs. the expected number, 
which depends on significance level and total number of objects. The light blue area shows 95% 
tolerance limits. See more details in [1].
}
\references{
1. A.L. Pomerantsev, O.Ye. Rodionova, Concept and role of extreme objects in PCA/SIMCA,
Journal of Chemometrics, 28 (2014) pp. 429-438.
}
