% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{plotExtremes.ddsimca}
\alias{plotExtremes.ddsimca}
\title{Extreme plot}
\usage{
\method{plotExtremes}{ddsimca}(
  obj,
  ncomp = obj$ncomp.selected,
  limType = "classic",
  col = mdaplot.getColors(2),
  legend.position = "topleft",
  ...
)
}
\arguments{
\item{obj}{a DD-SIMCA model (object of class \code{ddsimca})}

\item{ncomp}{number of components to show the plot for}

\item{limType}{limit type to show the plot for ('classic' or 'robust').}

\item{col}{vector with two colors (for calibration and PV-set sensitivity).}

\item{legend.position}{position of the legend}

\item{...}{other arguments, suitable for \code{\link{plotExtremes.ddsimcares}}}
}
\description{
Shows a plot with number of expected vs. number of observed extreme objects for different
significance levels (alpha values) for calibration and PV-set results.
}
