% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotExtremes.ddsimcares}
\alias{plotExtremes.ddsimcares}
\title{Extremes plot.}
\usage{
\method{plotExtremes}{ddsimcares}(
  obj,
  ncomp = obj$ncomp.selected,
  limType = "classic",
  main = sprintf("Extremes (A = \%d)", ncomp),
  xlab = "Number of extremes (expected)",
  ylab = "Number of extremes (observed)",
  pch = 1,
  col = "#2679b2",
  ellipse.col = "#eeeeee",
  show.plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{ncomp}{number of components in the model to show the plot for.}

\item{limType}{limit type to show the plot for ('classic' or 'robust').}

\item{main}{main title.}

\item{xlab}{label for x-axis.}

\item{ylab}{label for y-axis.}

\item{pch}{marker number or symbol for the plot points.}

\item{col}{color of the plot points.}

\item{ellipse.col}{color of the confidence ellipse elements.}

\item{show.plot}{logical, if FALSE returns only values and does not show the plot.}

\item{...}{other parameters compatible with \code{\link{mdaplot}} method.}
}
\description{
The method generates a sequence of values for significance level (alpha) and then
computes how many extremes the model produces by applying it to the target class members. After
that the computed number is plotted against the expected number.

The plot also shows a confidence ellipse for the expected values.
}
