% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotExtremes}
\alias{plotExtremes}
\title{Shows extreme plot for PCA and DD-SIMCA models}
\usage{
plotExtremes(obj, ...)
}
\arguments{
\item{obj}{a PCA or DD-SIMCA model}

\item{...}{other parameters}
}
\description{
Generic function for creating extreme plot for PCA and DD-SIMCA models
}
