% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotFactors}
\alias{plotFactors}
\title{Plot factors for a 3-way decomposition model}
\usage{
plotFactors(obj, ...)
}
\arguments{
\item{obj}{a model object.}

\item{...}{arguments passed to methods.}
}
\description{
Plot factors for a 3-way decomposition model
}
