% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotFoM.ddsimcares}
\alias{plotFoM.ddsimcares}
\title{Figure of merit plot.}
\usage{
\method{plotFoM}{ddsimcares}(
  obj,
  fom = "sens",
  limType = "classic",
  type = "b",
  ylim = c(0, 1.1),
  pch = 1,
  show.plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{fom}{name of the figure of merit to show the plot for ('sens', 'spec', 'sel', 'eff', 'acc').}

\item{limType}{limit type to show the plot for ('classic' or 'robust').}

\item{type}{type of plot (can be 'b', 'l', or 'h').}

\item{ylim}{limits for y-axis.}

\item{pch}{marker number or symbol for the plot points (if 'b' type is selected).}

\item{show.plot}{logical, if FALSE returns only values and does not show the plot.}

\item{...}{any parameters for \code{\link{mdaplot}} method can be provided.

Shows the selected figure of merit values vs. number of components. Pay attention, that
some of the FoMs may not be available, for example if result object is made only for
members of the target class, only sensitivity ('sens') is available.}
}
\description{
Figure of merit plot.
}
