% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotFoM}
\alias{plotFoM}
\title{Show plot with figure of merit vs. number of components (generic function).}
\usage{
plotFoM(obj, ...)
}
\arguments{
\item{obj}{result object, e.g. \code{ddsimcares}}

\item{...}{other parameters relevant for the method}
}
\description{
Show plot with figure of merit vs. number of components (generic function).
}
