% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotFoMs.ddsimcares}
\alias{plotFoMs.ddsimcares}
\title{Figures of merit plot (multiple FoMs).}
\usage{
\method{plotFoMs}{ddsimcares}(
  obj,
  foms = NULL,
  limType = "classic",
  type = "b",
  ylim = c(0, 1.1),
  col = NULL,
  legend.position = "bottom",
  main = "Figures of merit",
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{foms}{vector with FoM names to show on the plot (any combination of 'sens', 'spec', 'sel', 'eff', 'acc').}

\item{limType}{limit type to show the plot for ('classic' or 'robust').}

\item{type}{type of plot (can be 'b', 'l', or 'h').}

\item{ylim}{limits for y-axis.}

\item{col}{vector with color values (should contain one value for each fom), if not provided will use default colors.}

\item{legend.position}{position of legend on the plot.}

\item{main}{main title.}

\item{...}{any parameters for \code{\link{mdaplotg}} method can be provided.

Shows several figures of merit values vs. number of components as a group plot (see
details in \code{\link{mdaplotg}} method desciption.}
}
\description{
Figures of merit plot (multiple FoMs).
}
