% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotFoMs}
\alias{plotFoMs}
\title{Show plot with several figures of merit vs. number of components (generic function).}
\usage{
plotFoMs(obj, ...)
}
\arguments{
\item{obj}{result object, e.g. \code{ddsimcares}}

\item{...}{other parameters relevant for the method}
}
\description{
Show plot with several figures of merit vs. number of components (generic function).
}
