% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtest.R
\name{plotHist.randtest}
\alias{plotHist.randtest}
\title{Histogram plot for randomization test results}
\usage{
\method{plotHist}{randtest}(obj, ncomp = obj$ncomp.selected, bwd = 0.9, ...)
}
\arguments{
\item{obj}{results of randomization test (object of class `randtest`)}

\item{ncomp}{number of component to make the plot for}

\item{bwd}{width of bars (between 0 and 1)}

\item{...}{other optional arguments}
}
\description{
Makes a histogram for statistic values distribution for particular component, also
show critical value as a vertical line.
}
\details{
See examples in help for \code{\link{randtest}} function.
}
