% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotHist}
\alias{plotHist}
\title{Statistic histogram}
\usage{
plotHist(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Statistic histogram
}
\details{
Generic function for plotting statistic histogram plot
}

