\name{plotModelDistance.simcam}
\alias{plotModelDistance.simcam}

\title{
Model distance plot for SIMCA multiclass model
}

\description{
\code{plotModelDistance.simcam} is used to show plot for distance between individual SIMCA models. 
}
\usage{
plotModelDistance(model, nc = 1, ...)
}

\arguments{
  \item{model}{an object of class \code{simcam} (a SIMCA multiclass model).}
  \item{nc}{number of class (individual model) to plot the distance values for.}
  \item{\dots}{most of \code{\link{mdaplot}} parameters can be used.}
}

\details{
See examples for \code{\link{simcam}} class.
}

\seealso{
\code{\link{simcam}} to make SIMCAM model, \code{\link{simcamres}} to store and explore SIMCA multiclass classification results.
}

\examples{
}
