\name{plotModellingPower.simca}
\alias{plotModellingPower.simca}

\title{
Modelling power of variables
}

\description{
\code{plotModellingPower.simca} is used to show plot with modelling power of variables for a SIMCA model.
}
\usage{
plotModellingPower(model, ncomp = NULL)
}

\arguments{
  \item{model}{an object of class \code{simca} (a SIMCA model).}
  \item{ncomp}{number of components to show the plot for.}
  \item{\dots}{most of \code{\link{mdaplot}} parameters can be used.}
}

\details{
See examples for \code{\link{simca}} class.
}

\seealso{
\code{\link{simca}} to make SIMCA model, \code{\link{simcares}} to store and explore SIMCA one-class classification results.
}

\examples{
}
