% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcamres.R
\name{plotPredictions.simcamres}
\alias{plotPredictions.simcamres}
\title{Prediction plot for SIMCAM results}
\usage{
\method{plotPredictions}{simcamres}(obj, nc = seq_len(obj$nclasses), main = "Predictions", ...)
}
\arguments{
\item{obj}{classification results (object of class \code{plsdares}, \code{simcamres}, etc.).}

\item{nc}{vector with classes to show predictions for.}

\item{main}{title of the plot}

\item{...}{most of the graphical parameters from \code{\link{mdaplotg}} or \code{\link{mdaplot}} function
can be used.}
}
\description{
Makes a plot with predicted class values for classification results.
}
\details{
See examples in description of \code{\link{plsdares}}, \code{\link{simcamres}}, etc.
}
