% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classres.R
\name{plotProbabilities.classres}
\alias{plotProbabilities.classres}
\title{Plot for class belonging probability}
\usage{
\method{plotProbabilities}{classres}(obj, ncomp = obj$ncomp.selected,
  nc = 1, type = "h", show.lines = c(0.5, NA), ...)
}
\arguments{
\item{obj}{classification results (e.g. object of class \code{simcamres}).}

\item{ncomp}{number of components to use the probabilities for.}

\item{nc}{if there are several classes, which class to make the plot for.}

\item{type}{type of the plot}

\item{show.lines}{shows a horizontal line at p = 0.5}

\item{...}{most of the graphical parameters from \code{\link{mdaplot}} function can be used.}
}
\description{
Makes a plot with class belonging probabilities for each object of the classification results. 
Works only with classification methods, which compute this probability (e.g. SIMCA).
}
