% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrpure.R
\name{plotPurity.mcrpure}
\alias{plotPurity.mcrpure}
\title{Purity values plot}
\usage{
\method{plotPurity}{mcrpure}(
  obj,
  xticks = seq_len(obj$ncomp),
  type = "h",
  labels = "values",
  ...
)
}
\arguments{
\item{obj}{\code{mcrpure} object}

\item{xticks}{ticks for x axis}

\item{type}{type of the plot}

\item{labels}{what to use as data labels}

\item{...}{other parameters suitable for \code{mdaplot}

The plot shows largest weighted purity value for each component graphically.}
}
\description{
Purity values plot
}
