% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotPurity}
\alias{plotPurity}
\title{Plot purity values}
\usage{
plotPurity(obj, ...)
}
\arguments{
\item{obj}{object with mcr pure case}

\item{...}{other parameters}
}
\description{
Plot purity values
}
