% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotPuritySpectra}
\alias{plotPuritySpectra}
\title{Plot purity spectra}
\usage{
plotPuritySpectra(obj, ...)
}
\arguments{
\item{obj}{object with mcr pure case}

\item{...}{other parameters}
}
\description{
Plot purity spectra
}
