% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotQDoF}
\alias{plotQDoF}
\alias{plotQDoF.pca}
\title{Degrees of freedom plot for orthogonal distance (Nq)}
\usage{
plotQDoF(obj, ...)

\method{plotQDoF}{pca}(
  obj,
  type = "b",
  labels = "values",
  xticks = seq_len(obj$ncomp),
  ylab = "Nq",
  ...
)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}

\item{type}{type of the plot ("b", "l", "h")}

\item{labels}{what to show as data points labels}

\item{xticks}{vector with tick values for x-axis}

\item{ylab}{label for y-axis}
}
\description{
Shows a plot with degrees of freedom computed for orthogonal distances at given number
of components using data driven approach ("ddmoments" or "ddrobust").
}
\details{
Works only if parameter \code{lim.type} equal to "ddmoments" or "ddrobust".
}
