% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{plotRegressionLine}
\alias{plotRegressionLine}
\title{Add regression line for data points}
\usage{
plotRegressionLine(p, col = p$col, ...)
}
\arguments{
\item{p}{plot data returned by function `mdaplot()`}

\item{col}{color of line}

\item{...}{other parameters available for `abline()` function}
}
\description{
Shows linear fit line for data points.
}
