% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{plotResiduals.ldecomp}
\alias{plotResiduals.ldecomp}
\title{Residuals distance plot for a set of ldecomp objects (legacy, use \code{\link{plotDistances.ldecomp}} instead).}
\usage{
\method{plotResiduals}{ldecomp}(obj, ...)
}
\arguments{
\item{obj}{object of \code{ldecomp} class.}

\item{...}{other parameters.}
}
\description{
Residuals distance plot for a set of ldecomp objects (legacy, use \code{\link{plotDistances.ldecomp}} instead).
}
