% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotResiduals.pca}
\alias{plotResiduals.pca}
\title{Shortcut to plotDistances()}
\usage{
\method{plotResiduals}{pca}(obj, ...)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{...}{any parameter from \code{\link{plotDistances.pca}}}
}
\description{
Shortcut to plotDistances()
}
