% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcares.R
\name{plotResiduals.pcares}
\alias{plotResiduals.pcares}
\title{Residual distance plot}
\usage{
\method{plotResiduals}{pcares}(
  obj,
  ncomp = obj$ncomp.selected,
  cgroup = obj$categories,
  norm = TRUE,
  log = FALSE,
  show.labels = FALSE,
  labels = "names",
  main = NULL,
  show.plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{object of \code{ldecomp} class.}

\item{ncomp}{number of components to show the plot for (if NULL, selected by model value will be used).}

\item{cgroup}{color grouping of distance points, if "type" then they will be grouped by type ("normal", "extreme", "outliers")}

\item{norm}{logical, shall values be normalized or not}

\item{log}{logical, if TRUE, then log(1 + u) transformation is applied}

\item{show.labels}{logical, show or not labels for the plot objects}

\item{labels}{what to show as labels if necessary}

\item{main}{main title for the plot}

\item{...}{most of graphical parameters from \code{\link{mdaplot}} function can be used.}

\item{show.limits}{logical, shall extreme and outlier limits be shown or not}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{plot}{logical, shall plot be created or just plot series object is needed}
}
\description{
Shows a plot with orthogonal (Q, q) vs score (T2, h) distances for data objects. By default the
distance values are normalize using corresponding means (q/q0 and h/h0).
}
