% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Residuals plot}
\usage{
plotResiduals(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Residuals plot
}
\details{
Generic function for plotting residual values for data decomposition
}

