% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipls.R
\name{plotRMSE.ipls}
\alias{plotRMSE.ipls}
\title{RMSE development plot}
\usage{
\method{plotRMSE}{ipls}(
  obj,
  glob.ncomp = obj$gm$ncomp.selected,
  main = "RMSE development",
  xlab = "Iterations",
  ylab = "RMSECV",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{obj}{iPLS results (object of class ipls)}

\item{glob.ncomp}{number of components for global PLS model with all intervals}

\item{main}{main title for the plot}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{xlim}{limits for x-axis}

\item{ylim}{limits for y-axis}

\item{...}{other arguments}
}
\description{
Shows how RMSE develops for each iteration of iPLS selection
algorithm
}
\details{
The plot shows RMSE values obtained at each iteration of the iPLS algorithm as bars. The first
bar correspond to the global model with all variables included, second - to the model obtained
at the first iteration and so on. Number at the bottom of each bar corresponds to the interval
included or excluded at the particular iteration.
}
\seealso{
\code{\link{summary.ipls}}, \code{\link{plotSelection.ipls}}
}
