% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotRMSERatio}
\alias{plotRMSERatio}
\title{Plot for ratio RMSEC/RMSECV vs RMSECV}
\usage{
plotRMSERatio(obj, ...)
}
\arguments{
\item{obj}{object with any regression model}

\item{...}{other parameters}
}
\description{
Plot for ratio RMSEC/RMSECV vs RMSECV
}
