% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodel.R
\name{plotRMSERatio.regmodel}
\alias{plotRMSERatio.regmodel}
\title{RMSE plot for regression model}
\usage{
\method{plotRMSERatio}{regmodel}(
  obj,
  ny = 1,
  type = "b",
  xlab = "RMSECV/RMSEC ratio",
  ylab = "RMSECV",
  ...
)
}
\arguments{
\item{obj}{a regression model (object of class \code{regmodel})}
}
\description{
Shows plot with root mean squared error values vs. number of components for PLS model.
}
