% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{plotScatter}
\alias{plotScatter}
\title{Show plot series as set of points}
\usage{
plotScatter(
  ps,
  pch = 16,
  col = ps$col,
  bg = "white",
  lwd = 1,
  cex = 1,
  col.excluded = "lightgray",
  pch.colinv = FALSE,
  show.excluded = FALSE,
  ...
)
}
\arguments{
\item{ps}{`plotseries` object}

\item{pch}{size of point markers}

\item{col}{color of the points}

\item{bg}{background color of the points if `pch=21:25`}

\item{lwd}{line width for the error bars}

\item{cex}{scale factor for the marker}

\item{col.excluded}{color for excluded values (if must be shown)}

\item{pch.colinv}{logical, should `col` and `bg` be switched if `pch=21:25` and `cgroup` is used to create colors.}

\item{show.excluded}{logical, show or not the excluded data points}

\item{...}{other arguments for function `points()`.}
}
\description{
Show plot series as set of points
}
