% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotScores}
\alias{plotScores}
\title{Scores plot}
\usage{
plotScores(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Scores plot
}
\details{
Generic function for scores values for data decomposition
}

