% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotSelection}
\alias{plotSelection}
\title{Selected intervals plot}
\usage{
plotSelection(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Selected intervals plot
}
\details{
Generic function for plotting selected intervals or variables
}
