% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotSelectivityArea.ddsimcares}
\alias{plotSelectivityArea.ddsimcares}
\title{Selectivity area plot (similar to ROC curve).}
\usage{
\method{plotSelectivityArea}{ddsimcares}(
  obj,
  ncomp = obj$ncomp.selected,
  limType = "classic",
  col = "#2679b2",
  xlab = bquote("1 - selectivity, " ~ beta),
  ylab = bquote("Sensitivity, 1 - " ~ alpha),
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares}).}

\item{ncomp}{number of components in the model to show the plot for.}

\item{limType}{limit type to show the plot for ('classic' or 'robust').}

\item{col}{color of the plot elements.}

\item{xlab}{label for x-axis.}

\item{ylab}{label for y-axis.}

\item{...}{other parameters compatible with \code{\link{mdaplot}} method.}
}
\description{
The method generates a sequence of values for significance level (alpha) and then
computes corresponding selectivity (1 - beta) for each value from the sequence using model parameters
for given number of components and limit type. Then the values are shown as line plot using
coordinates (1 - alpha) vs beta and area under the curve is shown in the plot title.

Current values of alpha and beta are shown on the plot as a marker.
}
