% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotSelectivityArea}
\alias{plotSelectivityArea}
\title{Selectivity vs sensitivity plot for DD-SIMCA results (generic function)}
\usage{
plotSelectivityArea(obj, ...)
}
\arguments{
\item{obj}{result object, e.g. \code{ddsimcares}}

\item{...}{other parameters relevant for the plot}
}
\description{
Selectivity vs sensitivity plot for DD-SIMCA results (generic function)
}
