% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{plotSelectivityRatio.pls}
\alias{plotSelectivityRatio.pls}
\title{Selectivity ratio plot for PLS model}
\usage{
\method{plotSelectivityRatio}{pls}(obj, ny = 1, ncomp = obj$ncomp.selected, type = "l", ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{ny}{which response to plot the values for (if y is multivariate), can be a vector.}

\item{ncomp}{number of components to count}

\item{type}{type of the plot}

\item{...}{other plot parameters (see \code{mdaplot} for details)}
}
\description{
Computes and shows a plot for Selectivity ratio values for given number of components
and response variable
}
\details{
See \code{\link{vipscores}} for more details.
}
