% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{plotSensitivity.ddsimca}
\alias{plotSensitivity.ddsimca}
\title{Sensitivity plot.}
\usage{
\method{plotSensitivity}{ddsimca}(
  obj,
  limType = "classic",
  col = mdaplot.getColors(2),
  type = "b",
  pch = c(16, 16),
  legend.position = "bottomright",
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA model (object of class \code{ddsimca}).}

\item{limType}{limit type to show the plot for ('classic' or 'robust').}

\item{col}{vector with two colors (for calibration and PV-set sensitivity).}

\item{type}{type of the plot ("b", "l", or "h").}

\item{pch}{vector with two markers (for calibration and PV-set sensitivity).}

\item{legend.position}{position of legend on the plot.}

\item{...}{any parameters suitable for the \code{\link{plotSensitivity.ddsimcares}} method.}
}
\description{
Sensitivity plot.
}
\details{
The method shows sensitivity vs. number of components for calibration and PV-set results.
}
