% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotSensitivity.ddsimcares}
\alias{plotSensitivity.ddsimcares}
\title{Sensitivity plot.}
\usage{
\method{plotSensitivity}{ddsimcares}(
  obj,
  show.ci = FALSE,
  ci.col = "#a0a0a0",
  ci.lty = c(3, 2, 3),
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares}).}

\item{show.ci}{logical, show or not the expected sensitivity and 95\% confidence interval for it.}

\item{ci.col}{color for lines showing expected sensitivity and confidence interval.}

\item{ci.lty}{type of lines (lower, expected, upper) for the expected sensitivity and confidence interval.}

\item{...}{any parameters suitable for the \code{\link{plotFoM.ddsimcares}} method.}
}
\description{
Sensitivity plot.
}
\details{
The method is a wrapper for more general \code{\link{plotFoM.ddsimcares}} method with a
possibility to show confidence interval for the expected sensitivity.
}
