% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotSensitivity}
\alias{plotSensitivity}
\title{Sensitivity plot}
\usage{
plotSensitivity(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Sensitivity plot
}
\details{
Generic function for plotting sensitivity values for classification model or results
}
