\name{plotSensitivity.simca}
\alias{plotSensitivity.simca}

\title{
Sensitivity plot for SIMCA model
}

\description{
\code{plotSensitivity.simca} is used to show plot with sensitivity values for calibration and cross-validation predictions.
}

\usage{
plotSensitivity(model, ...)
}

\arguments{
  \item{model}{an object of class \code{simca} (a SIMCA model).}
  \item{\dots}{most of \code{\link{mdaplotg}} parameters can be used.}
}

\details{
See examples for \code{\link{simca}} class.
}

\seealso{
\code{\link{simca}} to make SIMCA model, \code{\link{simcares}} to store and explore SIMCA one-class classification results.
}
