\name{plotSpecificity}
\alias{plotSpecificity}
\title{Specificity plot}
\usage{
plotSpecificity(obj, ...)
}
\arguments{
  \item{obj}{a model or result object}

  \item{...}{other arguments}
}
\description{
Specificity plot
}
\details{
Generic function for plotting specificity values for
classification model or results
}

