% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Plot resolved spectra}
\usage{
plotSpectra(obj, ...)
}
\arguments{
\item{obj}{object with mcr case}

\item{...}{other parameters}
}
\description{
Plot resolved spectra
}
