% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotT2DoF}
\alias{plotT2DoF}
\alias{plotT2DoF.pca}
\title{Degrees of freedom plot for score distance (Nh)}
\usage{
plotT2DoF(obj, ...)

\method{plotT2DoF}{pca}(
  obj,
  type = "b",
  labels = "values",
  xticks = seq_len(obj$ncomp),
  ylab = "Nh",
  ...
)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}

\item{type}{type of the plot ("b", "l", "h")}

\item{labels}{what to show as data points labels}

\item{xticks}{vector with tick values for x-axis}

\item{ylab}{label for y-axis}
}
\description{
Shows a plot with degrees of freedom computed for score distances at given number
of components using data driven approach ("ddmoments" or "ddrobust").
}
\details{
Works only if parameter \code{lim.type} equal to "ddmoments" or "ddrobust".
}
