% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{plotVariance.mcr}
\alias{plotVariance.mcr}
\title{Show plot with explained variance}
\usage{
\method{plotVariance}{mcr}(
  obj,
  type = "h",
  labels = "values",
  main = "Variance",
  xticks = seq_len(obj$ncomp),
  ...
)
}
\arguments{
\item{obj}{object of clacc \code{mcr}}

\item{type}{type of the plot}

\item{labels}{what to use as data labels}

\item{main}{title of the plot}

\item{xticks}{vector with ticks for x-axis}

\item{...}{other parameters suitable for \code{mdaplot}}
}
\description{
Show plot with explained variance
}
