% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotVariance}
\alias{plotVariance}
\title{Variance plot}
\usage{
plotVariance(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Variance plot
}
\details{
Generic function for plotting explained variance for data decomposition
}

