% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotVIPScores}
\alias{plotVIPScores}
\title{VIP scores plot}
\usage{
plotVIPScores(obj, ...)
}
\arguments{
\item{obj}{a regression model}

\item{...}{other parameters}
}
\description{
Generic function for plotting VIP scores values for regression model (PCR, PLS, etc)
}

