% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{plotWeights.pls}
\alias{plotWeights.pls}
\title{X loadings plot for PLS}
\usage{
\method{plotWeights}{pls}(
  obj,
  comp = 1,
  type = (if (nrow(obj$weights) < 20) "h" else "l"),
  show.axes = TRUE,
  show.legend = TRUE,
  ...
)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{comp}{which components to show the plot for (one or vector with several values)}

\item{type}{type of the plot}

\item{show.axes}{logical, show or not a axes lines crossing origin (0,0)}

\item{show.legend}{logical, show or not a legend}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows plot with X loading values for selected components.
}
\details{
See examples in help for \code{\link{pls}} function.
}
