% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotWeights}
\alias{plotWeights}
\title{Plot for PLS weights}
\usage{
plotWeights(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Plot for PLS weights
}
\details{
Generic function for weight plot
}
