\name{plotXLoadings}
\alias{plotXLoadings}
\title{X loadings plot}
\usage{
plotXLoadings(obj, ...)
}
\arguments{
  \item{obj}{a model or result object}

  \item{...}{other arguments}
}
\description{
X loadings plot
}
\details{
Generic function for plotting loadings values for
decomposition of x data
}

