% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotXResiduals}
\alias{plotXResiduals}
\title{X residuals plot}
\usage{
plotXResiduals(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
X residuals plot
}
\details{
Generic function for plotting x residuals for classification or regression model or results
}

