% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotXScores}
\alias{plotXScores}
\title{X scores plot}
\usage{
plotXScores(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
X scores plot
}
\details{
Generic function for plotting scores values for decomposition of x data
}

