% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotXVariance}
\alias{plotXVariance}
\title{X variance plot}
\usage{
plotXVariance(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
X variance plot
}
\details{
Generic function for plotting explained variance for decomposition of x data
}

