% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotXYLoadings}
\alias{plotXYLoadings}
\title{XY loadings plot}
\usage{
plotXYLoadings(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
XY loadings plot
}
\details{
Generic function for plotting loadings values for decomposition of x and y data
}
