% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotXYResiduals}
\alias{plotXYResiduals}
\title{Plot for XY-residuals}
\usage{
plotXYResiduals(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Plot for XY-residuals
}
\details{
Generic function for XY-residuals plot
}
