\name{plotXYScores}
\alias{plotXYScores}
\title{XY scores plot}
\usage{
plotXYScores(obj, ...)
}
\arguments{
  \item{obj}{a model or result object}

  \item{...}{other arguments}
}
\description{
XY scores plot
}
\details{
Generic function for plotting scores values for
decomposition of x and y data
}

