% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotYCumVariance}
\alias{plotYCumVariance}
\title{Y cumulative variance plot}
\usage{
plotYCumVariance(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Y cumulative variance plot
}
\details{
Generic function for plotting cumulative explained variance for decomposition of y data
}

