% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodel.R
\name{plotYResiduals.regmodel}
\alias{plotYResiduals.regmodel}
\title{Y residuals plot for regression model}
\usage{
\method{plotYResiduals}{regmodel}(
  obj,
  ncomp = obj$ncomp.selected,
  ny = 1,
  show.lines = c(NA, 0),
  res = obj$res,
  ...
)
}
\arguments{
\item{obj}{a regression model (object of class \code{regmodel})}

\item{ncomp}{how many components to use (if NULL - user selected optimal value will be used)}

\item{ny}{number of response variable to make the plot for (if y is multivariate)}

\item{show.lines}{allows to show the horizonta line at 0 level}

\item{res}{list with result objects}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows plot with y residuals (predicted vs. reference values) for selected components.
}
