% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotYVariance}
\alias{plotYVariance}
\title{Y variance plot}
\usage{
plotYVariance(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Y variance plot
}
\details{
Generic function for plotting explained variance for decomposition of y data
}

