% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.fromjson}
\alias{pls.fromjson}
\title{Converts JSON string created in mda.tools/pls app to \code{pls} object}
\usage{
pls.fromjson(str)
}
\arguments{
\item{str}{stringified JSON (from model file)}
}
\value{
object of \code{\link{pls}} class
}
\description{
Converts JSON string created in mda.tools/pls app to \code{pls} object
}
