% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.getLimitsCoordinates}
\alias{pls.getLimitsCoordinates}
\title{Compute coordinates of lines or curves with critical limits}
\usage{
pls.getLimitsCoordinates(Qlim, T2lim, Zlim, nobj, ncomp, norm, log)
}
\arguments{
\item{Qlim}{matrix with critical limits for orthogonal distances (X)}

\item{T2lim}{matrix with critical limits for score distances (X)}

\item{Zlim}{matrix with critical limits for orthogonal distances (Y)}

\item{nobj}{number of objects to compute the limits for}

\item{ncomp}{number of components for computing the coordinates}

\item{norm}{logical, shall distance values be normalized or not}

\item{log}{logical, shall log transformation be applied or not}
}
\value{
list with two matrices (x and y coordinates of corresponding limits)
}
\description{
Compute coordinates of lines or curves with critical limits
}
