% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.getpredictions}
\alias{pls.getpredictions}
\title{Compute predictions for response values}
\usage{
pls.getpredictions(
  x,
  coeffs,
  ycenter,
  yscale,
  ynames = NULL,
  y.attrs = NULL,
  objnames = NULL,
  compnames = NULL
)
}
\arguments{
\item{x}{matrix with predictors, already preprocessed (e.g. mean centered) and cleaned}

\item{coeffs}{array with regression coefficients}

\item{ycenter}{`ycenter` property of PLS model}

\item{yscale}{`yscale` property of PLS model}

\item{ynames}{vector with names of the responses}

\item{y.attrs}{list with response attributes (e.g. from reference values if any)}

\item{objnames}{vector with names of objects (rows of x)}

\item{compnames}{vector with names used for components}
}
\value{
array with predicted y-values
}
\description{
Compute predictions for response values
}
