% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.getxdecomp}
\alias{pls.getxdecomp}
\title{Compute object with decomposition of x-values}
\usage{
pls.getxdecomp(
  x,
  xscores,
  xloadings,
  xeigenvals,
  xnames = NULL,
  x.attrs = NULL,
  objnames = NULL,
  compnames = NULL
)
}
\arguments{
\item{x}{matrix with predictors, already preprocessed (e.g. mean centered) and cleaned}

\item{xscores}{matrix with X-scores}

\item{xloadings}{matrix with X-loadings}

\item{xeigenvals}{matrix with eigenvalues for X}

\item{xnames}{vector with names of the predictors}

\item{x.attrs}{list with preditors attributes}

\item{objnames}{vector with names of objects (rows of x)}

\item{compnames}{vector with names used for components}
}
\value{
array `ldecomp` object for x-values
}
\description{
Compute object with decomposition of x-values
}
