% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.getxscores}
\alias{pls.getxscores}
\title{Compute matrix with X-scores}
\usage{
pls.getxscores(x, weights, xloadings)
}
\arguments{
\item{x}{matrix with predictors, already preprocessed and cleaned}

\item{weights}{matrix with PLS weights}

\item{xloadings}{matrix with X-loadings}
}
\value{
matrix with X-scores
}
\description{
Compute matrix with X-scores
}
