% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.getydecomp}
\alias{pls.getydecomp}
\title{Compute object with decomposition of y-values}
\usage{
pls.getydecomp(
  y,
  yscores,
  xscores,
  yloadings,
  yeigenvals,
  ynames = NULL,
  y.attrs = NULL,
  x.attrs = NULL,
  objnames = NULL,
  compnames = NULL
)
}
\arguments{
\item{y}{matrix with responses, already preprocessed (e.g. mean centered) and cleaned}

\item{yscores}{matrix with Y-scores}

\item{xscores}{matrix with X-scores}

\item{yloadings}{matrix with Y-loadings}

\item{yeigenvals}{matrix with eigenvalues for Y}

\item{ynames}{vector with names of the responses}

\item{y.attrs}{list with response attributes (e.g. from reference values if any)}

\item{x.attrs}{list with preditors attributes}

\item{objnames}{vector with names of objects (rows of x)}

\item{compnames}{vector with names used for components}
}
\value{
array `ldecomp` object for y-values (or NULL if y is not provided)
}
\description{
Compute object with decomposition of y-values
}
