% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.getZLimits}
\alias{pls.getZLimits}
\title{Compute critical limits for orthogonal distances (Q)}
\usage{
pls.getZLimits(lim.type, alpha, gamma, params)
}
\arguments{
\item{lim.type}{which method to use for calculation of critical limits for residuals}

\item{alpha}{significance level for extreme limits.}

\item{gamma}{significance level for outlier limits.}

\item{params}{distribution parameters returned by ldecomp.getLimParams}
}
\description{
Compute critical limits for orthogonal distances (Q)
}
