% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.readJSON}
\alias{pls.readJSON}
\title{Reads PLS model from JSON file made in web-application (mda.tools/pls).}
\usage{
pls.readJSON(fileName)
}
\arguments{
\item{fileName}{file name (or full path) to JSON file.}
}
\value{
list with PLS model similar to what \code{pls()} creates.
}
\description{
Reads PLS model from JSON file made in web-application (mda.tools/pls).
}
