% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.simplsold}
\alias{pls.simplsold}
\title{SIMPLS algorithm (old implementation)}
\usage{
pls.simplsold(x, y, ncomp, cv = FALSE)
}
\arguments{
\item{x}{a matrix with x values (predictors)}

\item{y}{a matrix with y values (responses)}

\item{ncomp}{number of components to calculate}

\item{cv}{logical, is model calibrated during cross-validation or not}
}
\value{
a list with computed regression coefficients, loadings and scores for x and y matrices,
and weights.
}
\description{
SIMPLS algorithm for calibration of PLS model (old version)
}
\references{
[1]. S. de Jong. SIMPLS: An Alternative approach to partial least squares regression.
Chemometrics and Intelligent Laboratory Systems, 18, 1993 (251-263).
}
