% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.syncResAliases}
\alias{pls.syncResAliases}
\title{Sync result aliases (calres, cvres, testres) from canonical res list}
\usage{
pls.syncResAliases(obj)
}
\arguments{
\item{obj}{PLS model (object of class \code{pls})}
}
\value{
model object with aliases updated
}
\description{
Sync result aliases (calres, cvres, testres) from canonical res list
}
