% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plsda.cal}
\alias{plsda.cal}
\title{Calibrate PLS-DA model}
\usage{
plsda.cal(x, c, ncomp, center, scale, cv, method, light, alpha, coeffs.ci,
  coeffs.alpha, info, exclcols = NULL, exclrows = NULL, ncomp.selcrit)
}
\description{
Calibrate PLS-DA model
}

