% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{predict.ddsimca}
\alias{predict.ddsimca}
\title{DD-SIMCA predictions}
\usage{
\method{predict}{ddsimca}(
  object,
  x,
  c.ref = NULL,
  alpha = object$alpha,
  gamma = object$gamma,
  ...
)
}
\arguments{
\item{object}{a DD-SIMCA model (object of class \code{ddsimca})}

\item{x}{a matrix with x values (predictors)}

\item{c.ref}{a vector with reference class names (same as class names for models)}

\item{alpha}{significance level for making the predictions.}

\item{gamma}{significance level for detection of outliers.}

\item{...}{other optional parameters.}
}
\value{
DD-SIMCA results (an object of class \code{ddsimcares})
}
\description{
Applies DD-SIMCA model to a new data set
}
\details{
See examples in help for \code{\link{ddsimca}} function.
}
