% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{predict.mcr}
\alias{predict.mcr}
\title{MCR predictions}
\usage{
\method{predict}{mcr}(object, x, ...)
}
\arguments{
\item{object}{an MCR model (object of class \code{mcr}).}

\item{x}{spectral values (matrix or data frame).}

\item{...}{other arguments.}
}
\value{
Matrix with contributions
}
\description{
Applies MCR model to a new set of spectra and returns matrix with contributions.
}
