% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrals.R
\name{predict.mcrals}
\alias{predict.mcrals}
\title{MCR ALS predictions}
\usage{
\method{predict}{mcrals}(object, x, ...)
}
\arguments{
\item{object}{an MCR model (object of class \code{mcr}).}

\item{x}{spectral values (matrix or data frame).}

\item{...}{other arguments.}
}
\value{
Matrix with contributions
}
\description{
Applies MCR-ALS model to a new set of spectra and returns matrix with contributions.
}
