% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{predict.pca}
\alias{predict.pca}
\title{PCA predictions}
\usage{
\method{predict}{pca}(object, x, ...)
}
\arguments{
\item{object}{a PCA model (object of class \code{pca}).}

\item{x}{data values (matrix or data frame).}

\item{...}{other arguments.}
}
\value{
PCA results (an object of class \code{pcares})
}
\description{
Applies PCA model to a new data set.
}
