% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.apply}
\alias{prep.apply}
\title{Applies a list with preprocessing methods to a dataset}
\usage{
prep.apply(obj, x)
}
\arguments{
\item{obj}{list with preprocssing methods (created using \code{prep} or \code{prep.fit} function).}

\item{x}{matrix with dataset}
}
\description{
Applies a list with preprocessing methods to a dataset
}
