% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.asjson}
\alias{prep.asjson}
\title{Converts preprocessing model to JSON elements.}
\usage{
prep.asjson(obj)
}
\arguments{
\item{obj}{list with preprocessing methods (created using \code{prep.fit} function).}
}
\value{
stringified JSON.
}
\description{
Converts preprocessing model to JSON elements.
}
