% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.center.params}
\alias{prep.center.params}
\title{Precomputes parameters for centering}
\usage{
prep.center.params(data, type = "mean", center = NULL)
}
\arguments{
\item{data}{a matrix with data values.}

\item{type}{type of statistic to use for centering (\code{'mean'}, or \code{'median'}).}

\item{center}{vector with precomputed values for centering.}
}
\value{
list with parameter values
}
\description{
Precomputes parameters for centering
}
