% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.center}
\alias{prep.center}
\title{Centering data columns.}
\usage{
prep.center(data, type = "mean", center = NULL)
}
\arguments{
\item{data}{a matrix with data values.}

\item{type}{type of statistic to use for centring ('mean', or 'median').}

\item{center}{do not use, required for training of preprocessing model.}
}
\value{
preprocessed data matrix
}
\description{
Centering data columns.
}
