% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.emsc.params}
\alias{prep.emsc.params}
\title{Precomputes parameters for EMSC}
\usage{
prep.emsc.params(data, degree = 0, mspectrum = NULL)
}
\arguments{
\item{data}{a matrix with data values.}

\item{degree}{polynomial degree.}

\item{mspectrum}{reference spectrum.}
}
\value{
list with parameter values
}
\description{
Precomputes parameters for EMSC
}
