% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.emsc}
\alias{prep.emsc}
\title{Applies Extended Multiplicative Scatter Correction to data rows}
\usage{
prep.emsc(data, degree = 0, mspectrum = NULL, lnorm = NULL, A = NULL)
}
\arguments{
\item{data}{a matrix with data values.}

\item{degree}{polynomial degree, if 0 then the result will be the same as for conventional MSC.}

\item{mspectrum}{optional reference spectrum (if not provided, mean spectrum will be used).}

\item{lnorm}{do not use, required for training of preprocessing model.}

\item{A}{do not use, required for training of preprocessing model.}
}
\description{
Applies Extended Multiplicative Scatter Correction to data rows
}
