% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.fit}
\alias{prep.fit}
\title{Fits preprocessing model}
\usage{
prep.fit(obj, x)
}
\arguments{
\item{obj}{list with preprocessing methods (created using \code{prep} or \code{prep.fit} function).}

\item{x}{matrix with training set to be used for computing data dependent parameters}
}
\value{
same list but with updated methods parameters computed based on
the training set.
}
\description{
Fits preprocessing model
}
