% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.from.json}
\alias{prep.from.json}
\title{Converts JSON string to preprocessing model}
\usage{
prep.from.json(str)
}
\arguments{
\item{str}{string with JSON}
}
\value{
list with the methods.
}
\description{
Converts JSON string to preprocessing model
}
