% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.generic}
\alias{prep.generic}
\title{Generic function for preprocessing}
\usage{
prep.generic(x, f, ...)
}
\arguments{
\item{x}{data matrix to be preprocessed}

\item{f}{function for preprocessing}

\item{...}{arguments for the function f}
}
\description{
Generic function for preprocessing
}
