% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.norm.asjson}
\alias{prep.norm.asjson}
\title{Converts preprocessing item from 'prep.norm' method to JSON elements}
\usage{
prep.norm.asjson(params, npred, left = 0, right = 1)
}
\arguments{
\item{params}{model parameters precomputed by using prep.fit()}

\item{npred}{number of predictors in original dataset}

\item{left}{index of first variable after trimming (if any)}

\item{right}{index of last variable after trimming (if any)}
}
\value{
list with main elements required for JSON with preprocessing model
compatible with preprocessing web-application (ml, mpl, mp, mpl_new, mp_new, info)
}
\description{
Converts preprocessing item from 'prep.norm' method to JSON elements
}
