% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.norm.params}
\alias{prep.norm.params}
\title{Precomputes parameters for normalization}
\usage{
prep.norm.params(data, type = "area", col.ind = NULL, ref.spectrum = NULL)
}
\arguments{
\item{data}{a matrix with data values.}

\item{type}{type of normalization.}

\item{col.ind}{indices of columns for IS normalization.}

\item{ref.spectrum}{reference spectrum for PQN normalization.}
}
\value{
list with parameter values
}
\description{
Precomputes parameters for normalization
}
