% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{Class for preprocessing object/item.}
\usage{
prep(name, ...)
}
\arguments{
\item{name}{short text with name for the preprocessing method.}

\item{...}{a list with named parameters for the method (if empty - default parameters will be used).}
}
\description{
Class for preprocessing object/item.
}
\details{
Use this class to create a list with a sequence of preprocessing methods to keep them together
in right order and with defined parameters. The list/object can be provided as an extra argument
to any modelling function (e.g. \code{pca}, \code{pls}, etc), so the optimal model parameters and
the optimal preprocessing will be stored together and can be applied to a raw data by using
method \code{predict}.

For your own preprocessing method you need to create a function, which takes matrix with values
(dataset) as the first argument, does something and then return a matrix with the same dimension
and same attributes as the result. The method can have any number of optional parameters.

See Bookdown tutorial for details.
}
