% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.readJSON}
\alias{prep.readJSON}
\title{Reads preprocessing model from JSON file made in web-application (mda.tools/prep).}
\usage{
prep.readJSON(fileName)
}
\arguments{
\item{fileName}{file name (or full path) to JSON file.}
}
\value{
list with preprocessing model similar to what \code{prep.fit()} creates.
}
\description{
Reads preprocessing model from JSON file made in web-application (mda.tools/prep).
}
