% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.ref2km}
\alias{prep.ref2km}
\title{Kubelka-Munk transformation}
\usage{
prep.ref2km(data)
}
\arguments{
\item{data}{a matrix with spectra values (absolute reflectance values)}
}
\value{
preprocessed spectra.
}
\description{
Applies Kubelka-Munk (km) transformation to data matrix (spectra)
}
\details{
Kubelka-Munk is useful preprocessing method for diffuse reflection spectra (e.g. taken for
powders or rough surface). It transforms the reflectance spectra R to K/M units as follows:
(1 - R)^2 / 2R
}
