% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.savgol}
\alias{prep.savgol}
\title{Savitzky-Golay filter}
\usage{
prep.savgol(data, width = 3, porder = 1, dorder = 0, w = NULL)
}
\arguments{
\item{data}{a matrix with data values}

\item{width}{width of the filter window}

\item{porder}{order of polynomial used for smoothing}

\item{dorder}{order of derivative to take (0 - no derivative)}

\item{w}{do not use, required for training of preprocessing model.}
}
\description{
Applies Savitzky-Golay filter to the rows of data matrix
}
\details{
The function implements algorithm described in [1] which handles the edge points correctly and
does not require to cut the spectra.
}
\references{
1. Peter A. Gorry. General least-squares smoothing and differentiation by the convolution
(Savitzky-Golay) method. Anal. Chem. 1990, 62, 6, 570–573, https://doi.org/10.1021/ac00205a007.
}
