% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.scale.fromjson}
\alias{prep.scale.fromjson}
\title{Converts JSON elements to preprocessing item for 'prep.scale' method}
\usage{
prep.scale.fromjson(mp, mp_new, left = 0, right = 1, left.tot = 0)
}
\arguments{
\item{mp}{model parameters from JSON (user defined)}

\item{mp_new}{mode parameters from JSON after training}

\item{left}{index of first variable after trimming (if any)}

\item{right}{index of last variable after trimming (if any)}

\item{left.tot}{total, accumulated shift on the left part}
}
\value{
\code{prep} object for the method
}
\description{
Converts JSON elements to preprocessing item for 'prep.scale' method
}
