% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.scale.params}
\alias{prep.scale.params}
\title{Precomputes parameters for scaling}
\usage{
prep.scale.params(data, type = "sd", max.cov = 0, scale = NULL)
}
\arguments{
\item{data}{a matrix with data values.}

\item{type}{type of statistic to use for scaling (\code{'sd'}, \code{'iqr'}, \code{'range'}, or \code{'pareto'})}

\item{max.cov}{columns that have coefficient of variation (in percent) below or equal to `max.cov` will not
be scaled.}

\item{scale}{vector with precomputed values for scaling.}
}
\value{
list with parameter values
}
\description{
Precomputes parameters for scaling
}
