% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.scale}
\alias{prep.scale}
\title{Scaling data columns.}
\usage{
prep.scale(data, type = "sd", max.cov = 0, scale = NULL)
}
\arguments{
\item{data}{a matrix with data values.}

\item{type}{type of statistic to use for scaling ('sd', 'iqr', 'range', 'pareto')}

\item{max.cov}{columns that have coefficient of variation (in percent) below or equal to `max.cov` will not
be scaled.}

\item{scale}{do not use, required for training of preprocessing model.}
}
\value{
preprocessed data matrix
}
\description{
Scaling data columns.
}
