% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.spikes}
\alias{prep.spikes}
\title{Remove spikes from Raman spectra}
\usage{
prep.spikes(data, width = 5, threshold = 6)
}
\arguments{
\item{data}{a matrix with data values}

\item{width}{width of the moving median filter}

\item{threshold}{threshold to compare modified z-score value with to detect the spike.}
}
\value{
data matrix with processed values
}
\description{
Removes spikes from Raman spectra using median absolute deviation computed for signal
differences. See [1] for more details
}
\references{
1. Darren A. Whitaker, Kevin Hayes, A simple algorithm for despiking Raman spectra,
Chemometrics and Intelligent Laboratory Systems, 179, 2018, pp. 82-84,
10.1016/j.chemolab.2018.06.009.
}
