% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.varsel}
\alias{prep.varsel}
\title{Variable selection}
\usage{
prep.varsel(data, var.ind)
}
\arguments{
\item{data}{a matrix with data values}

\item{var.ind}{indices of variables (columns) to select, can bet either numeric or logical}
}
\value{
data matrix with the selected variables (columns)
}
\description{
Returns dataset with selected variables
}
