% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.writeJSON}
\alias{prep.writeJSON}
\title{Saves preprocessing model to JSON file which can be loaded to web-application (mda.tools/prep).}
\usage{
prep.writeJSON(obj, fileName)
}
\arguments{
\item{obj}{list with preprocssing methods (created using \code{prep.fit} function).}

\item{fileName}{file name (or full path) to JSON file to save the model into.}
}
\description{
Saves preprocessing model to JSON file which can be loaded to web-application (mda.tools/prep).
}
