% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{preparePlotData}
\alias{preparePlotData}
\title{Take dataset and prepare them for plot}
\usage{
preparePlotData(data)
}
\arguments{
\item{data}{dataset (vector, matrix or data frame)}
}
\description{
The function checks that `data` contains correct numeric values, checks for
mandatory attributes (row and column names, x- and y-axis values and names, etc.)
and adds them if necessary.

Another thing is to remove hidden columns and split the rest to visible and hidden
values (if excluded rows are present).
}
