% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prepCalData}
\alias{prepCalData}
\title{Prepares calibration data}
\usage{
prepCalData(x, exclrows = NULL, exclcols = NULL)
}
\arguments{
\item{x}{matrix or data frame with values (calibration set)}

\item{exclrows}{rows to be excluded from calculations (numbers, names or vector with logical values)}

\item{exclcols}{columns to be excluded from calculations (numbers, names or vector with logical values)}
}
\description{
Prepares calibration data
}
