\name{print.classres}
\alias{print.classres}
\title{Print information about classification result object}
\usage{
\method{print}{classres}(x, str = NULL, ...)
}
\arguments{
  \item{x}{classification results (object of class
  \code{plsdares}, \code{simcamres}, etc.).}

  \item{str}{User specified text (e.g. to be used for
  particular method, like PLS-DA, etc).}

  \item{...}{other arguments}
}
\description{
Generic \code{print} function for classification results.
Prints information about major fields of the object.
}

