% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ipls.R
\name{print.ipls}
\alias{print.ipls}
\title{Print method for iPLS}
\usage{
\method{print}{ipls}(x, ...)
}
\arguments{
\item{x}{a iPLS (object of class \code{ipls})}

\item{...}{other arguments}
}
\description{
Prints information about the iPLS object structure
}

