\name{print.ldecomp}
\alias{print.ldecomp}
\title{Print function for linear decomposition}
\usage{
print.ldecomp(x, str = NULL, ...)
}
\arguments{
  \item{x}{object of class \code{ldecomp}}

  \item{str}{user specified text to show as a description
  of the object}

  \item{...}{other arguments}
}
\description{
Generic \code{print} function for linear decomposition.
Prints information about the \code{ldecomp} object.
}

