% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.R
\name{print.pca}
\alias{print.pca}
\title{Print method for PCA model object}
\usage{
\method{print}{pca}(x, ...)
}
\arguments{
\item{x}{a PCA model (object of class \code{pca})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}

