% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plsres.R
\name{print.plsres}
\alias{print.plsres}
\title{print method for PLS results object}
\usage{
\method{print}{plsres}(x, ...)
}
\arguments{
\item{x}{PLS results (object of class \code{plsres})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}

