% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{print.prepmodel}
\alias{print.prepmodel}
\title{Print the information about methods in the preprocessing model.}
\usage{
\method{print}{prepmodel}(x, ...)
}
\arguments{
\item{x}{preprocessing model (created by \code{\link{prep.fit}}).}

\item{...}{potential further arguments (required for Method/Generic reasons).}
}
\value{
the \code{x} argument (invisibly).
}
\description{
Print the information about methods in the preprocessing model.
}
