% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randtest.R
\name{print.randtest}
\alias{print.randtest}
\title{Print method for randtest object}
\usage{
\method{print}{randtest}(x, ...)
}
\arguments{
\item{x}{a randomization test results (object of class \code{randtest})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}

