% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcoeffs.R
\name{print.regcoeffs}
\alias{print.regcoeffs}
\title{print method for regression coefficients class}
\usage{
\method{print}{regcoeffs}(x, ...)
}
\arguments{
\item{x}{regression coefficients object (class \code{regcoeffs})}

\item{...}{other arguments}
}
\description{
prints regression coeffocoent values for given response number and amount of components
}
