% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodel.R
\name{print.regmodel}
\alias{print.regmodel}
\title{Print method for PLS model object}
\usage{
\method{print}{regmodel}(x, ...)
}
\arguments{
\item{x}{a regression model (object of class \code{regmodel})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}
