% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R, R/simca.R
\name{print.simca}
\alias{print.simca}
\title{Print method for SIMCA model object}
\usage{
\method{print}{simca}(x, ...)

\method{print}{simca}(x, ...)
}
\arguments{
\item{x}{a SIMCA model (object of class \code{simca})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure

Prints information about the object structure
}
