% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcamres.R
\name{print.simcamres}
\alias{print.simcamres}
\title{Print method for SIMCAM results object}
\usage{
\method{print}{simcamres}(x, ...)
}
\arguments{
\item{x}{SIMCAM results (object of class \code{simcamres})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}
