% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R, R/simcares.R
\name{print.simcares}
\alias{print.simcares}
\title{Print method for SIMCA results object}
\usage{
\method{print}{simcares}(x, ...)

\method{print}{simcares}(x, ...)
}
\arguments{
\item{x}{SIMCA results (object of class \code{simcares})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure

Prints information about the object structure
}
