% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{print.supure}
\alias{print.supure}
\title{Print method for SUPURE object}
\usage{
\method{print}{supure}(x, ...)
}
\arguments{
\item{x}{unmixing results (object of class \code{supure})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}

