% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{process_excluded_rows}
\alias{process_excluded_rows}
\title{Prepare x and y values for excluded rows}
\usage{
process_excluded_rows(plot_data, type)
}
\arguments{
\item{plot_data}{list with data and parameters prepared for plot by `prepare_data_for_plot()` function}

\item{type}{type of plot}
}
\description{
Prepare x and y values for excluded rows
}
