% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{processLimType}
\alias{processLimType}
\title{Make correction to limit types names.}
\usage{
processLimType(limType)
}
\arguments{
\item{limType}{the limit type provided by user}
}
\description{
Make correction to limit types names.
}
