% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{processMembers}
\alias{processMembers}
\title{Computes classification outcomes for target class members.}
\usage{
processMembers(res, indMembers)
}
\arguments{
\item{res}{a list with classification outcomes created by method \code{\link{classify}} (part of it will be filled by this method).}

\item{indMembers}{a vector with logical values pointing on data items corresponding to class members.}
}
\value{
the \code{res} list where roles vector is filled for class members, plus values for TP and FN.
}
\description{
Computes classification outcomes for target class members.
}
