% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{processStrangers}
\alias{processStrangers}
\title{Computes classification outcomes for members of non-target classes.}
\usage{
processStrangers(res, indStrangers)
}
\arguments{
\item{res}{a list with classification outcomes created by method \code{\link{classify}} (part of it will be filled by this method).}

\item{indStrangers}{a vector with logical values pointing on data items corresponding to members of non-target classes.}
}
\value{
the \code{res} list where roles vector is filled for class members, values for TN and FP, as
well as statistics for computing Type II error and related things (beta, s, f0, hz, Mz, Sz, k, m).
}
\description{
Computes classification outcomes for members of non-target classes.
}
