% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{readJSON}
\alias{readJSON}
\title{Reads models from JSON file made in web-application (mda.tools).}
\usage{
readJSON(fileName)
}
\arguments{
\item{fileName}{file name (or full path) to JSON file.}
}
\value{
object with model (prep, pls, pca, ddsimca, pls1da, pls2da).
}
\description{
Reads models from JSON file made in web-application (mda.tools).
}
