% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{regres.err}
\alias{regres.err}
\title{Error of prediction}
\usage{
regres.err(y.ref, y.pred)
}
\arguments{
\item{y.ref}{vector with reference values}

\item{y.pred}{matrix with predicted values}
}
\description{
Calculates array of differences between predicted and reference values.
}
