% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{regres.rmse}
\alias{regres.rmse}
\title{RMSE}
\usage{
regres.rmse(err)
}
\arguments{
\item{err}{vector with difference between reference and predicted y-values}
}
\description{
Calculates matrix with root mean squared error of prediction for every response and components.
}
