% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{regres.slope}
\alias{regres.slope}
\title{Slope}
\usage{
regres.slope(y.pred, y.ref)
}
\arguments{
\item{y.pred}{matrix with predicted values}

\item{y.ref}{vector with reference values}
}
\description{
Calculates matrix with slope of predicted and measured values for every response and components.
}
