% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{regress.addattrs}
\alias{regress.addattrs}
\title{Add names and attributes to matrix with statistics}
\usage{
regress.addattrs(stat, attrs, name)
}
\arguments{
\item{stat}{matrix with statistics}

\item{err}{array with error of predictions}
}
\description{
Add names and attributes to matrix with statistics
}
