% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{repmat}
\alias{repmat}
\title{Replicate matric x}
\usage{
repmat(x, nrows, ncols = nrows)
}
\arguments{
\item{x}{original matrix}

\item{nrows}{number of times replicate matrix row wise}

\item{ncols}{number of times replicate matrix columns wise}
}
\description{
Replicate matric x
}
