% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{reslim.chisq}
\alias{reslim.chisq}
\title{Calculates critical limits or statistic values for Q-residuals using Chi-squared distribution}
\usage{
reslim.chisq(Q, alpha = 0.05, gamma = 0.01, Qlim = NULL,
  return = "limits")
}
\arguments{
\item{Q}{vector with Q-residuals for selected component}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}

\item{Qlim}{vector with Q limits for selected number of components (from model)}

\item{return}{what to return: \code{'limits'} or \code{'probability'}}
}
\description{
The method is based on Chi-squared distribution with DF = 2 * (m(Q)/s(Q)^2
}
