% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{reslim.dd}
\alias{reslim.dd}
\title{Statistical limits for Q and T2 residuals using Data Driven approach}
\usage{
reslim.dd(Q, T2, type = "ddmoments", alpha = 0.05, gamma = 0.01,
  Qlim = NULL, T2lim = NULL, return = "limits")
}
\arguments{
\item{Q}{vector with Q-residuals for selected component}

\item{T2}{vector with T2-residuals for selected component}

\item{type}{which estimator to use: 'moments' or 'robust'}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}

\item{Qlim}{vector with Q limits for selected number of components (from model)}

\item{T2lim}{vector with T2 limits for selected number of components (from model)}

\item{return}{what to return: \code{'limits'} or \code{'probability'}}
}
\description{
Method is based on paper by Pomerantsev, Rodionova (JChem, 2014)
}
