% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{reslim.hotelling}
\alias{reslim.hotelling}
\title{Calculates critical limits for T2-residuals using Hotelling T2 distribution}
\usage{
reslim.hotelling(ncomp, T2 = NULL, alpha = 0.05, gamma = 0.01,
  T2lim = NULL, return = "limits")
}
\arguments{
\item{ncomp}{number of components}

\item{T2}{vector with T2-residuals for selected component}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}

\item{T2lim}{T2 limits from a PCA model (needed for probabilities)}

\item{return}{what to return: \code{'limits'} or \code{'probability'}}
}
\description{
The method is based on n
}
