% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{reslim.jm}
\alias{reslim.jm}
\title{Calculates critical limits for Q-residuals using classic JM approach}
\usage{
reslim.jm(eigenvals, Q, ncomp, alpha = 0.05, gamma = 0.01,
  return = "limits")
}
\arguments{
\item{eigenvals}{vector with eigenvalues for all variables}

\item{Q}{vector with Q-residuals for selected component}

\item{ncomp}{number of components}

\item{alpha}{significance level for extreme objects}

\item{gamma}{significance level for outliers}

\item{return}{what to return: \code{'limits'} or \code{'probability'}}
}
\description{
The method is based on
}
