% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcv.R
\name{rotationMatrixToX1}
\alias{rotationMatrixToX1}
\title{Creates a rotation matrix to map a vector x to [1 0 0 ... 0]}
\usage{
rotationMatrixToX1(x)
}
\arguments{
\item{x}{Vector (sequence with J coordinates)}
}
\value{
Rotation matrix (JxJ)
}
\description{
Creates a rotation matrix to map a vector x to [1 0 0 ... 0]
}
