% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{selectCompNum.pca}
\alias{selectCompNum.pca}
\title{Select optimal number of components for PCA model}
\usage{
\method{selectCompNum}{pca}(obj, ncomp, ...)
}
\arguments{
\item{ncomp}{number of components to select}

\item{...}{other parameters if any}

\item{model}{PCA model (object of class \code{pca})}
}
\value{
the same model with selected number of components
}
\description{
Allows user to select optimal number of components for a PCA model
}
