% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{selratio}
\alias{selratio}
\title{Selectivity ratio calculation}
\usage{
selratio(obj, ncomp = obj$ncomp.selected)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{ncomp}{number of components to count}
}
\value{
array \code{nvar x ncomp x ny} with selectivity ratio values
}
\description{
Calculates selectivity ratio for each component and response variable in
the PLS model
}
\references{
[1] Tarja Rajalahti et al. Chemometrics and Laboratory Systems, 95 (2009), pp. 35-48.
}
