% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{setParams}
\alias{setParams}
\title{Set model parameters other than number of components (generic function)}
\usage{
setParams(obj, ...)
}
\arguments{
\item{obj}{model object, e.g. \code{ddsimca}}

\item{...}{other parameters relevant for the method}
}
\description{
Set model parameters other than number of components (generic function)
}
