% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{setResLimits.pca}
\alias{setResLimits.pca}
\title{Recalculate and set critical limits for residuals}
\usage{
setResLimits.pca(model, lim.type, alpha, gamma)
}
\arguments{
\item{model}{PCA model (object of class \code{pca})}

\item{lim.type}{which type of limits to use (see help for \code{pca} class for details)}

\item{alpha}{significance level for critical limites (extreme objects)}

\item{gamma}{significance level for outliers}
}
\value{
object with PCA model and recalculated limits
}
\description{
Allows to recompute new critical limits for existing PCA/SIMCA model
}
