% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{setResLimits}
\alias{setResLimits}
\title{Set residual limits for PCA model}
\usage{
setResLimits(obj, ...)
}
\arguments{
\item{obj}{a SIMCA model}

\item{...}{other parameters}
}
\description{
Calculates and set critical limits for residuals of PCA model
}
