% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{showDistanceLimits}
\alias{showDistanceLimits}
\title{Show residual distance limits}
\usage{
showDistanceLimits(obj, ...)
}
\arguments{
\item{obj}{a model object}

\item{...}{other parameters}
}
\description{
Calculates and set critical limits for residuals of PCA model
}
