% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{showLabels}
\alias{showLabels}
\title{Show labels on plot}
\usage{
showLabels(ps, show.excluded = FALSE, pos = 3, cex = 0.65,
  col = "darkgray", force.x.values = NULL, bwd = 0.8)
}
\arguments{
\item{ps}{`plotseries` object}

\item{pos}{position of the labels relative to the data points}

\item{cex}{size of the labels text}

\item{col}{color of the labels text}
}
\description{
Show labels on plot
}
