\name{showPredictions}
\alias{showPredictions}
\title{Predictions}
\usage{
showPredictions(obj, ...)
}
\arguments{
  \item{obj}{a model or result object}

  \item{...}{other arguments}
}
\description{
Predictions
}
\details{
Generic function for showing predicted values for
classification or regression model or results
}

