% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simca.R
\name{simca.crossval}
\alias{simca.crossval}
\title{Cross-validation of a SIMCA model}
\usage{
simca.crossval(model, x, cv, center = T, scale = F)
}
\arguments{
\item{model}{a SIMCA model (object of class \code{simca})}

\item{x}{a matrix with x values (predictors from calibration set)}

\item{cv}{number of segments (if cv = 1, full cross-validation will be used)}

\item{center}{logical, do mean centering or not}

\item{scale}{logical, do standardization or not}
}
\value{
object of class \code{simcares} with results of cross-validation
}
\description{
Does the cross-validation of a SIMCA model
}
