\name{simcam.getPerformanceStatistics}
\alias{simcam.getPerformanceStatistics}
\title{Performance statistics for SIMCAM model}
\usage{
simcam.getPerformanceStatistics(model)
}
\arguments{
  \item{model}{SIMCAM model (object of class
  \code{simcam})}
}
\description{
Calculates discrimination power and distance between models
for SIMCAM model.
}

