% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcam.R
\name{simcam.getPerformanceStats}
\alias{simcam.getPerformanceStats}
\title{Performance statistics for SIMCAM model}
\usage{
simcam.getPerformanceStats(models, classnames)
}
\arguments{
\item{models}{list with SIMCA models (as provided to simcam class)}

\item{classnames}{names of the classes for each model}
}
\description{
Calculates discrimination power and distance between individual SIMCA models.
}
