\name{speye}
\alias{speye}
\title{Sparse identity matrix}
\usage{
speye(m = 1, n = NULL)
}
\arguments{
  \item{m}{number of rows in the matrix}

  \item{n}{number of columns in the matrix (if null, m x m
  matrix will be generated)}
}
\description{
Generates sparse identity matrix for given size
}

