% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{split_plot_data}
\alias{split_plot_data}
\title{Split dataset to x and y values depending on plot type}
\usage{
split_plot_data(plot_data, type)
}
\arguments{
\item{plot_data}{list with data and parameters prepared for plot by `prepare_data_for_plot()` function}

\item{type}{type of plot}
}
\description{
Split dataset to x and y values depending on plot type
}
