% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{splitExcludedData}
\alias{splitExcludedData}
\title{Split the excluded part of data}
\usage{
splitExcludedData(data, type)
}
\arguments{
\item{data}{matrix with hidden data values}

\item{type}{type of plot}
}
\description{
Split the excluded part of data
}
