% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{splitPlotData}
\alias{splitPlotData}
\title{Split dataset to x and y values depending on plot type}
\usage{
splitPlotData(data, type)
}
\arguments{
\item{data}{matrix with data values (visible or hidden)}

\item{type}{type of plot}
}
\description{
Split dataset to x and y values depending on plot type
}
