% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{str.prepmodel}
\alias{str.prepmodel}
\title{Show the information about methods in the preprocessing model.}
\usage{
\method{str}{prepmodel}(object, ...)
}
\arguments{
\item{object}{preprocessing model (created by \code{\link{prep.fit}}).}

\item{...}{potential further arguments (required for Method/Generic reasons).}
}
\description{
Show the information about methods in the preprocessing model.
}
