% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classres.R
\name{summary.classres}
\alias{summary.classres}
\title{Summary statistics about classification result object}
\usage{
\method{summary}{classres}(object, ncomp = NULL, nc = NULL, ...)
}
\arguments{
\item{object}{classification results (object of class \code{plsdares}, \code{simcamres}, etc.).}

\item{ncomp}{which number of components to make the plot for (can be one value for all classes or vector with
separate values for each, if NULL - model selected number will be used).}

\item{nc}{if there are several classes, which class to make the plot for (NULL - summary for all classes).}

\item{...}{other arguments}
}
\description{
Generic \code{summary} function for classification results. Prints performance values for the 
results.
}
