% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ipls.R
\name{summary.ipls}
\alias{summary.ipls}
\title{Summary for iPLS results}
\usage{
\method{summary}{ipls}(object, glob.ncomp = NULL, ...)
}
\arguments{
\item{object}{a iPLS (object of class \code{ipls})}

\item{glob.ncomp}{number of components for global PLS model with all intervals}

\item{...}{other arguments}
}
\description{
Shows statistics and algorithm parameters for iPLS results.
}
\details{
The method shows information on the algorithm parameters as well as a table with selected or
excluded interval. The table has the following columns: 'step' showing on which iteration
an interval was selected or excluded, 'start and 'end' show variable indices for the interval,
'nComp' is a number of components used in a model, 'RMSE' is RMSECV for the model and 'R2' is
coefficient of determination for the same model.
}

