% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{summary.ldecomp}
\alias{summary.ldecomp}
\title{Summary statistics for linear decomposition}
\usage{
\method{summary}{ldecomp}(object, str = NULL, ...)
}
\arguments{
\item{object}{object of class \code{ldecomp}}

\item{str}{user specified text to show as a description of the object}

\item{...}{other arguments}
}
\description{
Generic \code{summary} function for linear decomposition. Prints statistic about the decomposition.
}

