% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrals.R
\name{summary.mcrals}
\alias{summary.mcrals}
\title{Summary method for mcrals object}
\usage{
\method{summary}{mcrals}(object, ...)
}
\arguments{
\item{object}{\code{mcrals} object}

\item{...}{other arguments}
}
\description{
Shows some statistics (explained variance, etc) for the case.
}
