% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrpure.R
\name{summary.mcrpure}
\alias{summary.mcrpure}
\title{Summary method for mcrpure object}
\usage{
\method{summary}{mcrpure}(object, ...)
}
\arguments{
\item{object}{\code{mcrpure} object}

\item{...}{other arguments}
}
\description{
Shows some statistics (explained variance, etc) for the case.
}
