\name{summary.pcares}
\alias{summary.pcares}
\title{Summary method for PCA results object}
\usage{
\method{summary}{pcares}(object, ...)
}
\arguments{
  \item{object}{PCA results (object of class
  \code{pcares})}

  \item{...}{other arguments}
}
\description{
Shows some statistics (explained variance, eigenvalues)
about the results.
}

