% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{summary.prepmodel}
\alias{summary.prepmodel}
\title{Show summary of the preprocessing model.}
\usage{
\method{summary}{prepmodel}(object, ...)
}
\arguments{
\item{object}{preprocessing model (created by \code{\link{prep.fit}}).}

\item{...}{potential further arguments (required for Method/Generic reasons).}
}
\value{
the \code{object} argument (invisibly).
}
\description{
Show summary of the preprocessing model.
}
