% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtest.R
\name{summary.randtest}
\alias{summary.randtest}
\title{Summary method for randtest object}
\usage{
\method{summary}{randtest}(object, ...)
}
\arguments{
\item{object}{randomization test results (object of class \code{randtest})}

\item{...}{other arguments}
}
\description{
Shows summary for randomization test results.
}

