% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcoeffs.R
\name{summary.regcoeffs}
\alias{summary.regcoeffs}
\title{Summary method for regcoeffs object}
\usage{
\method{summary}{regcoeffs}(object, ncomp = 1, ny = 1, alpha = 0.05,
  ...)
}
\arguments{
\item{object}{object of class \code{regcoeffs}}

\item{ncomp}{how many components to use}

\item{ny}{which y variable to show the summary for}

\item{alpha}{significance level for confidence interval (if statistics available)}

\item{...}{other arguments}
}
\description{
Shows estimated coefficients and statistics (if available).
}
\details{
Statistcs are shown if Jack-Knifing was used when model is calibrated.
}
