% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcamres.R
\name{summary.simcamres}
\alias{summary.simcamres}
\title{Summary method for SIMCAM results object}
\usage{
\method{summary}{simcamres}(object, nc = seq_len(object$nclasses), ...)
}
\arguments{
\item{object}{SIMCAM results (object of class \code{simcamres})}

\item{nc}{number of class to show summary for (can be vector)}

\item{...}{other arguments}
}
\description{
Shows performance statistics for the results.
}
