% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R, R/simcares.R
\name{summary.simcares}
\alias{summary.simcares}
\title{Summary method for SIMCA results object}
\usage{
\method{summary}{simcares}(object, ...)

\method{summary}{simcares}(object, ...)
}
\arguments{
\item{object}{SIMCA results (object of class \code{simcares})}

\item{...}{other arguments}
}
\description{
Shows performance statistics for the results.

Shows performance statistics for the results.
}
