% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{summary.supure}
\alias{summary.supure}
\title{Summary method for SUPURE object}
\usage{
\method{summary}{supure}(object, ...)
}
\arguments{
\item{object}{unmixing results (object of class \code{supure})}

\item{...}{other arguments}
}
\description{
Shows some statistics () for the .
}

