% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{supure}
\alias{supure}
\title{Makes unmixing of set of spectra to pure components.}
\usage{
supure(data, ncomp, suggest = NULL, savgol.deriv = 2, savgol.width = NULL,
  savgol.porder = 2, wavelength = NULL)
}
\arguments{
\item{data}{a matrix with spectral values}

\item{ncomp}{number of pure components to identify}

\item{suggest}{a sequence with variable numbers suggested as pure variables for each component}

\item{savgol.deriv}{a derivative order}

\item{savgol.width}{a width of filter used to smooth signal prior to take a derivative}

\item{savgol.porder}{a polynomial order used for smoothing}

\item{wavelength}{a vector with wavelength if necessary}
}
\value{
A \code{supure} object - list with several fields, including:
}
\description{
Makes unmixing of set of spectra to pure components.
}

