% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrpure.R
\name{unmix.mcrpure}
\alias{unmix.mcrpure}
\title{Unmix spectral data using pure variables estimated before}
\usage{
unmix.mcrpure(obj, x)
}
\arguments{
\item{obj}{\code{mcrpure} object}

\item{x}{matrix with spectra}
}
\value{
Returns a list with resolved spectra and contributions (matrices).
}
\description{
Unmix spectral data using pure variables estimated before
}
