% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{writeCSV.ddsimcares}
\alias{writeCSV.ddsimcares}
\title{Save DD-SIMCA results to CSV file}
\usage{
\method{writeCSV}{ddsimcares}(
  res,
  fileName,
  name = "cal",
  sep = ",",
  dataFile = "",
  ncomp = res$ncomp.selected,
  limType = "classic",
  src = "mdatools for R",
  ...
)
}
\arguments{
\item{res}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{fileName}{name (or full path) to CSV file to be created.}

\item{name}{short name of the result object (e.g. \code{"cal"}, \code{"test"}. etc.).}

\item{sep}{values separator (either \code{","} or \code{";"}).}

\item{dataFile}{optional, name of the data file used to create the results.}

\item{ncomp}{model complexity (number of components) to compute the classiciation results for.}

\item{limType}{limit type to use ('classic' or 'robust')}

\item{src}{optional, source of the results}

\item{...}{other optional parameters}
}
\description{
Saves DD-SIMCA results to CSV file with structure identical to
the one generated by web-application at https://mda.tools/ddsimca
}
