% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcares.R
\name{writeCSV.pcares}
\alias{writeCSV.pcares}
\title{Save PCA results to CSV file}
\usage{
\method{writeCSV}{pcares}(
  res,
  fileName,
  name,
  sep = ",",
  dataFile = "",
  classes = NULL,
  model = NULL,
  ...
)
}
\arguments{
\item{res}{PLS results (object of class \code{plsres})}

\item{fileName}{name (or full path) to CSV file to be created.}

\item{name}{short name of the result object (e.g. \code{"cal"}, \code{"test"}. etc.).}

\item{sep}{values separator (either \code{","} or \code{";"}).}

\item{dataFile}{optional, name of the data file used to create the results.}

\item{classes}{vector with class names for every object in the dataset (if any)}

\item{model}{optional, if PCA model is provided, loadings and vectors for centering/scaling will be added}

\item{...}{other optional parameters}
}
\description{
Saves PLS results to CSV file with structure identical to
the one generated by web-application at https://mda.tools/pls
}
