% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsres.R
\name{writeCSV.plsres}
\alias{writeCSV.plsres}
\title{Save PLS results to CSV file}
\usage{
\method{writeCSV}{plsres}(res, fileName, name, sep = ",", dataFile = "", ...)
}
\arguments{
\item{res}{PLS results (object of class \code{plsres})}

\item{fileName}{name (or full path) to CSV file to be created.}

\item{name}{short name of the result object (e.g. \code{"cal"}, \code{"test"}. etc.).}

\item{sep}{values separator (either \code{","} or \code{";"}).}

\item{dataFile}{optional, name of the data file used to create the results.}

\item{...}{other optional parameters}
}
\description{
Saves PLS results to CSV file with structure identical to
the one generated by web-application at https://mda.tools/pls
}
