% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{writeCSV}
\alias{writeCSV}
\title{Method to write outcomes of any result object to CSV file}
\usage{
writeCSV(res, fileName, name, sep = ",", dataFile = "", ...)
}
\arguments{
\item{res}{result object (\code{plsres}, \code{pcares}, etc.).}

\item{fileName}{name (or full path) to CSV file to be created.}

\item{name}{short name of the result object (e.g. \code{"cal"}, \code{"test"}. etc.).}

\item{sep}{values separator (either \code{","} or \code{";"}).}

\item{dataFile}{optional, name of the data file used to create the results.}

\item{...}{other optional parameters}
}
\description{
Method to write outcomes of any result object to CSV file
}
