% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{writeJSON}
\alias{writeJSON}
\title{Save model as JSON file}
\usage{
writeJSON(obj, fileName, ...)
}
\arguments{
\item{obj}{model object, e.g. \code{pca}}

\item{fileName}{name (or full path) to JSON file to be created.}

\item{...}{other optional parameters}
}
\description{
Save model as JSON file
}
