% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{arr2int}
\alias{arr2int}
\title{Converts array (vector) with number of excluded variables to intervals}
\usage{
arr2int(arr)
}
\arguments{
\item{arr}{array with the variable indices}
}
\description{
Converts array (vector) with number of excluded variables to intervals
}
