% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{as.matrix.ldecomp}
\alias{as.matrix.ldecomp}
\title{as.matrix method for ldecomp object}
\usage{
\method{as.matrix}{ldecomp}(x, ...)
}
\arguments{
\item{x}{object of class \code{ldecomp}}

\item{...}{other arguments}
}
\description{
Generic \code{as.matrix} function for linear decomposition. Returns a matrix with information 
about the decomposition.
}
