% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsdares.R
\name{as.matrix.plsdares}
\alias{as.matrix.plsdares}
\title{as.matrix method for PLS-DA results}
\usage{
\method{as.matrix}{plsdares}(x, ncomp = NULL, nc = 1, ...)
}
\arguments{
\item{x}{PLS-DA results (object of class \code{plsdares})}

\item{ncomp}{number of components to calculate the statistics for (if NULL gets for all components)}

\item{nc}{for which class to calculate the statistics for}

\item{...}{other arguments}
}
\description{
Returns a matrix with model performance statistics for PLS-DA results
}
