% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{asjson.ddsimca}
\alias{asjson.ddsimca}
\title{Converts object with DD-SIMCA model to JSON string compatible with web-application.}
\usage{
\method{asjson}{ddsimca}(
  obj,
  limType = "classic",
  alpha = obj$alpha,
  gamma = obj$gamma,
  ncomp = obj$ncomp.selected
)
}
\arguments{
\item{obj}{Object with DD-SIMCA model (from \code{\link{ddsimca}}).}
}
\value{
stringigied JSON
}
\description{
Converts object with DD-SIMCA model to JSON string compatible with web-application.
}
