% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{asjson.pls}
\alias{asjson.pls}
\title{Converts object with PLS model to JSON string compatible with web-application.}
\usage{
\method{asjson}{pls}(obj, ...)
}
\arguments{
\item{obj}{Object with PLS model (from \code{\link{pls}}).}

\item{...}{other arguments}
}
\value{
stringified JSON
}
\description{
Converts object with PLS model to JSON string compatible with web-application.
}
