% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize text or vector with text values}
\usage{
capitalize(str)
}
\arguments{
\item{str}{text of vector with text values}
}
\description{
Capitalize text or vector with text values
}
