% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{chisq.prob}
\alias{chisq.prob}
\title{Calculate probabilities for distance values using Chi-square distribution}
\usage{
chisq.prob(u, param)
}
\arguments{
\item{u}{vector with distances}

\item{param}{vector with distribution parameters}
}
\description{
Calculate probabilities for distance values using Chi-square distribution
}
