% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcoeffs.R
\name{confint.regcoeffs}
\alias{confint.regcoeffs}
\title{Confidence intervals for regression coefficients}
\usage{
\method{confint}{regcoeffs}(object, parm = NULL, level = 0.95, ncomp = 1, ny = 1, ...)
}
\arguments{
\item{object}{regression coefficients object (class \code{regcoeffs})}

\item{parm}{not used, needed for compatiility with general method}

\item{level}{confidence level}

\item{ncomp}{number of components (one value)}

\item{ny}{index of response variable (one value)}

\item{...}{other arguments}
}
\description{
returns matrix with confidence intervals for regression coeffocoents
for given response number and number of components.
}
