% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraintNorm}
\alias{constraintNorm}
\title{Method for normalization constraint}
\usage{
constraintNorm(x, d, type = "length")
}
\arguments{
\item{x}{data matrix (spectra or contributions)}

\item{d}{matrix with the original spectral values}

\item{type}{type of normalization ("area", "length" or "sum")}
}
\description{
Normalize rows of matrix to unit length or area
}
