% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{employ.prep}
\alias{employ.prep}
\title{Applies a list with preprocessing methods to a dataset}
\usage{
employ.prep(obj, x, ...)
}
\arguments{
\item{obj}{list with preprocssing methods (created using \code{prep} function).}

\item{x}{matrix with dataset}

\item{...}{other arguments}
}
\description{
Applies a list with preprocessing methods to a dataset
}
