% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{employ}
\alias{employ}
\title{Generic function to apply a method (e.g. constraint)}
\usage{
employ(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{other parameters}
}
\description{
Generic function to apply a method (e.g. constraint)
}
