% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{employ}
\alias{employ}
\title{Generic function to apply a method (e.g. constraint, preprocessing, etc)}
\usage{
employ(obj, x, ...)
}
\arguments{
\item{obj}{object}

\item{x}{data in question}

\item{...}{other parameters}
}
\description{
Generic function to apply a method (e.g. constraint, preprocessing, etc)
}
