% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{employ}
\alias{employ}
\title{Generic function to apply a method (e.g. constraint)}
\usage{
employ(obj, x, d)
}
\arguments{
\item{obj}{constraint object}

\item{x}{data in question (e.g. resolved spectra)}

\item{d}{matrix with original data}
}
\description{
Generic function to apply a method (e.g. constraint)
}
