% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{getCalibrationData.pca}
\alias{getCalibrationData.pca}
\title{Get calibration data}
\usage{
\method{getCalibrationData}{pca}(obj, ...)
}
\arguments{
\item{obj}{PCA model (object of class \code{pca})}

\item{...}{other parameters}
}
\description{
Get data, used for calibration of the PCA model
}

