\name{getCalibrationData}
\alias{getCalibrationData}
\title{Calibration data}
\usage{
getCalibrationData(obj, ...)
}
\arguments{
  \item{obj}{a model object}

  \item{...}{other arguments}
}
\description{
Calibration data
}
\details{
Generic function getting calibration data from a linear
decomposition model (e.g. PCA)
}

