\name{getCalibrationData.simcam}
\alias{getCalibrationData.simcam}
\title{Get calibration data}
\usage{
\method{getCalibrationData}{simcam}(obj, ...)
}
\arguments{
  \item{obj}{SIMCAM model (object of class \code{simcam})}

  \item{...}{other arguments}
}
\description{
Get data, used for calibration of the SIMCAM model.
}
\details{
See examples in help for \code{\link{simcam}} function.
}

