% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{getProbabilities}
\alias{getProbabilities}
\title{Get class belonging probability}
\usage{
getProbabilities(obj, ...)
}
\arguments{
\item{obj}{an object with classification results (e.g. SIMCA)}

\item{...}{other parameters}
}
\description{
Compute class belonging probabilities for classification results.
}
