% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simca.R
\name{getProbabilities.simca}
\alias{getProbabilities.simca}
\title{Probability of class belonging for PCA/SIMCA results}
\usage{
\method{getProbabilities}{simca}(obj, ncomp, Q, T2, ...)
}
\arguments{
\item{obj}{object with SIMCA model}

\item{ncomp}{number of components to compute the probabilities for.}

\item{Q}{vector with Q values for selected component}

\item{T2}{vector with T2 values for selected component}

\item{...}{other arguments}
}
\description{
Computes probability of class belonging for each object based on Q and T2 residuls
}
