% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simca.R
\name{getProbabilities.simca}
\alias{getProbabilities.simca}
\title{Probabilities of class belonging for PCA/SIMCA results}
\usage{
\method{getProbabilities}{simca}(obj, ncomp, q, h, ...)
}
\arguments{
\item{obj}{object with PCA model}

\item{ncomp}{number of components to compute the probability for}

\item{q}{vector with squared orthogonal distances for given number of components}

\item{h}{vector with score distances for given number of components}

\item{...}{other parameters}
}
\description{
Probabilities of class belonging for PCA/SIMCA results
}
\details{
Computes p-value for every object being from the same populaion as calibration set
based on its orthogonal and score distances.
}
