% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrpure.R
\name{getPureVariables}
\alias{getPureVariables}
\title{Identifies pure variables}
\usage{
getPureVariables(D, ncomp, purevars, offset, use.deriv, savgol)
}
\arguments{
\item{D}{matrix with the spectra}

\item{ncomp}{number of pure components}

\item{purevars}{user provided values gor pure variables (no calculation will be run in this case)}

\item{offset}{offset (between 0 and 1) for calculation of parameter alpha}

\item{use.deriv}{a number which tells how to use derivative.}

\item{savgol}{list with parameters for Savitzky-Golay preprocessing (if \code{use.deriv} is not 0).}
}
\value{
The function returns a list with with following fields:
\item{ncomp }{number of pure components.}
\item{purvars }{vector with indices for pure variables.}
\item{purityspec }{matrix with purity values for each resolved components.}
\item{purity }{vector with purity values for resolved components.}
}
\description{
The method identifies indices of pure variables using the SIMPLISMA
algorithm.
}
