% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{hotelling.prob}
\alias{hotelling.prob}
\title{Calculate probabilities for distance values and given parameters using Hotelling T2 distribution}
\usage{
hotelling.prob(u, ncomp, nobj)
}
\arguments{
\item{u}{vector with distances}

\item{ncomp}{number of components}

\item{nobj}{number of objects in calibration set}
}
\description{
Calculate probabilities for distance values and given parameters using Hotelling T2 distribution
}
