% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ipls.R
\name{ipls.backward}
\alias{ipls.backward}
\title{Runs the backward iPLS algorithm}
\usage{
ipls.backward(x, y, obj)
}
\arguments{
\item{x}{a matrix with predictor values}

\item{y}{a vector with response values}

\item{obj}{object with initial settings for iPLS algorithm}
}
\description{
Runs the backward iPLS algorithm
}

