% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ipls.R
\name{ipls.forward}
\alias{ipls.forward}
\title{Runs the forward iPLS algorithm}
\usage{
ipls.forward(x, y, obj)
}
\arguments{
\item{x}{a matrix with predictor values}

\item{y}{a vector with response values}

\item{obj}{object with initial settings for iPLS algorithm}
}
\description{
Runs the forward iPLS algorithm
}

