% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.getDistances}
\alias{ldecomp.getDistances}
\title{Compute score and residual distances}
\usage{
ldecomp.getDistances(scores, loadings, residuals, eigenvals)
}
\arguments{
\item{scores}{matrix with scores (T).}

\item{loadings}{matrix with loadings (P).}

\item{residuals}{matrix with residuals (E).}

\item{eigenvals}{vector with eigenvalues for the components}
}
\value{
Returns a list with Q, T2 and tnorm values for each component.
}
\description{
Compute orthogonal Euclidean distance from object to PC space (Q, q) and Mahalanobis
squared distance between projection of the object to the space and its origin (T2, h).
}
\details{
The distances are calculated for every 1:n components, where n goes from 1 to ncomp
(number of columns in scores and loadings).
}
