% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.getLimParams}
\alias{ldecomp.getLimParams}
\title{Compute parameters for critical limits based on calibration results}
\usage{
ldecomp.getLimParams(Q, T2)
}
\arguments{
\item{Q}{matrix with Q distances}

\item{T2}{matrix with T2 distances}
}
\description{
Compute parameters for critical limits based on calibration results
}
