% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.getLimParams}
\alias{ldecomp.getLimParams}
\title{Compute parameters for critical limits based on calibration results}
\usage{
ldecomp.getLimParams(U)
}
\arguments{
\item{U}{matrix with residual distances}
}
\description{
Compute parameters for critical limits based on calibration results
}
