% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.getVariances}
\alias{ldecomp.getVariances}
\title{Explained variance for linear decomposition}
\usage{
ldecomp.getVariances(Q2, totvar)
}
\arguments{
\item{Q2}{Q2 values (squared residuals distance from object to component space).}

\item{totvar}{Total variance of the original data (after preprocessing).}
}
\value{
Returns a list with two vectors.
}
\description{
Computes explained variance and cumulative explained variance for a data decomposition X = TP' + E.
}

