% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{ldecomp.getVariances}
\alias{ldecomp.getVariances}
\title{Compute explained variance}
\usage{
ldecomp.getVariances(scores, loadings, residuals, Q)
}
\arguments{
\item{scores}{matrix with scores (T).}

\item{loadings}{matrix with loadings (P).}

\item{residuals}{matrix with residuals (E).}

\item{Q}{matrix with squared orthogonal distances.}
}
\value{
Returns a list with two vectors.
}
\description{
Computes explained variance and cumulative explained variance for data decomposition.
}
