% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrals.R
\name{mcrals.ols}
\alias{mcrals.ols}
\title{Ordinary least squares}
\usage{
mcrals.ols(D, A)
}
\arguments{
\item{D}{a matrix}

\item{A}{a matrix}
}
\description{
Ordinary least squares
}
\details{
Computes OLS solution for D = AB' (or D' = AB'), where D, A are known
}
