% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.getYTickLabels}
\alias{mdaplot.getYTickLabels}
\title{Prepare yticklabels for plot}
\usage{
mdaplot.getYTickLabels(yticklabels, yticks, excluded_rows)
}
\arguments{
\item{yticklabels}{yticklables provided by user (if any)}

\item{yticks}{yticks (provided or computed)}

\item{excluded_rows}{rows excluded from plot data (if any)}
}
\description{
Prepare yticklabels for plot
}
