% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot.plotAxes}
\alias{mdaplot.plotAxes}
\title{Create axes plane}
\usage{
mdaplot.plotAxes(
  xticklabels = NULL,
  yticklabels = NULL,
  xlim = xlim,
  ylim = ylim,
  xticks = NULL,
  yticks = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  xlas = 0,
  ylas = 0,
  show.grid = TRUE,
  grid.lwd = 0.5,
  grid.col = "lightgray"
)
}
\arguments{
\item{xticklabels}{labels for x ticks}

\item{yticklabels}{labels for y ticks}

\item{xlim}{vector with limits for x axis}

\item{ylim}{vector with limits for y axis}

\item{xticks}{values for x ticks}

\item{yticks}{values for y ticks}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{xlas}{orientation of xticklabels}

\item{ylas}{orientation of yticklabels}

\item{show.grid}{logical, show or not axes grid}

\item{grid.lwd}{line thinckness (width) for the grid}

\item{grid.col}{line color for the grid}
}
\description{
Creates an empty axes plane for given parameters
}
