% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.plotAxes}
\alias{mdaplot.plotAxes}
\title{Create axes plane}
\usage{
mdaplot.plotAxes(xticks, xticklabels, yticks, yticklabels, lim, main, xlab,
  ylab)
}
\arguments{
\item{xticks}{tick values for x axis}

\item{xticklabels}{labels for x ticks}

\item{yticks}{tick values for y axis}

\item{yticklabels}{labels for y ticks}

\item{lim}{vector with limits for x and y axis}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}
}
\description{
Creates an empty axes plane for given parameters
}

