% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.plotAxes}
\alias{mdaplot.plotAxes}
\title{Create axes plane}
\usage{
mdaplot.plotAxes(xticklabels = NULL, yticklabels = NULL, xticks = NULL,
  yticks = NULL, lim = NULL, main = NULL, xlab = NULL, ylab = NULL,
  xlas = 0, ylas = 0)
}
\arguments{
\item{xticklabels}{labels for x ticks}

\item{yticklabels}{labels for y ticks}

\item{xticks}{values for x ticks}

\item{yticks}{values for y ticks}

\item{lim}{vector with limits for x and y axis}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{xlas}{orientation of xticklabels}

\item{ylas}{orientation of yticklabels}
}
\description{
Creates an empty axes plane for given parameters
}
