% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mdaplots.R
\name{mdaplot.showColorbar}
\alias{mdaplot.showColorbar}
\title{Plot colorbar}
\usage{
mdaplot.showColorbar(cgroup, colmap = "default")
}
\arguments{
\item{cgroup}{a vector with values used to make color grouping of the elements}

\item{colmap}{a colormap to be used for color generation}
}
\description{
Shows a colorbar if plot has color grouping of elements (points or lines).
}

