% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdaplotg.R
\name{mdaplotg.getLegend}
\alias{mdaplotg.getLegend}
\title{Create and return vector with legend values}
\usage{
mdaplotg.getLegend(ps, data.names, legend = NULL)
}
\arguments{
\item{ps}{list with plot series}

\item{data.names}{names of the data sets}

\item{legend}{legend values provided by user}
}
\value{
vector of text values for the legend
}
\description{
Create and return vector with legend values
}
