% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.cal}
\alias{pca.cal}
\title{PCA model calibration}
\usage{
pca.cal(x, ncomp, center, scale, method, rand = NULL)
}
\arguments{
\item{x}{matrix with data values}

\item{ncomp}{number of principal components to calculate}

\item{center}{logical, do mean centering or not}

\item{scale}{logical, do standardization or not}

\item{method}{algorithm for compiting PC space (only 'svd' and 'nipals' are supported so far)}

\item{rand}{vector with parameters for randomized PCA methods (if NULL, conventional PCA is used instead)}
}
\value{
an object with calibrated PCA model
}
\description{
Calibrates (builds) a PCA model for given data and parameters
}
