% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.run}
\alias{pca.run}
\title{Runs one of the selected PCA methods}
\usage{
pca.run(x, ncomp, method, rand = NULL)
}
\arguments{
\item{x}{data matrix}

\item{ncomp}{number of components}

\item{method}{name of PCA methods ('svd', 'nipals')}

\item{rand}{parameters for randomized algorithm (if not NULL)}
}
\description{
Runs one of the selected PCA methods
}
