% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classres.R
\name{plot.classres}
\alias{plot.classres}
\title{Plot function for classification results}
\usage{
\method{plot}{classres}(x, ...)
}
\arguments{
\item{x}{classification results (object of class \code{plsdares}, \code{simcamres}, etc.).}

\item{...}{other arguments for \code{plotPredictions()} method.}
}
\description{
Generic plot function for classification results.
Alias for \code{\link{plotPredictions.classres}}.
}
