% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{plot.ddsimca}
\alias{plot.ddsimca}
\title{Model overview plot for SIMCA}
\usage{
\method{plot}{ddsimca}(x, comp = c(1, 2), ncomp = x$ncomp.selected, ...)
}
\arguments{
\item{x}{a SIMCA model (object of class \code{simca})}

\item{comp}{which components to show on scores and loadings plot}

\item{ncomp}{how many components to use for residuals plot}

\item{...}{other arguments}
}
\description{
Shows a set of plots for SIMCA model.
}
\details{
See examples in help for \code{\link{ddsimca}} function.
}
