% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plot.ddsimcares}
\alias{plot.ddsimcares}
\title{Plot method for DD-SIMCA results.}
\usage{
\method{plot}{ddsimcares}(obj, ...)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares})}
}
\description{
Plot method for DD-SIMCA results.
}
