% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pls.R
\name{plot.pls}
\alias{plot.pls}
\title{Model overview plot for PLS}
\usage{
\method{plot}{pls}(x, ncomp = NULL, ny = 1, show.legend = T,
  show.labels = F, ...)
}
\arguments{
\item{x}{a PLS model (object of class \code{pls})}

\item{ncomp}{how many components to use (if NULL - user selected optimal value will be used)}

\item{ny}{which y variable to show the summary for (if NULL, will be shown for all)}

\item{show.legend}{logical, show or not a legend on the plot}

\item{show.labels}{logical, show or not labels for the plot objects}

\item{...}{other arguments}
}
\description{
Shows a set of plots (x residuals, regression coefficients, RMSE and predictions) for PLS model.
}
\details{
See examples in help for \code{\link{pls}} function.
}

