% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcoeffs.R
\name{plot.regcoeffs}
\alias{plot.regcoeffs}
\title{Regression coefficients plot}
\usage{
\method{plot}{regcoeffs}(x, ncomp = 1, ny = 1, type = NULL, col = NULL,
  main = NULL, ylab = NULL, show.line = T, show.ci = T, ...)
}
\arguments{
\item{x}{regression coefficients object (class \code{regcoeffs})}

\item{ncomp}{number of components to return the coefficients for}

\item{ny}{number of response variable to return the coefficients for}

\item{type}{type of the plot}

\item{col}{vector with colors for the plot (vector or one value)}

\item{main}{main plot title}

\item{ylab}{label for y axis}

\item{show.line}{logical, show or not line for 0 value}

\item{show.ci}{logical, show or not confidence intervals if they are available}

\item{...}{other arguments}
}
\description{
Shows plot with regression coefficient values for every predictor variable (x)
}

