% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{plot.regres}
\alias{plot.regres}
\title{Plot method for regression results}
\usage{
\method{plot}{regres}(x, ...)
}
\arguments{
\item{x}{regression results (object of class \code{regres})}

\item{...}{other arguments}
}
\description{
Plot method for regression results
}
\details{
This is a shortcut for \code{\link{plotPredictions.regres}}
}
