% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotAcceptance.ddsimcares}
\alias{plotAcceptance.ddsimcares}
\title{Acceptance plot for DD-SIMCA results object.}
\usage{
\method{plotAcceptance}{ddsimcares}(
  obj,
  ncomp = obj$ncomp.selected,
  limType = "classic",
  show = "auto",
  log = FALSE,
  show.legend = TRUE,
  pch = 1,
  colors = c(regular = "#2679B2", extreme = "#F2B825", outlier = "#D22C2F", alien =
    "#2679B2", external = "#D22C2F", unknown = "#3c6784", excluded = "#00ff00"),
  lim.lty = c(2, 3),
  lim.col = c("darkgray", "darkgray"),
  lim.lwd = c(1, 1),
  ylim = NULL,
  xlim = NULL,
  show.excluded = FALSE,
  ...
)
}
\arguments{
\item{obj}{classification results (object of class \code{ddsimcamres}).}

\item{ncomp}{number of components to show the plot for.}

\item{limType}{method for estimation of critical limits ('classic' or 'robust').}

\item{show}{name of subset to show (\code{"members"}, \code{"strangers"}, \code{"all"}, or \code{"auto"}).}

\item{log}{logical, apply log transformation or not.}

\item{show.legend}{logica, show or not colorbar legend on the plot.}

\item{pch}{marker type}

\item{colors}{named vector with colors for every role}

\item{lim.lty}{vector with two values - types of the lines showing critical limits (extreme, outlier).}

\item{lim.col}{vector with two values - colors of the lines showing critical limits (extreme, outlier).}

\item{lim.lwd}{vector with two values - thickness of the lines showing critical limits (extreme, outlier).}

\item{ylim}{limits for y-axis, if NULL will be estimated automatically.}

\item{xlim}{limits for x-axis, if NULL will be estimated automatically.}

\item{show.excluded}{logical, show or not the excluded rows (objects).}

\item{...}{other parameters compatible with \code{\link{mdaplot}} method.}
}
\description{
Shows a plot with h/h0 and q/q0 distance values for given number of components and type of critical
limits.
}
