% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotAliens.ddsimcares}
\alias{plotAliens.ddsimcares}
\title{Aliens plot for DD-SIMCA results.}
\usage{
\method{plotAliens}{ddsimcares}(
  obj,
  ncomp = obj$ncomp.selected,
  limType = "classic",
  main = sprintf("Aliens (A = \%d)", ncomp),
  xlab = "Number of aliens (expected)",
  ylab = "Number of aliens (observed)",
  pch = 1,
  col = "#2679b2",
  ellipse.col = "#eeeeee",
  show.plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{ncomp}{number of components in the model to show the plot for.}

\item{limType}{limit type to show the plot for ('classic' or 'robust')}
}
\description{
Aliens plot for DD-SIMCA results.
}
