% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{plotCumVariance.ldecomp}
\alias{plotCumVariance.ldecomp}
\title{Cumulative explained variance plot}
\usage{
\method{plotCumVariance}{ldecomp}(obj, type = "b", labels = "values", show.plot = TRUE, ...)
}
\arguments{
\item{obj}{object of \code{ldecomp} class.}

\item{type}{type of the plot}

\item{labels}{what to show as labels for plot objects}

\item{show.plot}{logical, shall plot be created or just plot series object is needed}

\item{...}{most of graphical parameters from \code{\link{mdaplot}} function can be used.}
}
\description{
Shows a plot with cumulative explained variance vs. number of components.
}
