% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotCumVariance.pca}
\alias{plotCumVariance.pca}
\title{Cumulative explained variance plot for PCA model}
\usage{
\method{plotCumVariance}{pca}(obj, legend.position = "bottomright", ...)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{legend.position}{position of the legend}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows a plot with cumulative explained variance for components.
}
\details{
See examples in help for \code{\link{pca}} function.
}
