% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotCumVariance}
\alias{plotCumVariance}
\title{Variance plot}
\usage{
plotCumVariance(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Variance plot
}
\details{
Generic function for plotting explained variance for data decomposition
}

