% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotDiscriminationPower}
\alias{plotDiscriminationPower}
\title{Discrimination power plot}
\usage{
plotDiscriminationPower(obj, ...)
}
\arguments{
\item{obj}{a model object}

\item{...}{other arguments}
}
\description{
Discrimination power plot
}
\details{
Generic function for plotting discrimination power values for classification model
}

