% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotDistances}
\alias{plotDistances}
\title{Distance plot for model and results (generic function)}
\usage{
plotDistances(obj, ...)
}
\arguments{
\item{obj}{model or result object, e.g. \code{ddsimcares}}

\item{...}{other parameters relevant for the plot}
}
\description{
Distance plot for model and results (generic function)
}
