% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotDistDoF}
\alias{plotDistDoF}
\title{Degrees of freedom plot for both distances}
\usage{
plotDistDoF(
  obj,
  type = "b",
  main = "Degrees of freedom",
  ylab = c("Nh", "Nq"),
  labels = "values",
  ...
)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{type}{type of the plot ("b", "l", "h")}

\item{main}{main title for the plot}

\item{ylab}{label for y axis}

\item{labels}{what to show as data points labels}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows a plot with degrees of freedom computed for score and orthogonal distances at given number
of components using data driven approach ("ddmoments" or "ddrobust").
}
\details{
Work only if parameter \code{lim.type} equal to "ddmoments" or "ddrobust".
}
