% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotExtreme}
\alias{plotExtreme}
\title{Shows extreme plot for SIMCA model}
\usage{
plotExtreme(obj, ...)
}
\arguments{
\item{obj}{a SIMCA model}

\item{...}{other parameters}
}
\description{
Generic function for creating extreme plot for SIMCA model
}
