% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{plotFoMs.ddsimcares}
\alias{plotFoMs.ddsimcares}
\title{Figures of merit plot (multiple FoMs).}
\usage{
\method{plotFoMs}{ddsimcares}(
  obj,
  foms = NULL,
  limType = "classic",
  type = "b",
  pch = 1,
  ylim = c(0, 1.1),
  legend.position = "top",
  main = "Figures of merit",
  ...
)
}
\arguments{
\item{obj}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{foms}{vector with FoM names to show on the plot (any combination of 'sens', 'spec', 'sel', 'eff', 'acc').}

\item{limType}{limit type to show the plot for ('classic' or 'robust').}

\item{type}{type of plot (can be 'b', 'l', or 'h').}

\item{pch}{marker number or symbol for the plot points (if 'b' type is selected), can be single or individual for each FoM.}

\item{ylim}{limits for y-axis.}

\item{legend.position}{position of legend on the plot.}

\item{main}{main title.}

\item{...}{any parameters for \code{\link{mdaplotg}} method can be provided.

Shows several figures of merit values vs. number of components as a group plot (see
details in \code{\link{mdaplotg}} method desciption.}

\item{show.plot}{logical, if FALSE returns only values and does not show the plot.}
}
\description{
Figures of merit plot (multiple FoMs).
}
