% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/randtest.R
\name{plotHist.randtest}
\alias{plotHist.randtest}
\title{Histogram plot for randomization test results}
\usage{
\method{plotHist}{randtest}(obj, comp = NULL, main = NULL,
  xlab = "Test statistic", ylab = "Frequency", ...)
}
\arguments{
\item{obj}{results of randomization test (object of class `randtest`)}

\item{comp}{number of component to make the plot for}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{...}{other optional arguments}
}
\description{
Makes a histogram for statistic values distribution for particular component, also
show critical value as a vertical line.
}
\details{
See examples in help for \code{\link{randtest}} function.
}

