% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{plotLines}
\alias{plotLines}
\title{Show plot series as set of lines}
\usage{
plotLines(
  ps,
  col = ps$col,
  lty = 1,
  lwd = 1,
  cex = 1,
  col.excluded = "darkgray",
  show.excluded = FALSE,
  ...
)
}
\arguments{
\item{ps}{`plotseries` object}

\item{col}{a color for markers or lines (same as \code{plot} parameter).}

\item{lty}{line type}

\item{lwd}{line width}

\item{cex}{scale factor for the marker}

\item{col.excluded}{color for the excluded lines.}

\item{show.excluded}{logical, show or not the excluded data points}

\item{...}{other arguments for function `lines()`.}
}
\description{
Show plot series as set of lines
}
