% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotLoadings.pca}
\alias{plotLoadings.pca}
\title{Loadings plot for PCA model}
\usage{
\method{plotLoadings}{pca}(
  obj,
  comp = if (obj$ncomp > 1) c(1, 2) else 1,
  type = (if (length(comp == 2)) "p" else "l"),
  show.legend = TRUE,
  show.axes = TRUE,
  ...
)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{comp}{a value or vector with several values - number of components to show the plot for}

\item{type}{type of the plot ('b', 'l', 'h')}

\item{show.legend}{logical, show or not a legend on the plot}

\item{show.axes}{logical, show or not a axes lines crossing origin (0,0)}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows a loadings plot for selected components.
}
\details{
See examples in help for \code{\link{pca}} function.
}
