% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotLoadings}
\alias{plotLoadings}
\title{Loadings plot}
\usage{
plotLoadings(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Loadings plot
}
\details{
Generic function for plotting loadings values for data decomposition
}

