% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classres.R
\name{plotMisclassified.classres}
\alias{plotMisclassified.classres}
\title{Misclassified ratio plot for classification results}
\usage{
\method{plotMisclassified}{classres}(obj, ...)
}
\arguments{
\item{obj}{classification results (object of class \code{plsdares}, \code{simcamres}, etc.).}

\item{...}{other parameters for \code{\link{plotPerformance.classres}}}
}
\description{
Makes a plot with ms ratio values vs. model complexity (e.g. number of components) for
classification results.
}
\details{
See examples in description of \code{\link{plsdares}}, \code{\link{simcamres}}, etc.
}
