% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotMisclassified}
\alias{plotMisclassified}
\title{Misclassification ratio plot}
\usage{
plotMisclassified(obj, ...)
}
\arguments{
\item{obj}{a model or a result object}

\item{...}{other arguments}
}
\description{
Misclassification ratio plot
}
\details{
Generic function for plotting missclassification values for classification model or results
}

