% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotModelDistance}
\alias{plotModelDistance}
\title{Model distance plot}
\usage{
plotModelDistance(obj, ...)
}
\arguments{
\item{obj}{a model object}

\item{...}{other arguments}
}
\description{
Model distance plot
}
\details{
Generic function for plotting distance from object to a multivariate model
}
