\name{plotModellingPower}
\alias{plotModellingPower}
\title{Modelling power plot}
\usage{
plotModellingPower(obj, ...)
}
\arguments{
  \item{obj}{a model object}

  \item{...}{other arguments}
}
\description{
Modelling power plot
}
\details{
Generic function for plotting modelling power values for
classification model
}

