\name{plotModellingPower.simcam}
\alias{plotModellingPower.simcam}
\title{Modelling power plot for SIMCAM model}
\usage{
\method{plotModellingPower}{simcam}(obj, nc = 1, main = NULL, ...)
}
\arguments{
  \item{obj}{a SIMCAM model (object of class
  \code{simcam})}

  \item{nc}{which classe (SIMCA model) to show the plot
  for}

  \item{main}{main plot title}

  \item{...}{other plot parameters (see \code{mdaplotg} for
  details)}
}
\description{
Shows a plot with modelling power values for each predictor
of selected SIMCA model
}
\details{
See examples in help for \code{\link{simcam}} function.
}

