% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotPerformance}
\alias{plotPerformance}
\title{Classification performance plot}
\usage{
plotPerformance(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Classification performance plot
}
\details{
Generic function for plotting classification performance for model or results
}

