% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classmodel.R
\name{plotPredictions.classmodel}
\alias{plotPredictions.classmodel}
\title{Predictions plot for classification model}
\usage{
\method{plotPredictions}{classmodel}(obj, res = NULL, nc = NULL,
  ncomp = NULL, main = NULL, ...)
}
\arguments{
\item{obj}{a classification model (object of class \code{simca}, \code{plsda}, etc.). if \code{NULL} value
is specified, the result will be selected automatically by checking the nearest available from
test, cv and calibration results.}

\item{res}{which result to make the plot for (\code{'calres'}, \code{'cvres'} or \code{'testres'}).}

\item{nc}{if there are several classes, which class to make the plot for (NULL - for all).}

\item{ncomp}{what number of components to make the plot for (NULL - for selected in the model).}

\item{main}{main title for the plot}

\item{...}{most of the graphical parameters from \code{\link{mdaplotg}} function can be used.}
}
\description{
Makes a plot with class predictions for a classification model.
}
\details{
See examples in description of \code{\link{plsda}}, \code{\link{simca}} or \code{\link{simcam}}.
}

