% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classres.R
\name{plotPredictions.classres}
\alias{plotPredictions.classres}
\title{Prediction plot for classification results}
\usage{
\method{plotPredictions}{classres}(obj, nc = NULL, ncomp = NULL,
  type = "p", main = NULL, ylab = "", ...)
}
\arguments{
\item{obj}{classification results (object of class \code{plsdares}, \code{simcamres}, etc.).}

\item{nc}{if there are several classes, which class to make the plot for (NULL - summary for all classes).}

\item{ncomp}{which number of components to make the plot for (one value, if NULL - model selected number will 
be used).
This parameter shal not be used for multiclass models or results as predictions in this case are 
 only
for optimal number of components}

\item{type}{type of the plot}

\item{main}{main title for the plot}

\item{ylab}{label for y axis}

\item{...}{most of the graphical parameters from \code{\link{mdaplotg}} or \code{\link{mdaplot}} function 
can be used.}
}
\description{
Makes a plot with predicted class values for classification results.
}
\details{
See examples in description of \code{\link{plsdares}}, \code{\link{simcamres}}, etc.
}
