% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsres.R
\name{plotPredictions.plsres}
\alias{plotPredictions.plsres}
\title{Predictions plot for PLS results}
\usage{
\method{plotPredictions}{plsres}(obj, ny = 1, ncomp = NULL, main = NULL,
  ...)
}
\arguments{
\item{obj}{PLS results (object of class \code{plsres})}

\item{ny}{number of response variable to make the plot for (if y is multivariate)}

\item{ncomp}{how many components to use (if NULL - user selected optimal value will be used)}

\item{main}{main plot title}

\item{...}{other plot parameters (see \code{mdaplot} for details)}
}
\description{
Shows plot with predicted vs. reference (measured) y values for selected components.
}
\details{
See examples in help for \code{\link{plsres}} function.
}
\seealso{
\code{\link{plotPredictions.regres}} - prediction plot for regression results.
}
