% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcam.R
\name{plotPredictions.simcam}
\alias{plotPredictions.simcam}
\title{Predictions plot for SIMCAM model}
\usage{
\method{plotPredictions}{simcam}(
  obj,
  nc = seq_len(obj$nclasses),
  main = "SIMCAM Predictions (cal)",
  ...
)
}
\arguments{
\item{obj}{a SIMCAM model (object of class \code{simcam})}

\item{nc}{vector with class numbers to make the plot for.}

\item{main}{plot title.}

\item{...}{most of the graphical parameters from \code{\link{mdaplotg}} function can be used.}
}
\description{
Makes a plot with class predictions for calibration dataset.
}
\details{
See examples in description of \code{\link{plsda}}, \code{\link{simca}} or \code{\link{simcam}}.
}
