% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotProbabilities}
\alias{plotProbabilities}
\title{Plot for class belonging probability}
\usage{
plotProbabilities(obj, ...)
}
\arguments{
\item{obj}{an object with classification results (e.g. SIMCA)}

\item{...}{other parameters}
}
\description{
Makes a plot with class belonging probabilities for each object of the classification results.
Works only with classification methods, which compute this probability (e.g. SIMCA).
}
