% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{plotRegcoeffs.pls}
\alias{plotRegcoeffs.pls}
\title{Regression coefficient plot for PLS}
\usage{
\method{plotRegcoeffs}{pls}(obj, ncomp = NULL, ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{ncomp}{how many components to use (if NULL - user selected optimal value will be used)}

\item{...}{other plot parameters (see \code{mdaplotg} and \code{plot.regcoeffs} for details)}
}
\description{
Shows plot with regression coefficient values for selected components.
}
\details{
See examples in help for \code{\link{pls}} function.
}

