% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotRegcoeffs}
\alias{plotRegcoeffs}
\title{Regression coefficients plot}
\usage{
plotRegcoeffs(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Regression coefficients plot
}
\details{
Generic function for plotting regression coefficients values for a regression model
}
