% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodel.R
\name{plotRegcoeffs.regmodel}
\alias{plotRegcoeffs.regmodel}
\title{Regression coefficient plot for regression model}
\usage{
\method{plotRegcoeffs}{regmodel}(obj, ncomp = obj$ncomp.selected, ...)
}
\arguments{
\item{obj}{a regression model (object of class \code{regmodel})}

\item{ncomp}{number of components to show the plot for}

\item{...}{other plot parameters (see \code{link{plot.regcoeffs}} for details)}
}
\description{
Shows plot with regression coefficient values. Is a proxy for \code{link{plot.regcoeffs}} method.
}
