% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{plotResiduals.ldecomp}
\alias{plotResiduals.ldecomp}
\title{Residuals plot for linear decomposition}
\usage{
\method{plotResiduals}{ldecomp}(obj, ncomp = NULL, main = NULL,
  xlab = "T2", ylab = "Q2", show.labels = F, show.limits = T, ...)
}
\arguments{
\item{obj}{object of \code{ldecomp} class.}

\item{ncomp}{what number of components to show the plot for (if NULL, model selected value will be used).}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{show.labels}{logical, show or not labels for the plot objects}

\item{show.limits}{logical, show or not lines for statistical limits of the residuals}

\item{...}{most of graphical parameters from \code{\link{mdaplot}} function can be used.}
}
\description{
Shows a plot with T2 vs Q2 values for data objects.
}

