% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotResiduals.pca}
\alias{plotResiduals.pca}
\title{Residuals distance plot for PCA model (legacy, use \code{\link{plotResiduals}} instead).}
\usage{
\method{plotResiduals}{pca}(obj, ...)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{...}{any parameter from \code{\link{plotDistances.pca}}}
}
\description{
Residuals distance plot for PCA model (legacy, use \code{\link{plotResiduals}} instead).
}
