% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcares.R
\name{plotResiduals.pcares}
\alias{plotResiduals.pcares}
\title{Residual distance plot}
\usage{
\method{plotResiduals}{pcares}(
  obj,
  ncomp = obj$ncomp.selected,
  norm = TRUE,
  log = FALSE,
  show.plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{object of \code{ldecomp} class.}

\item{ncomp}{number of components to show the plot for (if NULL, selected by model value will be used).}

\item{norm}{logical, if TRUE disance values be normalized (u/u0)}

\item{log}{logical, if TRUE, then log(1 + u) transformation is applied}

\item{show.plot}{logical, show plot or just return plot data}

\item{...}{most of graphical parameters from \code{\link{mdaplot}} function can be used.}
}
\description{
Shows a plot with orthogonal (Q, q) vs score (T2, h) distances for data objects. By default the
distance values are normalize using corresponding means (q/q0 and h/h0).
}
