% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simcam.R
\name{plotResiduals.simcam}
\alias{plotResiduals.simcam}
\title{Residuals plot for SIMCAM model}
\usage{
\method{plotResiduals}{simcam}(obj, ...)
}
\arguments{
\item{obj}{a SIMCAM model (object of class \code{simcam})}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows a plot with residuals for SIMCAM model
}
\details{
See examples in help for \code{\link{simcam}} function.
}

