% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcamres.R
\name{plotResiduals.simcamres}
\alias{plotResiduals.simcamres}
\title{Residuals plot for SIMCAM results}
\usage{
\method{plotResiduals}{simcamres}(obj, nc = 1, main = NULL, ...)
}
\arguments{
\item{obj}{SIMCAM results (object of class \code{simcamres})}

\item{nc}{which class (SIMCA model) to show the plot for}

\item{main}{main plot title}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows a plot with Q vs. T2 residuals for SIMCAM results
}
\details{
See examples in help for \code{\link{simcamres}} function.
}

