\name{plotRMSE.plsres}
\alias{plotRMSE.plsres}
\title{RMSE plot for PLS results}
\usage{
\method{plotRMSE}{plsres}(obj, xlab = "Components", ...)
}
\arguments{
  \item{obj}{PLS results (object of class \code{plsres})}

  \item{xlab}{label for x axis}

  \item{...}{other plot parameters (see \code{mdaplotg} for
  details)}
}
\description{
Shows plot with root mean squared error values vs. number
of components for PLS results.
}
\details{
See examples in help for \code{\link{plsres}} function.
}

