% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotRMSE}
\alias{plotRMSE}
\title{RMSE plot}
\usage{
plotRMSE(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
RMSE plot
}
\details{
Generic function for plotting RMSE values vs. complexity of a regression model
}
