% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{plotRMSE.regres}
\alias{plotRMSE.regres}
\title{RMSE plot for regression results}
\usage{
\method{plotRMSE}{regres}(obj, ny = 1, type = "b", labels = "values", ...)
}
\arguments{
\item{obj}{regression results (object of class \code{regres})}

\item{ny}{number of predictor to show the plot for (if y is multivariate)}

\item{type}{type of the plot}

\item{labels}{what to show as labels for plot objects.}

\item{...}{other plot parameters (see \code{mdaplot} for details)}
}
\description{
Shows plot with RMSE values vs. model complexity (e.g. number of components).
}

