% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{plotScores.ldecomp}
\alias{plotScores.ldecomp}
\title{Scores plot}
\usage{
\method{plotScores}{ldecomp}(
  obj,
  comp = c(1, 2),
  type = "p",
  show.axes = TRUE,
  show.plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{object of \code{ldecomp} class.}

\item{comp}{which components to show the plot for (can be one value or vector with two values).}

\item{type}{type of the plot}

\item{show.axes}{logical, show or not a axes lines crossing origin (0,0)}

\item{show.plot}{logical, shall plot be created or just plot series object is needed}

\item{...}{most of graphical parameters from \code{\link{mdaplot}} function can be used.}
}
\description{
Shows a plot with scores values for data objects.
}
