\name{plotSelectivityRatio}
\alias{plotSelectivityRatio}
\title{Selectivity ratio plot}
\usage{
plotSelectivityRatio(obj, ...)
}
\arguments{
  \item{obj}{a regression model}

  \item{...}{other parameters}
}
\description{
Generic function for plotting selectivity ratio values for
regression model (PCR, PLS, etc)
}

