% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmodel.R
\name{plotSpecificity.classmodel}
\alias{plotSpecificity.classmodel}
\title{Specificity plot for classification model}
\usage{
\method{plotSpecificity}{classmodel}(obj, legend.position = "bottomright", ...)
}
\arguments{
\item{obj}{classification model (object of class \code{plsda}, \code{simca}, etc.).}

\item{legend.position}{position of the legend (as in \code{mdaplotg}).}

\item{...}{parameters for \code{\link{plotPerformance.classmodel}} function.}
}
\description{
Makes a plot with specificity values vs. model complexity (e.g. number of components)
}
\details{
See examples in description of \code{\link{plsda}}, \code{\link{simca}} or \code{\link{simcam}}.
}
