% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotSpecificity}
\alias{plotSpecificity}
\title{Specificity plot}
\usage{
plotSpecificity(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Specificity plot
}
\details{
Generic function for plotting specificity values for classification model or results
}
