% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotT2DoF}
\alias{plotT2DoF}
\title{Degrees of freedom plot for score distance (Nh)}
\usage{
plotT2DoF(obj, type = "b", labels = "values", xticks = seq_len(obj$ncomp), ...)
}
\arguments{
\item{obj}{a PCA model (object of class \code{pca})}

\item{type}{type of the plot ("b", "l", "h")}

\item{labels}{what to show as data points labels}

\item{xticks}{vector with tick values for x-axis}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows a plot with degrees of freedom computed for score distances at given number
of components using data driven approach ("ddmoments" or "ddrobust").
}
\details{
Work only if parameter \code{lim.type} equal to "ddmoments" or "ddrobust".
}
