% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsres.R
\name{plotVariance.plsres}
\alias{plotVariance.plsres}
\title{Explained X variance plot for PLS results}
\usage{
\method{plotVariance}{plsres}(obj, decomp = "xdecomp", variance = "expvar", ...)
}
\arguments{
\item{obj}{PLS results (object of class \code{plsres})}

\item{decomp}{which dcomposition to use ("xdecomp" or "ydecomp")}

\item{variance}{which variance to use ("expvar", "cumexpvar")}

\item{...}{other plot parameters (see \code{mdaplot} for details)}
}
\description{
Shows plot with explained X variance vs. number of components.
}
\details{
See examples in help for \code{\link{plsres}} function.
}
