% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pls.R
\name{plotVIPScores.pls}
\alias{plotVIPScores.pls}
\title{VIP scores plot for PLS model}
\usage{
\method{plotVIPScores}{pls}(obj, ny = 1, type = "l", main = NULL,
  xlab = "Variables", ylab = "", ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{ny}{which response to get the values for (if y is multivariate)}

\item{type}{type of the plot}

\item{main}{main title for the plot}

\item{xlab}{label for x axies}

\item{ylab}{label for y axis}

\item{...}{other plot parameters (see \code{mdaplot} for details)}
}
\description{
Shows a plot with VIP scores values for given number of components
and response variable
}
\references{
[1] Il-Gyo Chong, Chi-Hyuck Jun. Chemometrics and Laboratory Systems, 78 (2005), pp. 103-112.
}

