% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pls.R
\name{plotXCumVariance.pls}
\alias{plotXCumVariance.pls}
\title{Cumulative explained X variance plot for PLS}
\usage{
\method{plotXCumVariance}{pls}(obj, type = "b",
  main = "X cumulative variance", xlab = "Components",
  ylab = "Explained variance, \%", show.legend = T, ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{type}{type of the plot('b', 'l' or 'h')}

\item{main}{main plot title}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{show.legend}{logical, show or not legend for the plot}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows plot with cumulative explained X variance vs. number of components.
}
\details{
See examples in help for \code{\link{pls}} function.
}

