% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotXCumVariance}
\alias{plotXCumVariance}
\title{X cumulative variance plot}
\usage{
plotXCumVariance(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
X cumulative variance plot
}
\details{
Generic function for plotting cumulative explained variance for decomposition of x data
}
