% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotXLoadings}
\alias{plotXLoadings}
\title{X loadings plot}
\usage{
plotXLoadings(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
X loadings plot
}
\details{
Generic function for plotting loadings values for decomposition of x data
}

