% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{plotXResiduals.pls}
\alias{plotXResiduals.pls}
\title{X residuals plot for PLS}
\usage{
\method{plotXResiduals}{pls}(obj, ncomp = NULL, main = NULL, xlab = "T2",
  ylab = "Squared residual distance (Q)", ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{ncomp}{how many components to use (if NULL - user selected optimal value will be used)}

\item{main}{main title for the plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows a plot with Q residuals vs. Hotelling T2 values for PLS decomposition of x data.
}
\details{
See examples in help for \code{\link{pls}} function.
}

