% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{plotXScores.pls}
\alias{plotXScores.pls}
\title{X scores plot for PLS}
\usage{
\method{plotXScores}{pls}(
  obj,
  comp = if (obj$ncomp > 1) c(1, 2) else 1,
  show.axes = TRUE,
  main = "Scores (X)",
  res = obj$res,
  ...
)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{comp}{which components to show the plot for (one or vector with several values)}

\item{show.axes}{logical, show or not a axes lines crossing origin (0,0)}

\item{main}{main plot title}

\item{res}{list with result objects to show the plot for (by defaul, model results are used)}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows plot with X scores values for selected components.
}
\details{
See examples in help for \code{\link{pls}} function.
}
