% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{plotXVariance.pls}
\alias{plotXVariance.pls}
\title{Explained X variance plot for PLS}
\usage{
\method{plotXVariance}{pls}(obj, type = "b", main = "Variance (X)", ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{type}{type of the plot("b", "l" or "h")}

\item{main}{title for the plot}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows plot with explained X variance vs. number of components.
}
\details{
See examples in help for \code{\link{pls}} function.
}
