% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsres.R
\name{plotXVariance.plsres}
\alias{plotXVariance.plsres}
\title{Explained X variance plot for PLS results}
\usage{
\method{plotXVariance}{plsres}(obj, main = "X variance", ...)
}
\arguments{
\item{obj}{PLS results (object of class \code{plsres})}

\item{main}{main plot title}

\item{...}{other plot parameters (see \code{mdaplot} for details)}
}
\description{
Shows plot with explained X variance vs. number of components.
}
\details{
See examples in help for \code{\link{plsres}} function.
}
