% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{plotXYScores.pls}
\alias{plotXYScores.pls}
\title{XY scores plot for PLS}
\usage{
\method{plotXYScores}{pls}(obj, ncomp = 1, show.axes = T, res = obj$res, ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{ncomp}{which component to show the plot for}

\item{show.axes}{logical, show or not a axes lines crossing origin (0,0)}

\item{res}{list with result objects to show the plot for (by defaul, model results are used)}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows plot with X vs. Y scores values for selected component.
}
\details{
See examples in help for \code{\link{pls}} function.
}
