% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pls.R
\name{plotXYScores.pls}
\alias{plotXYScores.pls}
\title{XY scores plot for PLS}
\usage{
\method{plotXYScores}{pls}(obj, comp = 1, main = NULL, xlab = "X scores",
  ylab = "Y scores", show.axes = T, show.labels = F, show.legend = T,
  ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{comp}{which component to show the plot for}

\item{main}{main plot title}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{show.axes}{logical, show or not a axes lines crossing origin (0,0)}

\item{show.labels}{logical, show or not labels for the plot elements}

\item{show.legend}{logical, show or not legend for the plot}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows plot with X vs. Y scores values for selected component.
}
\details{
See examples in help for \code{\link{pls}} function.
}

