% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plsres.R
\name{plotXYScores.plsres}
\alias{plotXYScores.plsres}
\title{XY scores plot for PLS results}
\usage{
\method{plotXYScores}{plsres}(obj, comp = 1, type = "p",
  main = "XY scores", xlab = "X scores", ylab = "Y scores", ...)
}
\arguments{
\item{obj}{PLS results (object of class \code{plsres})}

\item{comp}{which component to show the plot for}

\item{type}{type of the plot}

\item{main}{main plot title}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{...}{other plot parameters (see \code{mdaplot} for details)}
}
\description{
Shows plot with X vs. Y scores values for PLS results.
}
\details{
See examples in help for \code{\link{plsres}} function.
}

