% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsres.R
\name{plotXYScores.plsres}
\alias{plotXYScores.plsres}
\title{XY scores plot for PLS results}
\usage{
\method{plotXYScores}{plsres}(obj, ncomp = 1, ...)
}
\arguments{
\item{obj}{PLS results (object of class \code{plsres})}

\item{ncomp}{which component to show the plot for}

\item{...}{other plot parameters (see \code{mdaplot} for details)}
}
\description{
Shows plot with X vs. Y scores values for PLS results.
}
\details{
See examples in help for \code{\link{plsres}} function.
}
