% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotXYScores}
\alias{plotXYScores}
\title{XY scores plot}
\usage{
plotXYScores(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
XY scores plot
}
\details{
Generic function for plotting scores values for decomposition of x and y data
}

