% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsres.R
\name{plotYResiduals.plsres}
\alias{plotYResiduals.plsres}
\title{Y residuals plot for PLS results}
\usage{
\method{plotYResiduals}{plsres}(obj, ncomp = obj$ncomp.selected, ...)
}
\arguments{
\item{obj}{PLS results (object of class \code{plsres})}

\item{ncomp}{how many components to use (if NULL - user selected optimal value will be used)}

\item{...}{other plot parameters (see \code{mdaplot} for details)}
}
\description{
Shows a plot with Y residuals vs reference Y values for selected component.
}
\details{
Proxy for \code{\link{plotResiduals.regres}} function.
}
