% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{plotYResiduals}
\alias{plotYResiduals}
\title{Y residuals plot}
\usage{
plotYResiduals(obj, ...)
}
\arguments{
\item{obj}{a model or result object}

\item{...}{other arguments}
}
\description{
Y residuals plot
}
\details{
Generic function for plotting y residuals for classification or regression model or results
}
