% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{plotYVariance.pls}
\alias{plotYVariance.pls}
\title{Explained Y variance plot for PLS}
\usage{
\method{plotYVariance}{pls}(obj, type = "b", main = "Y variance",
  xlab = "Components", ylab = "Explained variance, \%", ...)
}
\arguments{
\item{obj}{a PLS model (object of class \code{pls})}

\item{type}{type of the plot('b', 'l' or 'h')}

\item{main}{main plot title}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{...}{other plot parameters (see \code{mdaplotg} for details)}
}
\description{
Shows plot with explained Y variance vs. number of components.
}
\details{
See examples in help for \code{\link{pls}} function.
}

