% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls.getyscores}
\alias{pls.getyscores}
\title{Compute and orthogonalize matrix with Y-scores}
\usage{
pls.getyscores(y, yloadings, xscores)
}
\arguments{
\item{y}{matrix with response values, already preprocessed and cleaned}

\item{yloadings}{matrix with Y-loadings}

\item{xscores}{matrix with X-scores (needed for orthogonalization)}
}
\value{
matrix with Y-scores
}
\description{
Compute and orthogonalize matrix with Y-scores
}
