% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{predict.pca}
\alias{predict.pca}
\title{PCA predictions}
\usage{
\method{predict}{pca}(object, x, cal = FALSE, ...)
}
\arguments{
\item{object}{a PCA model (object of class \code{pca})}

\item{x}{a matrix with data values}

\item{cal}{logical, if TRUE the predictions are made for calibration set}

\item{...}{other arguments}
}
\value{
PCA results (an object of class \code{pcares})
}
\description{
Applies PCA model to a new data
}

