% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{predict.plsda}
\alias{predict.plsda}
\title{PLS-DA predictions}
\usage{
\method{predict}{plsda}(object, x, c.ref = NULL, ...)
}
\arguments{
\item{object}{a PLS-DA model (object of class \code{plsda})}

\item{x}{a matrix with x values (predictors)}

\item{c.ref}{a vector with reference class values (should be a factor)}

\item{...}{other arguments}
}
\value{
PLS-DA results (an object of class \code{plsdares})
}
\description{
Applies PLS-DA model to a new data set
}
\details{
See examples in help for \code{\link{plsda}} function.
}
