% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.autoscale}
\alias{prep.autoscale}
\title{Autoscale values}
\usage{
prep.autoscale(data, center = T, scale = F, max.cov = 0.1)
}
\arguments{
\item{data}{a matrix with data values}

\item{center}{a logical value or vector with numbers for centering}

\item{scale}{a logical value or vector with numbers for weighting}

\item{max.cov}{columns that have coefficient of variation (in percent) below `max.cv` will not be scaled}
}
\value{
data matrix with processed values
}
\description{
Autoscale (mean center and standardize) values in columns of data matrix.
}
