% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.norm}
\alias{prep.norm}
\title{Normalization}
\usage{
prep.norm(data, type = "area")
}
\arguments{
\item{data}{a matrix with data values}

\item{type}{type of normalization \code{"area"}, \code{"length"} or \code{"sum"}.}
}
\value{
data matrix with normalized values
}
\description{
Normalizes signals (rows of data matrix) to unit area, unit length or unit sum
}
