% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{print.ddsimca}
\alias{print.ddsimca}
\title{Print method for DD-SIMCA model object}
\usage{
\method{print}{ddsimca}(x, ...)
}
\arguments{
\item{x}{a DD-SIMCA model (object of class \code{ddsimca})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}
