% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{print.ddsimcares}
\alias{print.ddsimcares}
\title{Print method for DD-SIMCA results}
\usage{
\method{print}{ddsimcares}(x, ...)
}
\arguments{
\item{x}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}
