% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldecomp.R
\name{print.ldecomp}
\alias{print.ldecomp}
\title{Print method for linear decomposition}
\usage{
\method{print}{ldecomp}(x, str = NULL, ...)
}
\arguments{
\item{x}{object of class \code{ldecomp}}

\item{str}{user specified text to show as a description of the object}

\item{...}{other arguments}
}
\description{
Generic \code{print} function for linear decomposition. Prints information about 
the \code{ldecomp} object.
}
