% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrals.R
\name{print.mcrals}
\alias{print.mcrals}
\title{Print method for mcrpure object}
\usage{
\method{print}{mcrals}(x, ...)
}
\arguments{
\item{x}{\code{mcrpure} object}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}
