% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcares.R
\name{print.pcares}
\alias{print.pcares}
\title{Print method for PCA results object}
\usage{
\method{print}{pcares}(x, ...)
}
\arguments{
\item{x}{PCA results (object of class \code{pcares})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}

