% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regres.R
\name{print.regres}
\alias{print.regres}
\title{print method for regression results object}
\usage{
\method{print}{regres}(x, ...)
}
\arguments{
\item{x}{regression results (object of class \code{regres})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}

