\name{print.simcares}
\alias{print.simcares}
\title{Print method for SIMCA results object}
\usage{
\method{print}{simcares}(x, ...)
}
\arguments{
  \item{x}{SIMCA results (object of class \code{simcares})}

  \item{...}{other arguments}
}
\description{
Prints information about the object structure
}

