% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{regres.bias}
\alias{regres.bias}
\title{Prediction bias}
\usage{
regres.bias(err)
}
\arguments{
\item{err}{vector with difference between reference and predicted y-values}
}
\description{
Calculates matrix with bias (average prediction error) for every response and components
}
