% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{regres.r2}
\alias{regres.r2}
\title{Determination coefficient}
\usage{
regres.r2(err, ytot)
}
\arguments{
\item{err}{vector with difference between reference and predicted y-values}

\item{ytot}{total variance for y-values}
}
\description{
Calculates matrix with coeffient of determination for every response and components
}
