% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{regres.r2}
\alias{regres.r2}
\title{Determination coefficient}
\usage{
regres.r2(y.ref, y.pred)
}
\arguments{
\item{y.ref}{vector with reference values}

\item{y.pred}{matrix with predicted values}
}
\description{
Calculates matrix with coeffient of determination for every response and components
}
