% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regres.R
\name{regres.rmse}
\alias{regres.rmse}
\title{RMSE}
\usage{
regres.rmse(y.ref, y.pred)
}
\arguments{
\item{y.ref}{vector with reference values}

\item{y.pred}{matrix with predicted values}
}
\description{
Calculates matrix with root mean squared error of prediction for every response and components.
}

