\name{selectCompNum.pls}
\alias{selectCompNum.pls}
\title{Select optimal number of components for PLS model}
\usage{
\method{selectCompNum}{pls}(model, ncomp)
}
\arguments{
  \item{model}{PLS model (object of class \code{pls})}

  \item{ncomp}{number of components to select}
}
\value{
the same model with selected number of components
}
\description{
Allows user to select optimal number of components for PLS
model
}
\details{
See examples in help for \code{\link{pls}} function.
}

