% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{selectCompNum}
\alias{selectCompNum}
\title{Select optimal number of components for a model}
\usage{
selectCompNum(model, ncomp)
}
\arguments{
\item{model}{a model object}

\item{ncomp}{number of components to select}
}
\description{
Generic function for selecting number of components for multivariate models (e.g. PCA, PLS, ...)
}

