% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{setDistanceLimits}
\alias{setDistanceLimits}
\title{Set residual distance limits for PCA model}
\usage{
setDistanceLimits(obj, ...)
}
\arguments{
\item{obj}{a SIMCA model}

\item{...}{other parameters}
}
\description{
Calculates and set critical limits for residuals of PCA model
}
