% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{setParams.ddsimca}
\alias{setParams.ddsimca}
\title{Set default parameters for the DD-SIMCA model.}
\usage{
\method{setParams}{ddsimca}(m, alpha = m$alpha, gamma = m$gamma)
}
\arguments{
\item{m}{a DD-SIMCA model (object of class \code{ddsimca})}

\item{alpha}{significance level for making the predictions.}

\item{gamma}{significance level for detection of outliers.}
}
\value{
DD-SIMCA model with the redefined parameters.
}
\description{
Set default parameters for the DD-SIMCA model.
}
