% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotseries.R
\name{showLabels}
\alias{showLabels}
\title{Show labels on plot}
\usage{
showLabels(
  ps,
  show.excluded = FALSE,
  pos = 3,
  cex = 0.65,
  col = "darkgray",
  force.x.values = NULL,
  bwd = 0.8
)
}
\arguments{
\item{ps}{`plotseries` object}

\item{show.excluded}{logical, are excluded rows also shown on the plot}

\item{pos}{position of the labels relative to the data points}

\item{cex}{size of the labels text}

\item{col}{color of the labels text}

\item{force.x.values}{vector with forced x-values (or NULL)}

\item{bwd}{bar width in case of bar plot}
}
\description{
Show labels on plot
}
