% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simcam.R
\name{simcam.getPerformanceStatistics}
\alias{simcam.getPerformanceStatistics}
\title{Performance statistics for SIMCAM model}
\usage{
simcam.getPerformanceStatistics(model)
}
\arguments{
\item{model}{SIMCAM model (object of class \code{simcam})}
}
\description{
Calculates discrimination power and distance between models for SIMCAM model.
}

