% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classres.R
\name{summary.classres}
\alias{summary.classres}
\title{Summary statistics about classification result object}
\usage{
\method{summary}{classres}(
  object,
  ncomp = object$ncomp.selected,
  nc = seq_len(object$nclasses),
  ...
)
}
\arguments{
\item{object}{classification results (object of class \code{plsdares}, \code{simcamres}, etc.).}

\item{ncomp}{which number of components to make the plot for (use NULL to show results for all available).}

\item{nc}{vector with class numbers to show the summary for.}

\item{...}{other arguments}
}
\description{
Generic \code{summary} function for classification results. Prints performance values for the
results.
}
