% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimcares.R
\name{summary.ddsimcares}
\alias{summary.ddsimcares}
\title{Summary method for DD-SIMCA results object}
\usage{
\method{summary}{ddsimcares}(object, ncomp = object$ncomp.selected, limType = "classic", ...)
}
\arguments{
\item{object}{DD-SIMCA results (object of class \code{ddsimcares})}

\item{...}{other arguments}
}
\description{
Shows performance statistics for the DD-SIMCA results.
}
