% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{summary.pca}
\alias{summary.pca}
\title{Summary method for PCA model object}
\usage{
\method{summary}{pca}(object, ...)
}
\arguments{
\item{object}{a PCA model (object of class \code{pca})}

\item{...}{other arguments}
}
\description{
Shows some statistics (explained variance, eigenvalues) for the model.
}

