% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{summary.pls}
\alias{summary.pls}
\title{Summary method for PLS model object}
\usage{
\method{summary}{pls}(
  object,
  ncomp = object$ncomp.selected,
  ny = seq_len(nrow(object$yloadings)),
  ...
)
}
\arguments{
\item{object}{a PLS model (object of class \code{pls})}

\item{ncomp}{how many components to count.}

\item{...}{other arguments}
}
\description{
Shows performance statistics for the model.
}
