% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsres.R
\name{summary.plsres}
\alias{summary.plsres}
\title{summary method for PLS results object}
\usage{
\method{summary}{plsres}(object, ny = seq_len(object$nresp), ncomp = NULL, ...)
}
\arguments{
\item{object}{PLS results (object of class \code{plsres})}

\item{ny}{for which response variable show the summary for}

\item{ncomp}{how many components to use (if NULL - user selected optimal value will be used)}

\item{...}{other arguments}
}
\description{
Shows performance statistics for the results.
}
