\name{summary.regres}
\alias{summary.regres}
\title{summary method for regression results object}
\usage{
\method{summary}{regres}(object, ncomp = NULL, ny = NULL, ...)
}
\arguments{
  \item{object}{regression results (object of class
  \code{regres})}

  \item{ncomp}{model complexity to show the summary for}

  \item{ny}{for which response variable show the summary
  for}

  \item{...}{other arguments}
}
\description{
Shows performance statistics for the regression results.
}

