% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simca.R
\name{summary.simca}
\alias{summary.simca}
\title{Summary method for SIMCA model object}
\usage{
\method{summary}{simca}(object, ncomp = object$ncomp.selected, res = object$res, ...)
}
\arguments{
\item{object}{a SIMCA model (object of class \code{simca})}

\item{ncomp}{number of components to show summary for}

\item{res}{list of result objects to show summary for}

\item{...}{other arguments}
}
\description{
Shows performance statistics for the model.
}
