\name{summary.simca}
\alias{summary.simca}
\title{Summary method for SIMCA model object}
\usage{
\method{summary}{simca}(object, ...)
}
\arguments{
  \item{object}{a SIMCA model (object of class
  \code{simca})}

  \item{...}{other arguments}
}
\description{
Shows performance statistics for the model.
}

