% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcam.R
\name{summary.simcam}
\alias{summary.simcam}
\title{Summary method for SIMCAM model object}
\usage{
\method{summary}{simcam}(object, ...)
}
\arguments{
\item{object}{a SIMCAM model (object of class \code{simcam})}

\item{...}{other arguments}
}
\description{
Shows performance statistics for the model.
}
