\name{summary.simcamres}
\alias{summary.simcamres}
\title{Summary method for SIMCAM results object}
\usage{
\method{summary}{simcamres}(object, ...)
}
\arguments{
  \item{object}{SIMCAM results (object of class
  \code{simcamres})}

  \item{...}{other arguments}
}
\description{
Shows performance statistics for the results.
}

