% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsimca.R
\name{writeJSON.ddsimca}
\alias{writeJSON.ddsimca}
\title{Saves DD-SIMCA model as JSON file compatible with web-application (https://mda.tools/ddsimca).}
\usage{
\method{writeJSON}{ddsimca}(obj, fileName, ...)
}
\arguments{
\item{obj}{Object with DD-SIMCA model (from \code{\link{ddsimca}}).}

\item{fileName}{Name or full path to JSON file to be created.}

\item{...}{other arguments (passed to \code{\link{asjson.ddsimca}}).}
}
\description{
You can load created JSON file to web-app and use it for prediction.
}
