% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{writeJSON.pca}
\alias{writeJSON.pca}
\title{Saves PCA model as JSON file compatible with web-application (https://mda.tools/pca).}
\usage{
\method{writeJSON}{pca}(obj, fileName)
}
\arguments{
\item{obj}{Object with PCA model (from \code{\link{pca}}).}

\item{fileName}{Name or full path to JSON file to be created.}
}
\description{
You can load created JSON file to web-app and use it for prediction.
}
